% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aaTable}
\alias{aaTable}
\title{Amino Acid table}
\format{data.frame object with 22 rows and two columns}
\usage{
aaTable
}
\value{
A data.frame object
}
\description{
Amino acid table for converting between three letter code and one letter code.
}
\keyword{datasets}

