% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{addCustomColumn}
\alias{addCustomColumn}
\alias{LogFile$addCustomColumn}
\alias{.LF$addCustomColumn}
\title{SLiM method addCustomColumn}
\usage{
addCustomColumn(columnName, source, context)
}
\arguments{
\item{columnName}{An object of type string or string or any. Must be of length 1
(a singleton). See details for description.}

\item{source}{An object of type string or string or any. Must be of length 1 (a
singleton). See details for description.}

\item{context}{An object of type string or string or any. The default value is
\code{NULL}. See details for description.}
}
\value{
An object of type void or void or void or void or void or void or void
or void or void or void or void or void or void or void or void.
}
\description{
Documentation for SLiM function \code{addCustomColumn}, which is a method of the
SLiM class \code{LogFile}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Documentation for this function can be found in the official \href{http://
benhaller.com/slim/SLiM_Manual.pdf#page=555}{SLiM manual: page 555}.

Adds a new data column with its name provided by columnName. The value
for the column, when a given row is generated, will be produced by the code
supplied in source, which is expected to return either NULL (which will write
out NA), or a singleton value of any non-object type. The context parameter will
be set up as a pseudo-parameter, named context, when source is called, allowing
the same source code to be used to generate values for multiple data columns;
you might, for example, provide the particular Subpopulation object here that
you wish source to use for its calculations. This is optional; if the default
value of NULL is used, then context will be NULL when source is called. See
addMeanSDColumns() for a useful variant.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016–2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
