% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{addPopulationSexRatio}
\alias{addPopulationSexRatio}
\alias{LogFile$addPopulationSexRatio}
\alias{.LF$addPopulationSexRatio}
\title{SLiM method addPopulationSexRatio}
\usage{
addPopulationSexRatio(void)
}
\arguments{
\item{void}{An object of type . See details for description.}
}
\value{
An object of type void or void or void or void or void or void or void
or void or void or void or void or void or void or void or void.
}
\description{
Documentation for SLiM function \code{addPopulationSexRatio}, which is a method
of the SLiM class \code{\link{LogFile}}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Documentation for this function can be found in the official \href{http://
benhaller.com/slim/SLiM_Manual.pdf#page=555}{SLiM manual: page 555}.

Adds a new data column that provides the population sex ratio M:(M+F),
named sex_ratio. If the model is hermaphroditic, NA will be written.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016–2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\seealso{
Other LogFile: 
\code{\link{LogFile}},
\code{\link{addCustomColumn}()},
\code{\link{addGenerationStage}()},
\code{\link{addGeneration}()},
\code{\link{addKeysAndValuesFrom}()},
\code{\link{addMeanSDColumns}()},
\code{\link{addPopulationSize}()},
\code{\link{addSubpopulationSexRatio}()},
\code{\link{addSubpopulationSize}()},
\code{\link{clearKeysAndValues}()},
\code{\link{flush}()},
\code{\link{logRow}()},
\code{\link{setFilePath}()},
\code{\link{setLogInterval}()},
\code{\link{setValue}()}
}
\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
\concept{LogFile}
