% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{addSpatialMap}
\alias{addSpatialMap}
\alias{Subpopulation$addSpatialMap}
\alias{.P$addSpatialMap}
\title{SLiM method addSpatialMap}
\usage{
addSpatialMap(map)
}
\arguments{
\item{map}{An object of type SpatialMap object. Must be of length 1 (a
singleton). See details for description.}
}
\value{
An object of type void.
}
\description{
Documentation for SLiM function \code{addSpatialMap}, which is a method of the
SLiM class \code{\link{Subpopulation}}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Documentation for this function can be found in the official
\href{http://benhaller.com/slim/SLiM_Manual.pdf#page=737}{SLiM manual: page
737}.

Adds the given SpatialMap object, map, to the subpopulation. (The
spatial map would have been previously created with a call to defineSpatialMap()
on a different subpopulation; addSpatialMap() can then be used to add that
existing spatial map with other subpopulations, sharing the map between
subpopulations.) If the map is already added to the target subpopulation, this
method does nothing; if a different map with the same name is already added to
the subpopulation, an error results (because map names must be unique within
each subpopulation). The map being added must be compatible with the target
subpopulation; in particular, the spatial bounds utilized by the map must
exactly match the corresponding spatial bounds for the subpopulation, and the
dimensionality of the subpopulation must encompass the spatiality of the map.
For example, if the map has a spatiality of "xz" then the subpopulation must
have a dimensionality of "xyz" so that it encompasses both "x" and "z", and
the subpopulation's spatial bounds for "x" and "z" must match those for the
map (but the spatial bounds for "y" are unimportant, since the map does not use
that dimension). Adding a map to a subpopulation is not strictly necessary, at
present; one may query a SpatialMap object directly using mapValue(), regarding
points in a subpopulation, without the map actually having been added to that
subpopulation. However, it is a good idea to use addSpatialMap(), both for its
compatibility check that prevents unnoticed scripting errors, and because it
ensures correct display of the model in SLiMgui.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016-2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\seealso{
Other Subpopulation: 
\code{\link{P}},
\code{\link{addCloned}()},
\code{\link{addCrossed}()},
\code{\link{addEmpty}()},
\code{\link{addRecombinant}()},
\code{\link{addSelfed}()},
\code{\link{cachedFitness}()},
\code{\link{configureDisplay}()},
\code{\link{defineSpatialMap}()},
\code{\link{outputMSSample}()},
\code{\link{outputSample}()},
\code{\link{outputVCFSample}()},
\code{\link{pointDeviated}()},
\code{\link{pointInBounds}()},
\code{\link{pointPeriodic}()},
\code{\link{pointReflected}()},
\code{\link{pointStopped}()},
\code{\link{pointUniform}()},
\code{\link{removeSpatialMap}()},
\code{\link{removeSubpopulation}()},
\code{\link{sampleIndividuals}()},
\code{\link{setCloningRate}()},
\code{\link{setMigrationRates}()},
\code{\link{setSelfingRate}()},
\code{\link{setSexRatio}()},
\code{\link{setSpatialBounds}()},
\code{\link{setSubpopulationSize}()},
\code{\link{spatialMapColor}()},
\code{\link{spatialMapImage}()},
\code{\link{spatialMapValue}()},
\code{\link{subsetIndividuals}()},
\code{\link{takeMigrants}()}
}
\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
\concept{Subpopulation}
