% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{addSuppliedColumn}
\alias{addSuppliedColumn}
\alias{LogFile$addSuppliedColumn}
\alias{.LF$addSuppliedColumn}
\title{SLiM method addSuppliedColumn}
\usage{
addSuppliedColumn(columnName)
}
\arguments{
\item{columnName}{An object of type string. Must be of length 1 (a singleton).
See details for description.}
}
\value{
An object of type void or void or void or void or void or void or void
or void or void or void or void or void or void or void or void or void or void
or void or logical.
}
\description{
Documentation for SLiM function \code{addSuppliedColumn}, which is a method of
the SLiM class \code{\link{LogFile}}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Documentation for this function can be found in the official
\href{http://benhaller.com/slim/SLiM_Manual.pdf#page=702}{SLiM manual: page
702}.

Adds a new data column with its name provided by columnName. The value
for the column is initially undefined, and will be written as NA. A different
value may (optionally) be provided by calling setSuppliedValue() with a value
for columnName. That value will be used for the column the next time a row is
generated (whether automatically or by a call to logRow()), and the column's
value will subsequently be undefined again. In other words, for any given logged
row the default of NA may be kept, or a different value may be supplied. This
allows the value for the column to be set at any point during the tick cycle,
which can be convenient if the column's value depends upon transient state that
is no longer available at the time the row is logged.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016-2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\seealso{
Other LogFile: 
\code{\link{LF}},
\code{\link{addCustomColumn}()},
\code{\link{addCycleStage}()},
\code{\link{addCycle}()},
\code{\link{addKeysAndValuesFrom}()},
\code{\link{addMeanSDColumns}()},
\code{\link{addPopulationSexRatio}()},
\code{\link{addPopulationSize}()},
\code{\link{addSubpopulationSexRatio}()},
\code{\link{addSubpopulationSize}()},
\code{\link{addTick}()},
\code{\link{clearKeysAndValues}()},
\code{\link{flush}()},
\code{\link{logRow}()},
\code{\link{setFilePath}()},
\code{\link{setLogInterval}()},
\code{\link{setSuppliedValue}()},
\code{\link{setValue}()},
\code{\link{willAutolog}()}
}
\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
\concept{LogFile}
