% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{ancestry_deserts}
\alias{ancestry_deserts}
\title{Extract coordinates of deserts from a given VCF file.}
\usage{
ancestry_deserts(markers, cutoff = 0)
}
\arguments{
\item{markers}{GRanges object with marker genotypes.}

\item{cutoff}{Frequency cutoff for admixture markers.}
}
\value{
GRanges object with ancestry deserts.
}
\description{
Extract coordinates of deserts from a given VCF file.
}
