% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{ancestry_tracks}
\alias{ancestry_tracks}
\title{Given a chromosome with fixed ancestry markers, infer coordinates of
introgressed tracks.}
\usage{
ancestry_tracks(markers, chroms)
}
\arguments{
\item{markers}{GRanges object with marker genotypes.}

\item{chroms}{Identifiers of chromosomes to analyze.}
}
\value{
GRanges object with coordinates of detected tracks.
}
\description{
Given a chromosome with fixed ancestry markers, infer coordinates of
introgressed tracks.
}
\seealso{
mut_gt
}
