% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paletteDiscrete.R
\name{ArchRPalettes}
\alias{ArchRPalettes}
\title{ArchR colour palettes}
\format{
A named list of character vectors.
}
\source{
\url{https://github.com/GreenleafLab/ArchR}
}
\usage{
ArchRPalettes
}
\description{
A list of colour vectors reproducing the palettes used by the
\pkg{ArchR} package.  Names correspond to the \code{set} argument of
\code{paletteDiscrete()}.
}
