% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_run_script.R
\name{as.character.slim_script}
\alias{as.character.slim_script}
\title{Title}
\usage{
\method{as.character}{slim_script}(slim_script)
}
\arguments{
\item{slim_script}{}
}
\value{

}
\description{
Title
}
