% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_script.R
\name{as_slim_text.slimr_script}
\alias{as_slim_text.slimr_script}
\title{Convert a slimr_script to a length 1 character vector}
\usage{
\method{as_slim_text}{slimr_script}(x, ...)
}
\arguments{
\item{x}{\code{slimr_script} object to convert}

\item{...}{Further arguments, passed to or from other methods.}
}
\value{
A length 1 character vector
}
\description{
Convert a slimr_script to a length 1 character vector
}
\examples{
slim_script(
   slim_block(initialize(),
          {
            initializeMutationRate(1e-7);
            initializeMutationType("m1", 0.5, "f", 0.0);
            initializeGenomicElementType("g1", m1, 1.0);
            initializeGenomicElement(g1, 0, 99999);
            initializeRecombinationRate(1e-8);
          }),
   slim_block(1,
          {
            sim.addSubpop("p1", 500);
          }),
   slim_block(10000,
          {
            sim.simulationFinished();
          })
) -> script
as_slim_text(script)
}
