% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_code.R
\name{as_slimr_code}
\alias{as_slimr_code}
\title{Convert SLiM code text into equivalent `slimr` code to produce the same model}
\usage{
as_slimr_code(code_txt)
}
\arguments{
\item{code_txt}{}
}
\value{
`slimr` code as a character vector, invisibly
}
\description{
This function print the equivalent `slimr` code to the console where it can easily
be copied and pasted into an R script.
}
\examples{
as_slimr_code(slim_recipes$`5.3.4`)
}
