% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_script.R
\name{as_slimr_script}
\alias{as_slimr_script}
\title{Convert a character vector into a slim_script object}
\usage{
as_slimr_script(slim_script_text)
}
\arguments{
\item{slim_script_text}{A character vector giving the full SLiM script to
convert to a \code{slimr_script} object. Character vectors with length greater than
1 will be concatenated with newline separators}
}
\value{
A \code{slimr_script} object
}
\description{
Convert a character vector into a slim_script object
}
