% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_output.R
\name{calculate_indiv_load}
\alias{calculate_indiv_load}
\title{Calculate the total multiplicative genetic load in a given individual based
on a specified dominance.}
\usage{
calculate_indiv_load(indiv_id, muts_per_genome)
}
\description{
Calculate the total multiplicative genetic load in a given individual based
on a specified dominance.
}
