% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_callbacks.R
\name{initialize}
\alias{initialize}
\title{SLiM initialize() callback}
\usage{
initialize()
}
\value{
None
}
\description{
This callback specifies the code should be called to set up the simulation.
For details on what you can do in an \code{initialize()} callback code blocks see
\href{http://benhaller.com/slim/SLiM_Manual.pdf#page=516}{SLiM Manual: page 516}
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
modified from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016-2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\seealso{
Other callbacks: 
\code{\link{early}()},
\code{\link{fitness}()},
\code{\link{interaction}()},
\code{\link{late}()},
\code{\link{mateChoice}()},
\code{\link{modifyChild}()},
\code{\link{recombination}()},
\code{\link{reproduction}()},
\code{\link{slim_callbacks}()}
}
\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
\concept{callbacks}
