% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{changeValues}
\alias{changeValues}
\alias{SpatialMap$changeValues}
\alias{.SM$changeValues}
\title{SLiM method changeValues}
\usage{
changeValues(x)
}
\arguments{
\item{x}{An object of type integer or float or SpatialMap object. See details
for description.}
}
\value{
An object of type void.
}
\description{
Documentation for SLiM function \code{changeValues}, which is a method of the
SLiM class \code{\link{SpatialMap}}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Documentation for this function can be found in the official
\href{http://benhaller.com/slim/SLiM_Manual.pdf#page=713}{SLiM manual: page
713}.

Changes the grid values used for the target spatial map. The parameter
x should be either a SpatialMap object from which values are taken directly, or
a vector, matrix, or array of numeric values as described in the documentation
for defineSpatialMap(). Other characteristics of the spatial map, such as its
color mapping (if defined), its spatial bounds, and its spatiality, will remain
unchanged. The grid resolution of the spatial map is allowed to change with
this method. This method is useful for changing the values of a spatial map
over time, such as to implement changes to the landscape's characteristics due
to seasonality, climate change, processes such as fire or urbanization, and
so forth. As with the original map values provided to defineSpatialMap(), it
is often useful to read map values from a PNG image file using the Eidos class
Image.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016-2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\seealso{
Other SpatialMap: 
\code{\link{SM}},
\code{\link{add}()},
\code{\link{blend}()},
\code{\link{changeColors}()},
\code{\link{divide}()},
\code{\link{exp}()},
\code{\link{gridValues}()},
\code{\link{interpolate}()},
\code{\link{mapColor}()},
\code{\link{mapImage}()},
\code{\link{mapValue}()},
\code{\link{multiply}()},
\code{\link{power}()},
\code{\link{range}()},
\code{\link{rescale}()},
\code{\link{sampleImprovedNearbyPoint}()},
\code{\link{sampleNearbyPoint}()},
\code{\link{smooth}()},
\code{\link{subtract}()}
}
\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
\concept{SpatialMap}
