% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_script.R
\name{code}
\alias{code}
\alias{code<-}
\title{Extract or set code from a \code{slimr_script} object}
\usage{
code(x)

code(x) <- value
}
\arguments{
\item{x}{A \code{slimr_script} object}

\item{value}{Code to replace with.}
}
\description{
Extract or set code from a \code{slimr_script} object
}
\examples{
script <- slim_script(
  slim_block_init_minimal(),
  slim_block_finish(100)
)
code(script)
}
