% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{containsMutations}
\alias{containsMutations}
\alias{Genome$containsMutations}
\alias{.G$containsMutations}
\alias{Individual$containsMutations}
\alias{.I$containsMutations}
\title{SLiM method containsMutations}
\usage{
containsMutations(mutations)

containsMutations(mutations)
}
\arguments{
\item{mutations}{An object of type Mutation object. See details for description.}
}
\value{
An object of type logical.

An object of type logical.
}
\description{
Documentation for SLiM function \code{containsMutations}, which is a method of
the SLiM class \code{Genome}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.

Documentation for SLiM function \code{containsMutations}, which is a method of
the SLiM class \code{Individual}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Returns a logical vector indicating whether each of the mutations
in mutations is present in the genome; each element in the returned vector
indicates whether the corresponding mutation is present (T) or absent (F). This
method is provided for speed; it is much faster than the corresponding Eidos
code.

Returns a logical vector indicating whether each of the mutations in
mutations is present in the individual (in either of its genomes); each element
in the returned vector indicates whether the corresponding mutation is present
(T) or absent (F). This method is provided for speed; it is much faster than the
corresponding Eidos code.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016–2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}


This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016–2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})

Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
