% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_wsl_path}
\alias{convert_to_wsl_path}
\title{Convert a path into a path that works inside Windows Subsystem for Linux}
\usage{
convert_to_wsl_path(windows_path)
}
\arguments{
\item{windows_path}{The windows path to convert.}
}
\value{
The modified path.
}
\description{
WSL can access windows files through a directory mounted in the /mnt directory. This function converts a windows path from outside WSL to point to the correct directory inside WSL
}
