% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{createDB}
\alias{createDB}
\title{createDB}
\usage{
createDB(iupredPath = "/Users/buyar/Desktop/work/tools/iupred",
  uniprotAccessions, workingDirectory = getwd(), nodeN = 1,
  overwrite = FALSE)
}
\arguments{
\item{uniprotAccessions}{Vector of uniprot accessions (e.g. c('P04637',
'P11166'))}

\item{workingDirectory}{Path to folder where the database should be created}

\item{nodeN}{Number of cores to run the database generation.}

\item{overwrite}{TRUE/FALSE (default: FALSE) if the database is populated in
an existing folder, should the previously generated data be overwritten? Set
to TRUE to update/overwrite existing files/folders}

\item{saveIndividualRDS}{TRUE/FALSE (default: FALSE) Boolean value to decide if
the extracted data for each protein should be saved individually to separate
RDS files.}
}
\description{
Wrapper function to run extractProteinData in multiple nodes for for multiple
inputs and create a single database collecting all available information into
a single R object
}

