% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{downloadUniprotFiles}
\alias{downloadUniprotFiles}
\title{downloadUniprotFiles}
\usage{
downloadUniprotFiles(uniprotAccessions, outDir, format, overwrite = FALSE)
}
\arguments{
\item{uniprotAccessions}{A vector of uniprot accession numbers}

\item{outDir}{Path to folder where downloaded files should be stored.}

\item{format}{The format of files to be downloaded. Options are fasta, gff,
and txt.}

\item{overwrite}{TRUE or FALSE (default). Whether the existing files should
be overwritten with a new download.}
}
\description{
Given a uniprot accession number and a file format, download the annotation
file for a protein in uniprot in the given format. e.g. For human p53
protein, the uniprot accession is P04637. This function can be used to
download the fasta sequence of this protein from
http://www.uniprot.org/uniprot/P04637.fasta. This function is mostly useful
for bulk downloads when there are many uniprot accessions to be downloaded.
}
\examples{
ids <- c('P04637', 'P11166', 'P06400')
downloadUniprotFiles(uniprotAccessions = ids, outDir = getwd(),
                                format = 'fasta', overwrite = FALSE)

}

