% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{drawBreakpoints}
\alias{drawBreakpoints}
\alias{Chromosome$drawBreakpoints}
\alias{.c$drawBreakpoints}
\title{SLiM method drawBreakpoints}
\usage{
drawBreakpoints(parent, n)
}
\arguments{
\item{parent}{An object of type null or Individual object. Must be of length 1
(a singleton). The default value is \code{NULL}. See details for description.}

\item{n}{An object of type null or integer. Must be of length 1 (a singleton).
The default value is \code{NULL}. See details for description.}
}
\value{
An object of type integer.
}
\description{
Documentation for SLiM function \code{drawBreakpoints}, which is a method of the
SLiM class \code{Chromosome}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Draw recombination breakpoints, using the chromosome’s recombination
rate map, the current gene conversion parameters, and (in some cases – see
below) any active and applicable recombination() callbacks. The number of
breakpoints to generate, n, may be supplied; if it is NULL (the default), the
number of breakpoints will be drawn based upon the overall recombination rate
and the chromosome length (following the standard procedure in SLiM). Note that
if the double-stranded breaks model has been chosen, the number of breakpoints
generated will probably not be equal to the number requested, because most
breakpoints will entail gene conversion tracts, which entail additional
crossover breakpoints. It is generally recommended that the parent individual be
supplied to this method, but parent is NULL by default. The individual supplied
in parent is used for two purposes. First, in sexual models that define separate
recombination rate maps for males versus females, the sex of parent will be used
to determine which map is used; in this case, a non-NULL value must be supplied
for parent, since the choice of recombination rate map must be determined.
Second, in models that define recombination() callbacks, parent is used to
determine the various pseudo-parameters that are passed to recombination()
callbacks (individual, genome1, genome2, subpop), and the subpopulation to
which parent belongs is used to select which recombination() callbacks are
applicable; given the necessity of this information, recombination() callbacks
will not be called as a side effect of this method if parent is NULL. Apart from
these two uses, parent is not used, and the caller does not guarantee that the
generated breakpoints will actually be used to recombine the genomes of parent
in particular.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016–2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
