% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{drawSelectionCoefficient}
\alias{drawSelectionCoefficient}
\alias{MutationType$drawSelectionCoefficient}
\alias{.MT$drawSelectionCoefficient}
\title{SLiM method drawSelectionCoefficient}
\usage{
drawSelectionCoefficient(n)
}
\arguments{
\item{n}{An object of type integer. Must be of length 1 (a singleton). The
default value is \code{1}. See details for description.}
}
\value{
An object of type float or any but object or void or void.
}
\description{
Documentation for SLiM function \code{drawSelectionCoefficient}, which is a
method of the SLiM class \code{MutationType}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Draws and returns a vector of n selection coefficients using the
currently defined distribution of fitness effects (DFE) for the target mutation
type. See section 23.9 above for discussion of the supported distributions
and their uses. If the DFE is type "s", this method will result in synchronous
execution of the DFE’s script.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016–2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
