% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_classes.R, R/slim_lang.R
\name{E}
\alias{E}
\alias{Eidos}
\title{Eidos}
\description{
Documentation for Eidos class from SLiM
}
\details{
NA
This class has the following methods (functions):
\itemize{
\item{\code{\link{abs}}}
\item{\code{\link{acos}}}
\item{\code{\link{asin}}}
\item{\code{\link{atan}}}
\item{\code{\link{atan2}}}
\item{\code{\link{ceil}}}
\item{\code{\link{cos}}}
\item{\code{\link{cumProduct}}}
\item{\code{\link{cumSum}}}
\item{\code{\link{exp}}}
\item{\code{\link{floor}}}
\item{\code{\link{integerDiv}}}
\item{\code{\link{integerMod}}}
\item{\code{\link{isFinite}}}
\item{\code{\link{isInfinite}}}
\item{\code{\link{isNAN}}}
\item{\code{\link{log}}}
\item{\code{\link{log10}}}
\item{\code{\link{log2}}}
\item{\code{\link{product}}}
\item{\code{\link{round}}}
\item{\code{\link{setDifference}}}
\item{\code{\link{setIntersection}}}
\item{\code{\link{setSymmetricDifference}}}
\item{\code{\link{setUnion}}}
\item{\code{\link{sin}}}
\item{\code{\link{sqrt}}}
\item{\code{\link{sum}}}
\item{\code{\link{sumExact}}}
\item{\code{\link{tan}}}
\item{\code{\link{trunc}}}
\item{\code{\link{cor}}}
\item{\code{\link{cov}}}
\item{\code{\link{max}}}
\item{\code{\link{mean}}}
\item{\code{\link{min}}}
\item{\code{\link{pmax}}}
\item{\code{\link{pmin}}}
\item{\code{\link{quantile}}}
\item{\code{\link{range}}}
\item{\code{\link{rank}}}
\item{\code{\link{sd}}}
\item{\code{\link{ttest}}}
\item{\code{\link{var}}}
\item{\code{\link{dmvnorm}}}
\item{\code{\link{dbeta}}}
\item{\code{\link{dexp}}}
\item{\code{\link{dgamma}}}
\item{\code{\link{dnorm}}}
\item{\code{\link{findInterval}}}
\item{\code{\link{pnorm}}}
\item{\code{\link{qnorm}}}
\item{\code{\link{rbeta}}}
\item{\code{\link{rbinom}}}
\item{\code{\link{rcauchy}}}
\item{\code{\link{rdunif}}}
\item{\code{\link{rexp}}}
\item{\code{\link{rf}}}
\item{\code{\link{rgamma}}}
\item{\code{\link{rgeom}}}
\item{\code{\link{rlnorm}}}
\item{\code{\link{rmvnorm}}}
\item{\code{\link{rnbinom}}}
\item{\code{\link{rnorm}}}
\item{\code{\link{rpois}}}
\item{\code{\link{runif}}}
\item{\code{\link{rweibull}}}
\item{\code{\link{c}}}
\item{\code{\link{float}}}
\item{\code{\link{integer}}}
\item{\code{\link{logical}}}
\item{\code{\link{object}}}
\item{\code{\link{rep}}}
\item{\code{\link{repEach}}}
\item{\code{\link{sample}}}
\item{\code{\link{seq}}}
\item{\code{\link{seqAlong}}}
\item{\code{\link{seqLen}}}
\item{\code{\link{string}}}
\item{\code{\link{all}}}
\item{\code{\link{any}}}
\item{\code{\link{cat}}}
\item{\code{\link{catn}}}
\item{\code{\link{format}}}
\item{\code{\link{identical}}}
\item{\code{\link{ifelse}}}
\item{\code{\link{length}}}
\item{\code{\link{match}}}
\item{\code{\link{order}}}
\item{\code{\link{paste}}}
\item{\code{\link{paste0}}}
\item{\code{\link{print}}}
\item{\code{\link{rev}}}
\item{\code{\link{size}}}
\item{\code{\link{sort}}}
\item{\code{\link{sortBy}}}
\item{\code{\link{str}}}
\item{\code{\link{tabulate}}}
\item{\code{\link{unique}}}
\item{\code{\link{which}}}
\item{\code{\link{whichMax}}}
\item{\code{\link{whichMin}}}
\item{\code{\link{asFloat}}}
\item{\code{\link{asInteger}}}
\item{\code{\link{asLogical}}}
\item{\code{\link{asString}}}
\item{\code{\link{elementType}}}
\item{\code{\link{isFloat}}}
\item{\code{\link{isInteger}}}
\item{\code{\link{isLogical}}}
\item{\code{\link{isNULL}}}
\item{\code{\link{isObject}}}
\item{\code{\link{isString}}}
\item{\code{\link{type}}}
\item{\code{\link{nchar}}}
\item{\code{\link{strcontains}}}
\item{\code{\link{strfind}}}
\item{\code{\link{strprefix}}}
\item{\code{\link{strsplit}}}
\item{\code{\link{strsuffix}}}
\item{\code{\link{substr}}}
\item{\code{\link{apply}}}
\item{\code{\link{array}}}
\item{\code{\link{cbind}}}
\item{\code{\link{diag}}}
\item{\code{\link{dim}}}
\item{\code{\link{drop}}}
\item{\code{\link{lowerTri}}}
\item{\code{\link{matrix}}}
\item{\code{\link{matrixMult}}}
\item{\code{\link{ncol}}}
\item{\code{\link{nrow}}}
\item{\code{\link{rbind}}}
\item{\code{\link{t}}}
\item{\code{\link{upperTri}}}
\item{\code{\link{createDirectory}}}
\item{\code{\link{deleteFile}}}
\item{\code{\link{fileExists}}}
\item{\code{\link{filesAtPath}}}
\item{\code{\link{flushFile}}}
\item{\code{\link{getwd}}}
\item{\code{\link{readCSV}}}
\item{\code{\link{readFile}}}
\item{\code{\link{setwd}}}
\item{\code{\link{tempdir}}}
\item{\code{\link{writeFile}}}
\item{\code{\link{writeTempFile}}}
\item{\code{\link{cmColors}}}
\item{\code{\link{colors}}}
\item{\code{\link{color2rgb}}}
\item{\code{\link{heatColors}}}
\item{\code{\link{hsv2rgb}}}
\item{\code{\link{rainbow}}}
\item{\code{\link{rgb2color}}}
\item{\code{\link{rgb2hsv}}}
\item{\code{\link{terrainColors}}}
\item{\code{\link{assert}}}
\item{\code{\link{beep}}}
\item{\code{\link{citation}}}
\item{\code{\link{clock}}}
\item{\code{\link{date}}}
\item{\code{\link{debugIndent}}}
\item{\code{\link{defineConstant}}}
\item{\code{\link{defineGlobal}}}
\item{\code{\link{exists}}}
\item{\code{\link{functionSignature}}}
\item{\code{\link{functionSource}}}
\item{\code{\link{getSeed}}}
\item{\code{\link{license}}}
\item{\code{\link{ls}}}
\item{\code{\link{rm}}}
\item{\code{\link{sapply}}}
\item{\code{\link{setSeed}}}
\item{\code{\link{source}}}
\item{\code{\link{stop}}}
\item{\code{\link{suppressWarnings}}}
\item{\code{\link{sysinfo}}}
\item{\code{\link{system}}}
\item{\code{\link{time}}}
\item{\code{\link{usage}}}
\item{\code{\link{version}}}
}
This class has the following properties:
\describe{
\item{None.}{This class has no properties.}
}
}
\seealso{
Other Eidos: 
\code{\link{eidos_abs}()},
\code{\link{eidos_acos}()},
\code{\link{eidos_all}()},
\code{\link{eidos_any}()},
\code{\link{eidos_apply}()},
\code{\link{eidos_array}()},
\code{\link{eidos_asFloat}()},
\code{\link{eidos_asInteger}()},
\code{\link{eidos_asLogical}()},
\code{\link{eidos_asString}()},
\code{\link{eidos_asin}()},
\code{\link{eidos_assert}()},
\code{\link{eidos_atan2}()},
\code{\link{eidos_atan}()},
\code{\link{eidos_beep}()},
\code{\link{eidos_catn}()},
\code{\link{eidos_cat}()},
\code{\link{eidos_cbind}()},
\code{\link{eidos_ceil}()},
\code{\link{eidos_citation}()},
\code{\link{eidos_clock}()},
\code{\link{eidos_cmColors}()},
\code{\link{eidos_color2rgb}()},
\code{\link{eidos_colors}()},
\code{\link{eidos_cor}()},
\code{\link{eidos_cos}()},
\code{\link{eidos_cov}()},
\code{\link{eidos_createDirectory}()},
\code{\link{eidos_cumProduct}()},
\code{\link{eidos_cumSum}()},
\code{\link{eidos_c}()},
\code{\link{eidos_date}()},
\code{\link{eidos_dbeta}()},
\code{\link{eidos_debugIndent}()},
\code{\link{eidos_defineConstant}()},
\code{\link{eidos_defineGlobal}()},
\code{\link{eidos_deleteFile}()},
\code{\link{eidos_dexp}()},
\code{\link{eidos_dgamma}()},
\code{\link{eidos_diag}()},
\code{\link{eidos_dim}()},
\code{\link{eidos_dmvnorm}()},
\code{\link{eidos_dnorm}()},
\code{\link{eidos_drop}()},
\code{\link{eidos_elementType}()},
\code{\link{eidos_exists}()},
\code{\link{eidos_exp}()},
\code{\link{eidos_fileExists}()},
\code{\link{eidos_filesAtPath}()},
\code{\link{eidos_findInterval}()},
\code{\link{eidos_float}()},
\code{\link{eidos_floor}()},
\code{\link{eidos_flushFile}()},
\code{\link{eidos_format}()},
\code{\link{eidos_functionSignature}()},
\code{\link{eidos_functionSource}()},
\code{\link{eidos_getSeed}()},
\code{\link{eidos_getwd}()},
\code{\link{eidos_heatColors}()},
\code{\link{eidos_hsv2rgb}()},
\code{\link{eidos_identical}()},
\code{\link{eidos_ifelse}()},
\code{\link{eidos_integerDiv}()},
\code{\link{eidos_integerMod}()},
\code{\link{eidos_integer}()},
\code{\link{eidos_isFinite}()},
\code{\link{eidos_isFloat}()},
\code{\link{eidos_isInfinite}()},
\code{\link{eidos_isInteger}()},
\code{\link{eidos_isLogical}()},
\code{\link{eidos_isNAN}()},
\code{\link{eidos_isNULL}()},
\code{\link{eidos_isObject}()},
\code{\link{eidos_isString}()},
\code{\link{eidos_length}()},
\code{\link{eidos_license}()},
\code{\link{eidos_log10}()},
\code{\link{eidos_log2}()},
\code{\link{eidos_logical}()},
\code{\link{eidos_log}()},
\code{\link{eidos_lowerTri}()},
\code{\link{eidos_ls}()},
\code{\link{eidos_match}()},
\code{\link{eidos_matrixMult}()},
\code{\link{eidos_matrix}()},
\code{\link{eidos_max}()},
\code{\link{eidos_mean}()},
\code{\link{eidos_min}()},
\code{\link{eidos_nchar}()},
\code{\link{eidos_ncol}()},
\code{\link{eidos_nrow}()},
\code{\link{eidos_object}()},
\code{\link{eidos_order}()},
\code{\link{eidos_paste0}()},
\code{\link{eidos_paste}()},
\code{\link{eidos_pmax}()},
\code{\link{eidos_pmin}()},
\code{\link{eidos_pnorm}()},
\code{\link{eidos_print}()},
\code{\link{eidos_product}()},
\code{\link{eidos_qnorm}()},
\code{\link{eidos_quantile}()},
\code{\link{eidos_rainbow}()},
\code{\link{eidos_range}()},
\code{\link{eidos_rank}()},
\code{\link{eidos_rbeta}()},
\code{\link{eidos_rbind}()},
\code{\link{eidos_rbinom}()},
\code{\link{eidos_rcauchy}()},
\code{\link{eidos_rdunif}()},
\code{\link{eidos_readCSV}()},
\code{\link{eidos_readFile}()},
\code{\link{eidos_repEach}()},
\code{\link{eidos_rep}()},
\code{\link{eidos_rev}()},
\code{\link{eidos_rexp}()},
\code{\link{eidos_rf}()},
\code{\link{eidos_rgamma}()},
\code{\link{eidos_rgb2color}()},
\code{\link{eidos_rgb2hsv}()},
\code{\link{eidos_rgeom}()},
\code{\link{eidos_rlnorm}()},
\code{\link{eidos_rmvnorm}()},
\code{\link{eidos_rm}()},
\code{\link{eidos_rnbinom}()},
\code{\link{eidos_rnorm}()},
\code{\link{eidos_round}()},
\code{\link{eidos_rpois}()},
\code{\link{eidos_runif}()},
\code{\link{eidos_rweibull}()},
\code{\link{eidos_sample}()},
\code{\link{eidos_sapply}()},
\code{\link{eidos_sd}()},
\code{\link{eidos_seqAlong}()},
\code{\link{eidos_seqLen}()},
\code{\link{eidos_seq}()},
\code{\link{eidos_setDifference}()},
\code{\link{eidos_setIntersection}()},
\code{\link{eidos_setSeed}()},
\code{\link{eidos_setSymmetricDifference}()},
\code{\link{eidos_setUnion}()},
\code{\link{eidos_setwd}()},
\code{\link{eidos_sin}()},
\code{\link{eidos_size}()},
\code{\link{eidos_sortBy}()},
\code{\link{eidos_sort}()},
\code{\link{eidos_source}()},
\code{\link{eidos_sqrt}()},
\code{\link{eidos_stop}()},
\code{\link{eidos_strcontains}()},
\code{\link{eidos_strfind}()},
\code{\link{eidos_string}()},
\code{\link{eidos_strprefix}()},
\code{\link{eidos_strsplit}()},
\code{\link{eidos_strsuffix}()},
\code{\link{eidos_str}()},
\code{\link{eidos_substr}()},
\code{\link{eidos_sumExact}()},
\code{\link{eidos_sum}()},
\code{\link{eidos_suppressWarnings}()},
\code{\link{eidos_sysinfo}()},
\code{\link{eidos_system}()},
\code{\link{eidos_tabulate}()},
\code{\link{eidos_tan}()},
\code{\link{eidos_tempdir}()},
\code{\link{eidos_terrainColors}()},
\code{\link{eidos_time}()},
\code{\link{eidos_trunc}()},
\code{\link{eidos_ttest}()},
\code{\link{eidos_type}()},
\code{\link{eidos_t}()},
\code{\link{eidos_unique}()},
\code{\link{eidos_upperTri}()},
\code{\link{eidos_usage}()},
\code{\link{eidos_var}()},
\code{\link{eidos_version}()},
\code{\link{eidos_whichMax}()},
\code{\link{eidos_whichMin}()},
\code{\link{eidos_which}()},
\code{\link{eidos_writeFile}()},
\code{\link{eidos_writeTempFile}()}
}
\concept{Eidos}
