% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_classes.R, R/slim_lang.R
\name{E}
\alias{E}
\alias{Eidos}
\title{Eidos}
\description{
Documentation for Eidos class from SLiM
}
\details{
NA
This class has the following methods (functions):
\itemize{
\item{\code{\link{eidos_abs}}}
\item{\code{\link{eidos_acos}}}
\item{\code{\link{eidos_asin}}}
\item{\code{\link{eidos_atan}}}
\item{\code{\link{eidos_atan2}}}
\item{\code{\link{eidos_ceil}}}
\item{\code{\link{eidos_cos}}}
\item{\code{\link{eidos_cumProduct}}}
\item{\code{\link{eidos_cumSum}}}
\item{\code{\link{eidos_exp}}}
\item{\code{\link{eidos_floor}}}
\item{\code{\link{eidos_integerDiv}}}
\item{\code{\link{eidos_integerMod}}}
\item{\code{\link{eidos_isFinite}}}
\item{\code{\link{eidos_isInfinite}}}
\item{\code{\link{eidos_isNAN}}}
\item{\code{\link{eidos_log}}}
\item{\code{\link{eidos_log10}}}
\item{\code{\link{eidos_log2}}}
\item{\code{\link{eidos_product}}}
\item{\code{\link{eidos_round}}}
\item{\code{\link{eidos_setDifference}}}
\item{\code{\link{eidos_setIntersection}}}
\item{\code{\link{eidos_setSymmetricDifference}}}
\item{\code{\link{eidos_setUnion}}}
\item{\code{\link{eidos_sin}}}
\item{\code{\link{eidos_sqrt}}}
\item{\code{\link{eidos_sum}}}
\item{\code{\link{eidos_sumExact}}}
\item{\code{\link{eidos_tan}}}
\item{\code{\link{eidos_trunc}}}
\item{\code{\link{eidos_cor}}}
\item{\code{\link{eidos_cov}}}
\item{\code{\link{eidos_max}}}
\item{\code{\link{eidos_mean}}}
\item{\code{\link{eidos_min}}}
\item{\code{\link{eidos_pmax}}}
\item{\code{\link{eidos_pmin}}}
\item{\code{\link{eidos_quantile}}}
\item{\code{\link{eidos_range}}}
\item{\code{\link{eidos_rank}}}
\item{\code{\link{eidos_sd}}}
\item{\code{\link{eidos_ttest}}}
\item{\code{\link{eidos_var}}}
\item{\code{\link{eidos_dmvnorm}}}
\item{\code{\link{eidos_dbeta}}}
\item{\code{\link{eidos_dexp}}}
\item{\code{\link{eidos_dgamma}}}
\item{\code{\link{eidos_dnorm}}}
\item{\code{\link{eidos_findInterval}}}
\item{\code{\link{eidos_pnorm}}}
\item{\code{\link{eidos_qnorm}}}
\item{\code{\link{eidos_rbeta}}}
\item{\code{\link{eidos_rbinom}}}
\item{\code{\link{eidos_rcauchy}}}
\item{\code{\link{eidos_rdunif}}}
\item{\code{\link{eidos_rexp}}}
\item{\code{\link{eidos_rf}}}
\item{\code{\link{eidos_rgamma}}}
\item{\code{\link{eidos_rgeom}}}
\item{\code{\link{eidos_rlnorm}}}
\item{\code{\link{eidos_rmvnorm}}}
\item{\code{\link{eidos_rnbinom}}}
\item{\code{\link{eidos_rnorm}}}
\item{\code{\link{eidos_rpois}}}
\item{\code{\link{eidos_runif}}}
\item{\code{\link{eidos_rweibull}}}
\item{\code{\link{eidos_c}}}
\item{\code{\link{eidos_float}}}
\item{\code{\link{eidos_integer}}}
\item{\code{\link{eidos_logical}}}
\item{\code{\link{eidos_object}}}
\item{\code{\link{eidos_rep}}}
\item{\code{\link{eidos_repEach}}}
\item{\code{\link{eidos_sample}}}
\item{\code{\link{eidos_seq}}}
\item{\code{\link{eidos_seqAlong}}}
\item{\code{\link{eidos_seqLen}}}
\item{\code{\link{eidos_string}}}
\item{\code{\link{eidos_all}}}
\item{\code{\link{eidos_any}}}
\item{\code{\link{eidos_cat}}}
\item{\code{\link{eidos_catn}}}
\item{\code{\link{eidos_format}}}
\item{\code{\link{eidos_identical}}}
\item{\code{\link{eidos_ifelse}}}
\item{\code{\link{eidos_length}}}
\item{\code{\link{eidos_match}}}
\item{\code{\link{eidos_order}}}
\item{\code{\link{eidos_paste}}}
\item{\code{\link{eidos_paste0}}}
\item{\code{\link{eidos_print}}}
\item{\code{\link{eidos_rev}}}
\item{\code{\link{eidos_size}}}
\item{\code{\link{eidos_sort}}}
\item{\code{\link{eidos_sortBy}}}
\item{\code{\link{eidos_str}}}
\item{\code{\link{eidos_tabulate}}}
\item{\code{\link{eidos_unique}}}
\item{\code{\link{eidos_which}}}
\item{\code{\link{eidos_whichMax}}}
\item{\code{\link{eidos_whichMin}}}
\item{\code{\link{eidos_asFloat}}}
\item{\code{\link{eidos_asInteger}}}
\item{\code{\link{eidos_asLogical}}}
\item{\code{\link{eidos_asString}}}
\item{\code{\link{eidos_elementType}}}
\item{\code{\link{eidos_isFloat}}}
\item{\code{\link{eidos_isInteger}}}
\item{\code{\link{eidos_isLogical}}}
\item{\code{\link{eidos_isNULL}}}
\item{\code{\link{eidos_isObject}}}
\item{\code{\link{eidos_isString}}}
\item{\code{\link{eidos_type}}}
\item{\code{\link{eidos_nchar}}}
\item{\code{\link{eidos_strcontains}}}
\item{\code{\link{eidos_strfind}}}
\item{\code{\link{eidos_strprefix}}}
\item{\code{\link{eidos_strsplit}}}
\item{\code{\link{eidos_strsuffix}}}
\item{\code{\link{eidos_substr}}}
\item{\code{\link{eidos_apply}}}
\item{\code{\link{eidos_array}}}
\item{\code{\link{eidos_cbind}}}
\item{\code{\link{eidos_diag}}}
\item{\code{\link{eidos_dim}}}
\item{\code{\link{eidos_drop}}}
\item{\code{\link{eidos_lowerTri}}}
\item{\code{\link{eidos_matrix}}}
\item{\code{\link{eidos_matrixMult}}}
\item{\code{\link{eidos_ncol}}}
\item{\code{\link{eidos_nrow}}}
\item{\code{\link{eidos_rbind}}}
\item{\code{\link{eidos_t}}}
\item{\code{\link{eidos_upperTri}}}
\item{\code{\link{eidos_createDirectory}}}
\item{\code{\link{eidos_deleteFile}}}
\item{\code{\link{eidos_fileExists}}}
\item{\code{\link{eidos_filesAtPath}}}
\item{\code{\link{eidos_flushFile}}}
\item{\code{\link{eidos_getwd}}}
\item{\code{\link{eidos_readCSV}}}
\item{\code{\link{eidos_readFile}}}
\item{\code{\link{eidos_setwd}}}
\item{\code{\link{eidos_tempdir}}}
\item{\code{\link{eidos_writeFile}}}
\item{\code{\link{eidos_writeTempFile}}}
\item{\code{\link{eidos_cmColors}}}
\item{\code{\link{eidos_colors}}}
\item{\code{\link{eidos_color2rgb}}}
\item{\code{\link{eidos_heatColors}}}
\item{\code{\link{eidos_hsv2rgb}}}
\item{\code{\link{eidos_rainbow}}}
\item{\code{\link{eidos_rgb2color}}}
\item{\code{\link{eidos_rgb2hsv}}}
\item{\code{\link{eidos_terrainColors}}}
\item{\code{\link{eidos_assert}}}
\item{\code{\link{eidos_beep}}}
\item{\code{\link{eidos_citation}}}
\item{\code{\link{eidos_clock}}}
\item{\code{\link{eidos_date}}}
\item{\code{\link{eidos_debugIndent}}}
\item{\code{\link{eidos_defineConstant}}}
\item{\code{\link{eidos_defineGlobal}}}
\item{\code{\link{eidos_exists}}}
\item{\code{\link{eidos_functionSignature}}}
\item{\code{\link{eidos_functionSource}}}
\item{\code{\link{eidos_getSeed}}}
\item{\code{\link{eidos_license}}}
\item{\code{\link{eidos_ls}}}
\item{\code{\link{eidos_rm}}}
\item{\code{\link{eidos_sapply}}}
\item{\code{\link{eidos_setSeed}}}
\item{\code{\link{eidos_source}}}
\item{\code{\link{eidos_stop}}}
\item{\code{\link{eidos_suppressWarnings}}}
\item{\code{\link{eidos_sysinfo}}}
\item{\code{\link{eidos_system}}}
\item{\code{\link{eidos_time}}}
\item{\code{\link{eidos_usage}}}
\item{\code{\link{eidos_version}}}
}
This class has the following properties:
\describe{
\item{None.}{This class has no properties.}
}
}
\seealso{
Other Eidos: 
\code{\link{eidos_abs}()},
\code{\link{eidos_acos}()},
\code{\link{eidos_all}()},
\code{\link{eidos_any}()},
\code{\link{eidos_apply}()},
\code{\link{eidos_array}()},
\code{\link{eidos_asFloat}()},
\code{\link{eidos_asInteger}()},
\code{\link{eidos_asLogical}()},
\code{\link{eidos_asString}()},
\code{\link{eidos_asin}()},
\code{\link{eidos_assert}()},
\code{\link{eidos_atan}()},
\code{\link{eidos_atan2}()},
\code{\link{eidos_beep}()},
\code{\link{eidos_c}()},
\code{\link{eidos_cat}()},
\code{\link{eidos_catn}()},
\code{\link{eidos_cbind}()},
\code{\link{eidos_ceil}()},
\code{\link{eidos_citation}()},
\code{\link{eidos_clock}()},
\code{\link{eidos_cmColors}()},
\code{\link{eidos_color2rgb}()},
\code{\link{eidos_colors}()},
\code{\link{eidos_cor}()},
\code{\link{eidos_cos}()},
\code{\link{eidos_cov}()},
\code{\link{eidos_createDirectory}()},
\code{\link{eidos_cumProduct}()},
\code{\link{eidos_cumSum}()},
\code{\link{eidos_date}()},
\code{\link{eidos_dbeta}()},
\code{\link{eidos_debugIndent}()},
\code{\link{eidos_defineConstant}()},
\code{\link{eidos_defineGlobal}()},
\code{\link{eidos_deleteFile}()},
\code{\link{eidos_dexp}()},
\code{\link{eidos_dgamma}()},
\code{\link{eidos_diag}()},
\code{\link{eidos_dim}()},
\code{\link{eidos_dmvnorm}()},
\code{\link{eidos_dnorm}()},
\code{\link{eidos_drop}()},
\code{\link{eidos_elementType}()},
\code{\link{eidos_exists}()},
\code{\link{eidos_exp}()},
\code{\link{eidos_fileExists}()},
\code{\link{eidos_filesAtPath}()},
\code{\link{eidos_findInterval}()},
\code{\link{eidos_float}()},
\code{\link{eidos_floor}()},
\code{\link{eidos_flushFile}()},
\code{\link{eidos_format}()},
\code{\link{eidos_functionSignature}()},
\code{\link{eidos_functionSource}()},
\code{\link{eidos_getSeed}()},
\code{\link{eidos_getwd}()},
\code{\link{eidos_heatColors}()},
\code{\link{eidos_hsv2rgb}()},
\code{\link{eidos_identical}()},
\code{\link{eidos_ifelse}()},
\code{\link{eidos_integer}()},
\code{\link{eidos_integerDiv}()},
\code{\link{eidos_integerMod}()},
\code{\link{eidos_isFinite}()},
\code{\link{eidos_isFloat}()},
\code{\link{eidos_isInfinite}()},
\code{\link{eidos_isInteger}()},
\code{\link{eidos_isLogical}()},
\code{\link{eidos_isNAN}()},
\code{\link{eidos_isNULL}()},
\code{\link{eidos_isObject}()},
\code{\link{eidos_isString}()},
\code{\link{eidos_length}()},
\code{\link{eidos_license}()},
\code{\link{eidos_log}()},
\code{\link{eidos_log10}()},
\code{\link{eidos_log2}()},
\code{\link{eidos_logical}()},
\code{\link{eidos_lowerTri}()},
\code{\link{eidos_ls}()},
\code{\link{eidos_match}()},
\code{\link{eidos_matrix}()},
\code{\link{eidos_matrixMult}()},
\code{\link{eidos_max}()},
\code{\link{eidos_mean}()},
\code{\link{eidos_min}()},
\code{\link{eidos_nchar}()},
\code{\link{eidos_ncol}()},
\code{\link{eidos_nrow}()},
\code{\link{eidos_object}()},
\code{\link{eidos_order}()},
\code{\link{eidos_paste}()},
\code{\link{eidos_paste0}()},
\code{\link{eidos_pmax}()},
\code{\link{eidos_pmin}()},
\code{\link{eidos_pnorm}()},
\code{\link{eidos_print}()},
\code{\link{eidos_product}()},
\code{\link{eidos_qnorm}()},
\code{\link{eidos_quantile}()},
\code{\link{eidos_rainbow}()},
\code{\link{eidos_range}()},
\code{\link{eidos_rank}()},
\code{\link{eidos_rbeta}()},
\code{\link{eidos_rbind}()},
\code{\link{eidos_rbinom}()},
\code{\link{eidos_rcauchy}()},
\code{\link{eidos_rdunif}()},
\code{\link{eidos_readCSV}()},
\code{\link{eidos_readFile}()},
\code{\link{eidos_rep}()},
\code{\link{eidos_repEach}()},
\code{\link{eidos_rev}()},
\code{\link{eidos_rexp}()},
\code{\link{eidos_rf}()},
\code{\link{eidos_rgamma}()},
\code{\link{eidos_rgb2color}()},
\code{\link{eidos_rgb2hsv}()},
\code{\link{eidos_rgeom}()},
\code{\link{eidos_rlnorm}()},
\code{\link{eidos_rm}()},
\code{\link{eidos_rmvnorm}()},
\code{\link{eidos_rnbinom}()},
\code{\link{eidos_rnorm}()},
\code{\link{eidos_round}()},
\code{\link{eidos_rpois}()},
\code{\link{eidos_runif}()},
\code{\link{eidos_rweibull}()},
\code{\link{eidos_sample}()},
\code{\link{eidos_sapply}()},
\code{\link{eidos_sd}()},
\code{\link{eidos_seq}()},
\code{\link{eidos_seqAlong}()},
\code{\link{eidos_seqLen}()},
\code{\link{eidos_setDifference}()},
\code{\link{eidos_setIntersection}()},
\code{\link{eidos_setSeed}()},
\code{\link{eidos_setSymmetricDifference}()},
\code{\link{eidos_setUnion}()},
\code{\link{eidos_setwd}()},
\code{\link{eidos_sin}()},
\code{\link{eidos_size}()},
\code{\link{eidos_sort}()},
\code{\link{eidos_sortBy}()},
\code{\link{eidos_source}()},
\code{\link{eidos_sqrt}()},
\code{\link{eidos_stop}()},
\code{\link{eidos_str}()},
\code{\link{eidos_strcontains}()},
\code{\link{eidos_strfind}()},
\code{\link{eidos_string}()},
\code{\link{eidos_strprefix}()},
\code{\link{eidos_strsplit}()},
\code{\link{eidos_strsuffix}()},
\code{\link{eidos_substr}()},
\code{\link{eidos_sum}()},
\code{\link{eidos_sumExact}()},
\code{\link{eidos_suppressWarnings}()},
\code{\link{eidos_sysinfo}()},
\code{\link{eidos_system}()},
\code{\link{eidos_t}()},
\code{\link{eidos_tabulate}()},
\code{\link{eidos_tan}()},
\code{\link{eidos_tempdir}()},
\code{\link{eidos_terrainColors}()},
\code{\link{eidos_time}()},
\code{\link{eidos_trunc}()},
\code{\link{eidos_ttest}()},
\code{\link{eidos_type}()},
\code{\link{eidos_unique}()},
\code{\link{eidos_upperTri}()},
\code{\link{eidos_usage}()},
\code{\link{eidos_var}()},
\code{\link{eidos_version}()},
\code{\link{eidos_which}()},
\code{\link{eidos_whichMax}()},
\code{\link{eidos_whichMin}()},
\code{\link{eidos_writeFile}()},
\code{\link{eidos_writeTempFile}()}
}
\concept{Eidos}
