% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_script.R
\name{end_gen}
\alias{end_gen}
\alias{end_gen<-}
\title{Extract or set end generation}
\usage{
end_gen(x)

end_gen(x) <- value
}
\arguments{
\item{x}{A \code{slimr_script} object}

\item{value}{A end generation value to replace with.}
}
\description{
Extract or set end generation
}
\examples{
script <- slim_script(
  slim_block_init_minimal(),
  slim_block(1, 100, {
    sim.outputFull()
  })
)
script
end_gen(script)
end_gen(script)[2] <- 1000
script
}
