% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{extractProteinData}
\alias{extractProteinData}
\title{extractProteinData}
\usage{
extractProteinData(iupredPath, uni, workingDirectory = getwd(),
  overwrite = FALSE, saveAsRDS = FALSE)
}
\arguments{
\item{iupredPath}{The path to the IUPred Disorder score prediction tool
executable}

\item{uni}{The uniprot accession of the protein for which to extract data
(e.g. P04637)}

\item{workingDirectory}{The name of the folder to write downloaded/processed
files and RDS objects}

\item{overwrite}{TRUE/FALSE (default: FALSE). Boolean to decide if downloaded
files should overwrite existing files}

\item{saveAsRDS}{TRUE/FALSE (default: FALSE). Boolean to decide if all
extracted data should be saved as an RDS file in the working directory
under folder named 'RDS'}
}
\value{
A list object with variable data types directory
}
\description{
Given a uniprot accession, extract and process all available sequence
features including: 1. Fasta sequences 2. Disorder predictions using IUPred
3. Uniprot feature files downloaded in gff format 4. Uniprot variants
including disease-causing and polymorphic substitutions 5. Short linear
motifs - all substrings matching all available patterns 6. SLiM Changes - The
collection of changes in the proteins' SLiM content when variants at 4) are
applied to the sequence
}

