% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{fill_lost}
\alias{fill_lost}
\title{Fill in frequencies of alleles lost during the simulation (to prevent a SFS
to be "skewed" at the end of the simulation).}
\usage{
fill_lost(sim_sites, all_sites)
}
\arguments{
\item{sim_sites}{GRanges object of simulated sites.}

\item{all_sites}{GRanges object of all sites at the start of a simulation.}
}
\value{
GRanges object with frequencies of lost sites set to zero. Otherwise
  contains the same data as given by the sim_sites argument.
}
\description{
Fill in frequencies of alleles lost during the simulation (to prevent a SFS
to be "skewed" at the end of the simulation).
}
