% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{findModifiedSlims}
\alias{findModifiedSlims}
\title{findModifiedSlims}
\usage{
findModifiedSlims(regex)
}
\arguments{
\item{regex}{A regular expression pattern from ELM database}
}
\description{
Find out if a given slim pattern is only functional if it is
post-translationally modified
}
\details{
Modified residues are denoted within round brackets according to the
convention employed by the ELM annotations e.g. The regex "S..([ST]).."
suggests that [ST] residue needs to be modified for the function of this
slim. However, round brackets are also used as part of universal regexes to
group items. E.g.: motifRegex$TRG_ER_diArg_1 =>
"([LIVMFYWPR]R[^YFWDE]{0,1}R)|(R[^YFWDE]{0,1}R[LIVMFYWPR])' So, we need a
function to distinguish the usual usage of round brackets from those that
suggest a modified residue.
}
