% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifAnalysis.R
\name{findMotifChanges}
\alias{findMotifChanges}
\title{findMotifKnockIns}
\usage{
findMotifChanges(sequence, variants, motifRegex)
}
\arguments{
\item{sequence}{A character string of amino acid sequence}

\item{variants}{A data.frame consisting of minimum three columns: 1.wtAA,
2.mutAA, 3.pos where pos is the mutation position in the sequence, wtAA is
the wild-type amino acid in the sequence and mutAA is the mutant amino
acid.}
}
\description{
Find out which SLiMs are created de novo via point amino acid substitutions
in protein sequences
}

