% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifAnalysis.R
\name{findMotifChangesBulk}
\alias{findMotifChangesBulk}
\title{findMotifChangesBulk}
\usage{
findMotifChangesBulk(variants, uniprotDataDir)
}
\arguments{
\item{variants}{A Granges object of variants parsed from Humsavar
using getHumSavar() function or a subset of it with the same structure
  consisting of minimum two meta-data columns: 1.wtAA,
  2.mutAA, wtAA is the wild-type amino acid (one letter code) in the
  sequence and mutAA is the mutant amino acid (one letter code).}

\item{uniprotDataDir}{The folder that is used to download/store uniprot
data files}
}
\description{
Find out which SLiMs are gained or lost (no longer matching the regex
pattern) via point amino acid substitutions in protein sequences
}

