% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifAnalysis.R
\name{findMotifChangesMulti}
\alias{findMotifChangesMulti}
\title{findMotifChangesMulti}
\usage{
findMotifChangesMulti(sequences, variants, motifRegex = slimR::motifRegex,
  nodeN = 1)
}
\arguments{
\item{sequences}{List of strings where the strings are amino-acid sequences
and the list names are uniprot accessions}

\item{variants}{A data.frame consisting of minimum four columns: 1.
uniprotAccession 2.wtAA, 3.mutAA, 4.pos where pos is the mutation position
in the sequence, wtAA is the wild-type amino acid (one letter code) in the
sequence and mutAA is the mutant amino acid (one letter code).}

\item{motifRegex}{List of slim regular expressions. By default, the built-in
slimR::motifRegex from the ELM database is used.}

\item{nodeN}{Number of cores needed to run the analysis (default: 1)}
}
\value{
List of data.frame objects. One data.frame per each uniprot accession
}
\description{
A wrapper function that runs findMotifChanges function for multiple inputs
using multiple cores
}
\examples{
sequences <- downloadUniprotFiles(uniprotAccessions = c('P04637', 'P11166'),
                         format = 'fasta', overwrite = 'FALSE', nodeN = 2)
variants <- slimR::glutMutations
motifRegex <- list('mymotif' = '.ABSDBASDBAS.')
motifChanges <- findMotifChangesMulti(sequences = sequences,
                                     variants = variants,
                                     motifRegex = motifRegex,
                                     nodeN = 1)
}
