% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_callbacks.R
\name{fitness}
\alias{fitness}
\title{SLiM fitness() callback}
\usage{
fitness(mut_type_id, subpop_id)
}
\arguments{
\item{mut_type_id}{The id of the mutationType to which this callback should apply. Can be an
integer 1, 2, etc., or character "m1", "m2", etc.}

\item{subpop_id}{The id(s) of the subpopulation(s) to which this callback should apply. Can
be an integer 1, 2, etc., or character "p1", "p2", etc.}
}
\value{
None
}
\description{
This callback specifies that a code block is providing logic to decide the fitness of a mutation in an individual.
The callback should return a the relative fitness value of the mutation, where a value of 1.0 is neutral. This is
called once per mutation per individual per generation in which it is active. If \code{mut_id_type = NULL} then
the callback will be called only once per individual and should return a global relative fitness value that is
independent of mutations the individual possesses.
see \href{http://benhaller.com/slim/SLiM_Manual.pdf#page=596}{SLiM Manual: page 596}
}
\details{
Global variables available in reproduction callbacks:
\describe{
  \item{mut}{A Mutation object, the mutation whose relative fitness is being evaluated}
  \item{homozygous}{A value of T (the mutation is homozygous), F (heterozygous), or NULL (it is
  paired with a null chromosome, which can occur with sex chromosomes)}
  \item{relFitness}{The default relative fitness value calculated by SLiM}
  \item{individual}{The individual carrying this mutation (an object of class Individual)}
  \item{genome1}{One genome of the individual carrying this mutation}
  \item{genome2}{The other genome of that individual}
  \item{subpop}{The subpopulation in which that individual lives}
}
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
modified from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016-2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\examples{
slim_block(fitness(m2), {
  ## sets up frequency dependent selection for m2
  return(1.5 - sim.mutationFrequencies(p1, mut))
})
}
\seealso{
Other callbacks: 
\code{\link{early}()},
\code{\link{first}()},
\code{\link{fitnessEffect}()},
\code{\link{initialize}()},
\code{\link{interaction}()},
\code{\link{late}()},
\code{\link{mateChoice}()},
\code{\link{modifyChild}()},
\code{\link{mutationEffect}()},
\code{\link{mutation}()},
\code{\link{recombination}()},
\code{\link{reproduction}()},
\code{\link{slim_callbacks}()},
\code{\link{survival}()}
}
\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
\concept{callbacks}
