% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_callbacks.R
\name{fitnessEffect}
\alias{fitnessEffect}
\title{SLiM fitnessEffect() callback}
\usage{
fitnessEffect(subpop_id)
}
\arguments{
\item{subpop_id}{The id(s) of the subpopulation(s) to which this callback should apply. Can
be an integer 1, 2, etc., or character "p1", "p2", etc.}
}
\value{
None
}
\description{
A fitnessEffect() callback is called by SLiM when it is determining the fitness of an individual
– typically, but not always, once per tick during the fitness calculation tick cycle stage. Normally,
the fitness of a given individual is determined by multiplying together the fitness effects of all
mutations possessed by that individual (see section 25.2 for further discussion). Supplying a
fitnessEffect() callback allows you to add another multiplicative fitness effect into that
calculation. As with mutationEffect() callbacks, the value returned by fitnessEffect()
callbacks is a fitness effect, so 1.0 is neutral. For details
see \href{http://benhaller.com/slim/SLiM_Manual.pdf#page=715}{SLiM Manual: page 715}
}
\details{
Global variables available in reproduction callbacks:
\describe{
  \item{individual}{The individual carrying this mutation (an object of class Individual)}
  \item{subpop}{The subpopulation in which that individual lives}
}
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
modified from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016-2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\examples{
slim_block(fitnessEffect(p3), {
  # multiplies all individual's fitness by 0.75
  return(0.75)
})
}
\seealso{
Other callbacks: 
\code{\link{early}()},
\code{\link{first}()},
\code{\link{fitness}()},
\code{\link{initialize}()},
\code{\link{interaction}()},
\code{\link{late}()},
\code{\link{mateChoice}()},
\code{\link{modifyChild}()},
\code{\link{mutationEffect}()},
\code{\link{mutation}()},
\code{\link{recombination}()},
\code{\link{reproduction}()},
\code{\link{slim_callbacks}()},
\code{\link{survival}()}
}
\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
\concept{callbacks}
