% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\docType{data}
\name{Genome}
\alias{Genome}
\alias{.G}
\title{Genome}
\format{
An object of class \code{list} of length 22.
}
\usage{
Genome
}
\description{
Documentation for Genome class from SLiM
}
\details{
This class has the following methods (functions):
\itemize{
\item{\code{\link{addMutations}}}
\item{\code{\link{addNewDrawnMutation}}}
\item{\code{\link{addNewMutation}}}
\item{\code{\link{containsMarkerMutation}}}
\item{\code{\link{containsMutations}}}
\item{\code{\link{countOfMutationsOfType}}}
\item{\code{\link{mutationsOfType}}}
\item{\code{\link{nucleotides}}}
\item{\code{\link{output}}}
\item{\code{\link{outputMS}}}
\item{\code{\link{outputVCF}}}
\item{\code{\link{positionsOfMutationsOfType}}}
\item{\code{\link{readFromMS}}}
\item{\code{\link{readFromVCF}}}
\item{\code{\link{removeMutations}}}
\item{\code{\link{sumOfMutationsOfType}}}
}
This class has the following properties:
\describe{
\item{genomePedigreeID}{A property of type integer. It is of length one
(a singleton). This property is a constant, so it is not modifiable.
\strong{Property Description:} If pedigree tracking is turned on with
initializeSLiMOptions(keepPedigrees=T), genomePedigreeID is a unique non-
negative identifier for each genome in a simulation, never reused throughout
the duration of the simulation run. Furthermore, the genomePedigreeID of a
given genome will be equal to either (2*pedigreeID) or (2*pedigreeID + 1) of the
individual that the genome belongs to (the former for the first genome of the
individual, the latter for the second genome of the individual); this invariant
relationship is guaranteed. This property is also available when treesequence
recording is enabled. If neither pedigree tracking nor tree-sequence recording
is enabled, this property is unavailable. }
\item{genomeType}{A property of type string. It is of length one (a singleton).
This property is a constant, so it is not modifiable. \strong{Property
Description:} The type of chromosome represented by this genome; one of "A",
"X", or "Y". }
\item{individual}{A property of type Individual object. It is of length
one (a singleton). This property is a constant, so it is not modifiable.
\strong{Property Description:} The Individual object to which this genome
belongs. }
\item{isNullGenome}{A property of type logical. It is of length one
(a singleton). This property is a constant, so it is not modifiable.
\strong{Property Description:} T if the genome is a "null” genome, F if it is
an ordinary genome object. When a sex chromosome (X or Y) is simulated, the
other sex chromosome also exists in the simulation, but it is a "null” genome
that does not carry any mutations. Instead, it is a placeholder, present to
allow SLiM’s code to operate in much the same way as it does when an autosome is
simulated. Null genomes should not be accessed or manipulated. }
\item{mutations}{A property of type Mutation object. This property is a
constant, so it is not modifiable. \strong{Property Description:} All of the
Mutation objects present in this genome. }
\item{tag}{A property of type integer. It is of length one (a singleton). This
property is a variable, so it is modifiable. \strong{Property Description:}
A user-defined integer value. The value of tag is initially undefined, and
it is an error to try to read it; if you wish it to have a defined value, you
must arrange that yourself by explicitly setting its value prior to using it
elsewhere in your code. The value of tag is not used by SLiM; it is free for you
to use. Note that the Genome objects used by SLiM are new with every generation,
so the tag value of each new offspring generated in each generation will be
initially undefined. If you set a tag value for an offspring genome inside
a modifyChild() callback, that tag value will be preserved as the offspring
individual becomes a parent (across the generation boundary, in other words).
If you take advantage of this, however, you should be careful to set up initial
values for the tag values of all offspring, otherwise undefined initial values
might happen to match the values that you are trying to use to tag particular
individuals. A rule of thumb in programming: undefined values should always be
assumed to take on the most inconvenient value possible.}

}
}
\keyword{datasets}
