% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{genomicElementTypesWithIDs}
\alias{genomicElementTypesWithIDs}
\alias{Community$genomicElementTypesWithIDs}
\alias{.Co$genomicElementTypesWithIDs}
\title{SLiM method genomicElementTypesWithIDs}
\usage{
genomicElementTypesWithIDs(ids)
}
\arguments{
\item{ids}{An object of type integer. See details for description.}
}
\value{
An object of type GenomicElementType object.
}
\description{
Documentation for SLiM function \code{genomicElementTypesWithIDs}, which is a
method of the SLiM class \code{\link{Community}}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Documentation for this function can be found in the official
\href{http://benhaller.com/slim/SLiM_Manual.pdf#page=637}{SLiM manual: page
637}.

Find and return the GenomicElementType objects with id values matching
the values in ids. If no matching GenomicElementType object can be found with a
given id, an error results.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016-2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\seealso{
Other Community: 
\code{\link{Co}},
\code{\link{createLogFile}()},
\code{\link{deregisterScriptBlock}()},
\code{\link{interactionTypesWithIDs}()},
\code{\link{mutationTypesWithIDs}()},
\code{\link{outputUsage}()},
\code{\link{registerEarlyEvent}()},
\code{\link{registerFirstEvent}()},
\code{\link{registerInteractionCallback}()},
\code{\link{registerLateEvent}()},
\code{\link{rescheduleScriptBlock}()},
\code{\link{scriptBlocksWithIDs}()},
\code{\link{simulationFinished}()},
\code{\link{speciesWithIDs}()},
\code{\link{subpopulationsWithIDs}()}
}
\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
\concept{Community}
