% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_script.R
\name{get_block}
\alias{get_block}
\title{Extract a single code block from a \code{slimr_script}}
\usage{
get_block(x, i)
}
\arguments{
\item{x}{\code{slimr_script} object to extract block from}

\item{i}{The block to extract. Can be either an integer specifying what block(s)
index to extract, or a character, in which case it pulls the block(s) with
the corresponding name.}
}
\description{
Extract a single code block from a \code{slimr_script}
}
\examples{
script <- slim_script(
  slim_block_init_minimal(),
  slim_block_finish(100)
)
get_block(script, "block_1")
get_block(script, 2)
}
