% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_output.R
\name{get_deserts}
\alias{get_deserts}
\title{Extract coordinates of deserts from a given VCF file.}
\usage{
get_deserts(markers, cutoff = 0)
}
\description{
Extract coordinates of deserts from a given VCF file.
}
