% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_utils.R
\name{get_gaps}
\alias{get_gaps}
\title{Download the coordinates of centromeres.
These are used for "splitting" admixture tracts that overlap a centromere
and would therefore appear extensively long.
http://rohsdb.usc.edu/GBshape/cgi-bin/hgTables?db=hg19&hgta_group=map&hgta_track=gap&hgta_table=gap&hgta_doSchema=describe+table+schema}
\usage{
get_gaps()
}
\description{
Download the coordinates of centromeres.
These are used for "splitting" admixture tracts that overlap a centromere
and would therefore appear extensively long.
http://rohsdb.usc.edu/GBshape/cgi-bin/hgTables?db=hg19&hgta_group=map&hgta_track=gap&hgta_table=gap&hgta_doSchema=describe+table+schema
}
