% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_output.R
\name{get_markers}
\alias{get_markers}
\title{Load the simulated Neanderthal fixed markers and their frequencies
(either all, or the ones within regions or gaps). Return as a GRanges
object.}
\usage{
get_markers(vcf, sites_coords, within_region = NULL, fill_freq = TRUE)
}
\description{
Load the simulated Neanderthal fixed markers and their frequencies
(either all, or the ones within regions or gaps). Return as a GRanges
object.
}
