% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_os}
\alias{get_os}
\title{Function to return the OS}
\usage{
get_os()
}
\value{
Lowercase character stating the OS: linux, windows, or osx
}
\description{
This function shamelessly stolen from: https://www.r-bloggers.com/identifying-the-os-from-r/
}
