% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_recipes.R
\name{get_recipe}
\alias{get_recipe}
\title{Get a SLiM recipe}
\usage{
get_recipe(recipe = 1)
}
\arguments{
\item{recipe}{Integer or Character specifying the recipe number, or the recipe name.}
}
\value{

}
\description{
Get a SLiM recipe
}
