% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_slim.R
\name{get_slim_call}
\alias{get_slim_call}
\title{Get the call information for running SLiM. Doubles as a check for SLiM availability.}
\usage{
get_slim_call()
}
\value{

}
\description{
Function to test if SLiM is available on user's system.
}
