% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariationData.R
\name{getClinVarData}
\alias{getClinVarData}
\title{getClinVarData}
\usage{
getClinVarData(url, overwrite = FALSE)
}
\arguments{
\item{url}{The url to the ftp location of the clinvar dataset
(e.g. ftp://ftp.ncbi.nlm.nih.gov/pub/clinvar/vcf_GRCh38/clinvar_20170404.vcf.gz)}
}
\value{
A tibble object extracted from the downloaded vcf file
}
\description{
This function will fetch the clinvar data from the given url and parse the contents
of the downloaded file.
}

