% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariationData.R
\name{getClinVarVEPmissenseVariants}
\alias{getClinVarVEPmissenseVariants}
\title{getClinVarVEPmissenseVariants}
\usage{

  getClinVarVEPmissenseVariants(vepPath = "/home/buyar/.local/bin/variant_effect_predictor.pl",
  clinvarDataURL, overwrite = FALSE, nodeN = 4)
}
\value{
A data.table object
}
\description{
This function downloads ClinVar variants and runs variant_effect_predictor
(VEP) tool and merges the clinvar variants with missense variants predicted
by VEP.
}
