% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getELMdomainInteractions}
\alias{getELMdomainInteractions}
\title{getELMdomainInteractions}
\usage{
getELMdomainInteractions()
}
\value{
A data.frame object with ELM classes versus cognate PFAM domains
}
\description{
Scrape known motif class - domain pairs from ELM
http://elm.eu.org/interactiondomains
}
\examples{

elm2pfam <- getELMdomainInteractions()

}
