% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getElmInstances}
\alias{getElmInstances}
\title{getElmInstances}
\usage{
getElmInstances(query = "*", instanceLogic = "true+positive",
  taxon = "homo+sapiens")
}
\arguments{
\item{query}{(default: * -all instances-) any query term to search the
database for instances}

\item{instanceLogic}{(default: true+positive) other options: true+negative,
false+positive, false+negative}

\item{taxon}{(default: homo+sapiens) other option examples: mus+musculus,
drosophila+melanogaster etc.}
}
\value{
A GRanges object where seqnames are Uniprot accessions and
ranges are the short linear motif locations in the proteins.
}
\description{
Using REST services, download ELM instances table from elm.eu.org
}
\examples{
elmInst <- getElmInstances('*', 'true+positive', 'homo+sapiens')
}
