% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getHumSavar}
\alias{getHumSavar}
\title{getHumSavar}
\usage{
getHumSavar(outdir)
}
\arguments{
\item{outdir}{Path to the folder, where the downloaded file should be saved.}
}
\value{
A Granges object containing the coordinates of mutated sites in
  proteins

A GRanges object where seqnames are Uniprot sequence accessions and
  ranges are the coordinates of the variants.
}
\description{
Download and parse protein mutation data from UniProt
}
\examples{
\dontrun{
hs <- getHumSavar(getwd())
}
}
