% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getUniprotData}
\alias{getUniprotData}
\title{getUniprotData}
\usage{
getUniprotData(format, reviewed = TRUE, organism = 9606,
  overwrite = FALSE)
}
\arguments{
\item{format}{File format. Accepted formats are gff, fasta, txt}

\item{reviewed}{Only download reviewed uniprot entries (default: TRUE)}

\item{organism}{Organism code e.g. 9606 for human}
}
\description{
Batch download different types of data from Uniprot using
REST services
}
