% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glutFasta}
\alias{glutFasta}
\title{GLUT1 protein fasta sequence}
\format{An object of class \code{AAStringSet} of length 1.}
\source{
Downloaded from uniprot \url{www.uniprot.org/uniprot/P11166.fasta}
}
\usage{
glutFasta
}
\value{
A AAStringSet object
}
\description{
The amino acid sequence for P11166 protein (GLUT1)
}
\keyword{datasets}

