% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glutMutations}
\alias{glutMutations}
\title{GLUT1 disease mutations from uniprot}
\format{An object of class \code{data.frame} with 43 rows and 4 columns.}
\source{
GLUT1 protein \url{www.uniprot.org/uniprot/P11166}
and humsavar \url{www.uniprot.org/docs/humsavar.txt}
}
\usage{
glutMutations
}
\value{
A data.frame object
}
\description{
Table of disease associated substitutions parsed from Humsavar (UniProt)
}
\keyword{datasets}
