% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_shims.R
\name{\%.\%}
\alias{\%.\%}
\title{Use this to replace "." operator in SLiM code}
\usage{
lhs \%.\% rhs
}
\arguments{
\item{lhs}{Left hand side}

\item{rhs}{Right hand side}
}
\value{
NULL (invisibly)
}
\description{
This is a simple shim used to allow SLiM-like code to be parsed by the R interpreter.
When writing SLiM code in R use this instead of "." to maintain code highlighting and
code completion. See \code{vignette("SLiM_IDE_tutorial")} for details on how to use this
in an R IDE to write SLiM code.
}
