% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%+\%}
\alias{\%+\%}
\title{Glue two strings together.}
\usage{
a \%+\% b
}
\arguments{
\item{a}{First string}

\item{b}{Second sring}
}
\value{
New string with a and b pasted together with a space in between.
}
\description{
Glue two strings together.
}
