% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%+n\%}
\alias{\%+n\%}
\title{Glue two string together with a newline separator}
\usage{
a \%+n\% b
}
\arguments{
\item{a}{First string}

\item{b}{Second sring}
}
\value{
New string with a and b pasted together with a newline in between.
}
\description{
Glue two string together with a newline separator
}
