% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{initializeHotspotMap}
\alias{initializeHotspotMap}
\alias{Initialize$initializeHotspotMap}
\alias{.Init$initializeHotspotMap}
\title{SLiM method initializeHotspotMap}
\usage{
initializeHotspotMap(multipliers, ends, sex)
}
\arguments{
\item{multipliers}{An object of type numeric. See details for description.}

\item{ends}{An object of type null or integer. The default value is \code{NULL}.
See details for description.}

\item{sex}{An object of type string. Must be of length 1 (a singleton). The
default value is \code{"*"}. See details for description.}
}
\value{
An object of type void.
}
\description{
Documentation for SLiM function \code{initializeHotspotMap}, which is a method
of the SLiM class \code{Initialize}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
In nucleotide-based models, set the mutation rate multiplier along the
chromosome. Nucleotidebased models define sequence-based mutation rates that
are set up with the mutationMatrix parameter to initializeGenomicElementType().
If no hotspot map is specified by calling initializeHotspotMap(), a hotspot
map with a multiplier of 1.0 across the whole chromosome is assumed (and so the
sequence-based rates are the absolute mutation rates used by SLiM). A hotspot
map modifies the sequence-based rates by scaling them up in some regions,
with multipliers greater than 1.0 (representing mutational hot spots), and/or
scaling them down in some regions, with multipliers less than 1.0 (representing
mutational cold spots). There are two ways to call this function. If the
optional ends parameter is NULL (the default), then multipliers must be a
singleton value that specifies a single multiplier to be used along the entire
chromosome (typically 1.0, but not required to be). If, on the other hand, ends
is supplied, then multipliers and ends must be the same length, and the values
in ends must be specified in ascending order. In that case, multipliers and ends
taken together specify the multipliers to be used along successive contiguous
stretches of the chromosome, from beginning to end; the last position specified
in ends should extend to the end of the chromosome (i.e. at least to the end of
the last genomic element, if not further). For example, if the following call is
made: initializeHotspotMap(c(1.0, 1.2), c(5000, 9999)); then the result is that
the mutation rate multiplier for bases 0...5000 (inclusive) will be 1.0 (and
so the specified sequence-based mutation rates will be used verbatim), and the
multiplier for bases 5001...9999 (inclusive) will be 1.2 (and so the sequence-
based mutation rates will be multiplied by 1.2 within the region). Note that
mutations are generated by SLiM only within genomic elements, regardless of the
hotspot map. In effect, the hotspot map given is intersected with the coverage
area of the genomic elements defined; areas outside of any genomic element are
given a multiplier of zero. There is no harm in supplying a hotspot map that
specifies multipliers for areas outside of the genomic elements defined; the
excess information is simply not used. If the optional sex parameter is "*" (the
default), then the supplied hotspot map will be used for both sexes (which is
the only option for hermaphroditic simulations). In sexual simulations sex may
be "M" or "F" instead, in which case the supplied hotspot map is used only for
that sex (i.e., when generating a gamete from a parent of that sex). In this
case, two calls must be made to initializeHotspotMap(), one for each sex, even
if a multiplier of 1.0 is desired for the other sex; no default hotspot map is
supplied.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016–2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
