% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{initializeMutationType}
\alias{initializeMutationType}
\alias{Initialize$initializeMutationType}
\alias{.Init$initializeMutationType}
\title{SLiM method initializeMutationType}
\usage{
initializeMutationType(id, dominanceCoeff, distributionType, ...)
}
\arguments{
\item{id}{An object of type integer or string. Must be of length 1 (a
singleton). See details for description.}

\item{dominanceCoeff}{An object of type numeric. Must be of length 1 (a
singleton). See details for description.}

\item{distributionType}{An object of type string. Must be of length 1 (a
singleton). See details for description.}

\item{...}{An object of type NA. NA See details for description.}
}
\value{
An object of type MutationType object. Return will be of length 1 (a
singleton)
}
\description{
Documentation for SLiM function \code{initializeMutationType}, which is a method
of the SLiM class \code{Initialize}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Add a mutation type at initialization time. The id must not already
be used for any mutation type in the simulation. The id parameter may be
either an integer giving the ID of the new mutation type, or a string giving
the name of the new mutation type (such as "m5" to specify an ID of 5). The
dominanceCoeff parameter supplies the dominance coefficient for the mutation
type; 0.0 produces no dominance, 1.0 complete dominance, and values greater
than 1.0, overdominance. The distributionType may be "f", in which case the
ellipsis ... should supply a numeric$ fixed selection coefficient; "e", in
which case the ellipsis should supply a numeric$ mean selection coefficient
for an exponential distribution; "g", in which case the ellipsis should supply
a numeric$ mean selection coefficient and a numeric$ alpha shape parameter for
a gamma distribution; "n", in which case the ellipsis should supply a numeric$
mean selection coefficient and a numeric$ sigma (standard deviation) parameter
for a normal distribution; "w", in which case the ellipsis should supply a
numeric$ λ scale parameter and a numeric$ k shape parameter for a Weibull
distribution; or "s", in which case the ellipsis should supply a string$ Eidos
script parameter. See section 23.9 for discussion of the various DFEs and their
uses. The global symbol for the new mutation type is immediately available; the
return value also provides the new object. Note that by default in WF models,
all mutations of a given mutation type will be converted into Substitution
objects when they reach fixation, for efficiency reasons. If you need to disable
this conversion, to keep mutations of a given type active in the simulation
even after they have fixed, you can do so by setting the convertToSubstitution
property of MutationType to T. In contrast, by default in nonWF models mutations
will not be converted into Substitution objects when they reach fixation;
convertToSubstitution is F by default in nonWF models. To enable conversion in
nonWF models for neutral mutation types with no indirect fitness effects, you
should therefore set convertToSubstitution to T. See sections 21.3, 22.5, and
23.9.1 for further discussion regarding the convertToSubstitution property.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016–2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
