% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{initializeRecombinationRate}
\alias{initializeRecombinationRate}
\alias{Initialize$initializeRecombinationRate}
\alias{.Init$initializeRecombinationRate}
\title{SLiM method initializeRecombinationRate}
\usage{
initializeRecombinationRate(rates, ends, sex)
}
\arguments{
\item{rates}{An object of type numeric. See details for description.}

\item{ends}{An object of type null or integer. The default value is \code{NULL}.
See details for description.}

\item{sex}{An object of type string. Must be of length 1 (a singleton). The
default value is \code{"*"}. See details for description.}
}
\value{
An object of type void.
}
\description{
Documentation for SLiM function \code{initializeRecombinationRate}, which is a
method of the SLiM class \code{Initialize}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Set the recombination rate per base position per generation along the
chromosome. To be precise, this recombination rate is the probability that a
breakpoint will occur between one base and the next base; note that this is
different from how the mutation rate is defined (see initializeMutationRate()).
All rates must be in the interval [0.0, 0.5]. A rate of 0.5 implies complete
independence between the adjacent bases, which might be used to implement
independent assortment of loci located on different chromosomes (see the
example below). Whether a breakpoint occurs between two bases is then, in
effect, determined by a binomial draw with a single trial and the given rate
as probability (but under the hood SLiM uses a mathematically equivalent but
much more efficient strategy). Unlike the mutational process in SLiM, then,
which can generate more than one mutation at a given site (in one generation/
genome), the recombinational process in SLiM will never generate more then
one crossover between one base and the next (in one generation/genome), and
a supplied rate of 0.5 will therefore result in an actual probability of 0.5
for a crossover at the relevant position. (Note that this was not true in SLiM
2.x and earlier, however; their implementation of recombination resulted in a
crossover probability of about 39.3% for a rate of 0.5, due to the use of an
inaccurate approximation method. Recombination rates lower than about 0.01 would
have been essentially exact, since the approximation error became large only
as the rate approached 0.5.) There are two ways to call this function. If the
optional ends parameter is NULL (the default), then rates must be a singleton
value that specifies a single recombination rate to be used along the entire
chromosome. If, on the other hand, ends is supplied, then rates and ends must
be the same length, and the values in ends must be specified in ascending
order. In that case, rates and ends taken together specify the recombination
rates to be used along successive contiguous stretches of the chromosome, from
beginning to end; the last position specified in ends should extend to the end
of the chromosome (i.e. at least to the end of the last genomic element, if
not further). Note that a recombination rate of 1 centimorgan/Mbp corresponds
to a recombination rate of 1e-8 in the units used by SLiM. For example, if the
following call is made: initializeRecombinationRate(c(0, 0.5, 0), c(5000, 5001,
9999)); then the result is that the recombination rates between bases 0 / 1,
1 / 2, ..., 4999 / 5000 will be 0, the rate between bases 5000 / 5001 will be
0.5, and the rate between bases 5001 / 5002 onward (up to 9998 / 9999) will
again be 0. Setting the recombination rate between one specific pair of bases
to 0.5 forces recombination to occur with a probability of 0.5 between those
bases, which effectively breaks the simulated locus into separate chromosomes
at that point; this example effectively has one simulated chromosome from base
position 0 to 5000, and another from 5001 to 9999. If the optional sex parameter
is "*" (the default), then the supplied recombination rate map will be used
for both sexes (which is the only option for hermaphroditic simulations). In
sexual simulations sex may be "M" or "F" instead, in which case the supplied
recombination map is used only for that sex. In this case, two calls must be
made to initializeRecombinationRate(), one for each sex, even if a rate of zero
is desired for the other sex; no default recombination map is supplied.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016–2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
