% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{initializeSpecies}
\alias{initializeSpecies}
\alias{Initialize$initializeSpecies}
\alias{.Init$initializeSpecies}
\title{SLiM method initializeSpecies}
\usage{
initializeSpecies(tickModulo, tickPhase, avatar, color)
}
\arguments{
\item{tickModulo}{An object of type integer or integer or string or string. Must
be of length 1 (a singleton). The default value is \code{1}. See details for
description.}

\item{tickPhase}{An object of type integer or integer or string or string. Must
be of length 1 (a singleton). The default value is \code{1}. See details for
description.}

\item{avatar}{An object of type integer or integer or string or string. Must
be of length 1 (a singleton). The default value is \code{""}. See details for
description.}

\item{color}{An object of type integer or integer or string or string. Must
be of length 1 (a singleton). The default value is \code{""}. See details for
description.}
}
\value{
An object of type void.
}
\description{
Documentation for SLiM function \code{initializeSpecies}, which is a method of
the SLiM class \code{\link{Initialize}}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Documentation for this function can be found in the official
\href{http://benhaller.com/slim/SLiM_Manual.pdf#page=656}{SLiM manual: page
656}.

Configure options for the species being initialized. This
initialization function may only be called in multispecies models (i.e.,
models with explicit species declarations); in single-species models, the
default values are assumed and cannot be changed. The tickModulo and tickPhase
parameters determine the activation schedule for the species. The active
property of the species will be set to T (thus activating the species) every
tickModulo ticks, beginning in tick tickPhase. (However, when the species
is activated in a given tick, the skipTick() method may still be called in a
first() event to deactivate it.) See the active property of Species (section
25.15.1) for more details. The avatar parameter, if not "", sets a string value
used to represent the species graphically, particularly in SLiMgui but perhaps
in other contexts also. The avatar should generally be a single character -
usually an emoji corresponding to the species, such as "🦊 " for foxes or "🐭
" for mice. If avatar is the empty string, "", SLiMgui will choose a default
avatar. The color parameter, if not "", sets a string color value used to
represent the species in SLiMgui. Colors may be specified by name, or with
hexadecimal RGB values of the form "#RRGGBB" (see the Eidos manual for details).
If color is the empty string, "", SLiMgui will choose a default color.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016-2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\seealso{
Other Initialize: 
\code{\link{Init}},
\code{\link{initializeAncestralNucleotides}()},
\code{\link{initializeGeneConversion}()},
\code{\link{initializeGenomicElementType}()},
\code{\link{initializeGenomicElement}()},
\code{\link{initializeHotspotMap}()},
\code{\link{initializeInteractionType}()},
\code{\link{initializeMutationRate}()},
\code{\link{initializeMutationTypeNuc}()},
\code{\link{initializeMutationType}()},
\code{\link{initializeRecombinationRate}()},
\code{\link{initializeSLiMModelType}()},
\code{\link{initializeSLiMOptions}()},
\code{\link{initializeSex}()},
\code{\link{initializeTreeSeq}()}
}
\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
\concept{Initialize}
