% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_callbacks.R
\name{interaction}
\alias{interaction}
\title{SLiM interaction() callback}
\usage{
interaction(int_type_id, subpop_id)
}
\arguments{
\item{int_type_id}{The id of the InteractionType to apply this callback to. Can
be an integer 1, 2, etc., or character "i1", "i2", etc.}

\item{subpop_id}{The id(s) of the subpopulation(s) to which this callback should apply. Can
be an integer 1, 2, etc., or character "p1", "p2", etc.}
}
\value{
None
}
\description{
This callback specifies that a code block is providing logic to determine the
strength of interaction between individuals. It should return the strength of the interaction
as a numeric value. You must explicitly use \code{return(value)} at the end of the code block.
For more information on how to use \code{interaction()} callback
see \href{http://benhaller.com/slim/SLiM_Manual.pdf#page=604}{SLiM Manual: page 606}
}
\details{
Global variables available in reproduction callbacks:
\describe{
  \item{distance}{The distance from receiver to exerter, in spatial simulations; NAN otherwise}
  \item{strength}{The default interaction strength calculated by the interaction function}
  \item{receiver}{The individual receiving the interaction (an object of class Individual)}
  \item{exerter}{The individual exerting the interaction (an object of class Individual)}
  \item{subpop}{The subpopulation in which the receiver and exerter live}
}
}
\examples{
slim_block(interaction(), {
  # custom interaction strength as a function of values held in exerter and receiver tagF element.
  return(dnorm(exerter.tagF, receiver.tagF, 0.1) / dnorm(0, 0, 0.1))
})
}
