% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_herm}
\alias{load_herm}
\title{Download and load herm.rdata file containing SNP data from Sandy Inland Mouse
Details can be found in the article at https://rdinnager.github.io/slimr/articles/Main_manuscript_example_v2.html}
\usage{
load_herm(
  file_name = file.path(system.file(package = "slimr", "extdata"), "herm.rdata"),
  download = TRUE
)
}
\arguments{
\item{file_name}{File name to download data file to. It is recommended to use the default.}

\item{download}{Should the file be downloaded? Default is `TRUE`. If `FALSE`, only the URL to
the file will be returned.}
}
\value{
A `genlight` object with SNP data
}
\description{
Download and load herm.rdata file containing SNP data from Sandy Inland Mouse
Details can be found in the article at https://rdinnager.github.io/slimr/articles/Main_manuscript_example_v2.html
}
\examples{
herm <- load_herm(download = FALSE)
herm
}
