% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifAnalysis.R
\name{locateAllRegex}
\alias{locateAllRegex}
\title{locateAllRegex}
\usage{
locateAllRegex(sequence, pattern)
}
\arguments{
\item{sequence}{A character vector}

\item{pattern}{PERL like regular expression}
}
\description{
Finds the positions and substrings of all overlapping
matches of regexes in a given sequence
}
