% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database_Markers_list_PCSCT.R
\docType{data}
\name{Markers_list_PCSCT}
\alias{Markers_list_PCSCT}
\title{List of cell type markers in the article "Pan-cancer single cell landscapeof tumor-infiltrating T cells"}
\format{
A list with ten tables.
}
\source{
\doi{10.1126/science.abe6474}
}
\usage{
Markers_list_PCSCT
}
\description{
A dataset containing marker genes for different T cell types from the article "Pan-cancer single cell landscapeof tumor-infiltrating T cells"
}
\details{
This list is a table of 40 types of T cell markers obtained from the article "Pan-cancer single cell landscapeof tumor-infiltrating T cells".
The data source is "https://doi.org/10.1126/science.abe6474",
and the reference literature is: L. Zheng et al. (2021) \url{doi:10.1126/science.abe6474}.
}
\seealso{
Other Section_0_Database: 
\code{\link{Cellmarker2}},
\code{\link{Cellmarker2_raw}},
\code{\link{Cellmarker2_table}},
\code{\link{Markers_list_TCellSI}},
\code{\link{Markers_list_scIBD}},
\code{\link{PanglaoDB}},
\code{\link{PanglaoDB_raw}},
\code{\link{PanglaoDB_table}}
}
\concept{Section_0_Database}
\keyword{datasets}
