% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database_Markers_list_PCTAM.R
\docType{data}
\name{Markers_list_PCTAM}
\alias{Markers_list_PCTAM}
\title{List of Macrophage subtype markers in the article "Macrophage diversity in cancer revisited in the era of single-cell omics"}
\format{
A list with 7 tables.
}
\source{
\doi{10.1016/j.it.2022.04.008}
}
\usage{
Markers_list_PCTAM
}
\description{
A dataset containing marker genes for different Macrophage subtypes from the article "Macrophage diversity in cancer revisited in the era of single-cell omics"
}
\details{
This list is a table of 7 types of Tumor-associated macrophages (TAMs) markers obtained from the article "Macrophage diversity in cancer revisited in the era of single-cell omics".
The data source is "https://doi.org/10.1016/j.it.2022.04.008",
and the reference literature is: Ruo-Yu Ma et al. (2022) \url{doi:10.1016/j.it.2022.04.008}.
}
\seealso{
Other Section_0_Database:
\code{\link{Cellmarker2}},
\code{\link{Cellmarker2_raw}},
\code{\link{Cellmarker2_table}},
\code{\link{Markers_list_PCTIT}},
\code{\link{Markers_list_TCellSI}},
\code{\link{Markers_list_scIBD}},
\code{\link{PanglaoDB}},
\code{\link{PanglaoDB_raw}},
\code{\link{PanglaoDB_table}},
\code{\link{ScType}},
\code{\link{ScType_raw}},
\code{\link{ScType_table}}
}
\concept{Section_0_Database}
\keyword{datasets}
