% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database_Markers_list_scIBD.R
\docType{data}
\name{Markers_list_scIBD}
\alias{Markers_list_scIBD}
\title{List of cell type markers in the article scIBD}
\format{
A list with one hundred and one tables.
}
\source{
\doi{10.1038/s43588-023-00464-9}
}
\usage{
Markers_list_scIBD
}
\description{
A dataset containing marker genes for different human intestine cell types from scIBD
}
\details{
This list is a table of 101 types of human intestine cell types markers
obtained from scIBD. The article doi source is "https://doi.org/10.1038/s43588-023-00464-9",
and the reference literature is: Nie et al. (2023) \url{doi:10.1038/s43588-023-00464-9}.
Note: The 'Markers_list_scIBD' was generated using section 2.4 and the parameters
'sort_by = "logFC"' and 'gene_filter = 20' were set.
}
\seealso{
Other Section_0_Database:
\code{\link{Cellmarker2}},
\code{\link{Cellmarker2_raw}},
\code{\link{Cellmarker2_table}},
\code{\link{Markers_list_PCTAM}},
\code{\link{Markers_list_PCTIT}},
\code{\link{Markers_list_TCellSI}},
\code{\link{PanglaoDB}},
\code{\link{PanglaoDB_raw}},
\code{\link{PanglaoDB_table}},
\code{\link{ScType}},
\code{\link{ScType_raw}},
\code{\link{ScType_table}}
}
\concept{Section_0_Database}
\keyword{datasets}
