% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_callbacks.R
\name{mateChoice}
\alias{mateChoice}
\title{SLiM mateChoice() callback}
\usage{
mateChoice(subpop_id)
}
\arguments{
\item{subpop_id}{The id(s) of the subpopulation(s) to which this callback should apply. Can
be an integer 1, 2, etc., or character "p1", "p2", etc.}
}
\value{
None
}
\description{
This callback specifies that a code block is providing logic for an individual to choose a mate. The first parent is
chosen based on fitness values, then this callback is called to determine the second parent (if no callback is specified, it is chosen
randomly proportional to fitness). The callback code should return a vector of weights specifying the proportional probability of each individual in the population
being the second parent (including the first parent, as selfing is possible). Alternatively, it could return an \code{Individual} object
referring to the individual chosen as the second parent. Returning a numeric \code{0} tells SLiM that no mate could be found. Lastly, returning
\code{NULL} tells SLiM to use the default \code{weights} vector, which is the same as just returning \code{weights} but is more efficient.
see \href{http://benhaller.com/slim/SLiM_Manual.pdf#page=599}{SLiM Manual: page 599}
}
\details{
Global variables available in reproduction callbacks:
\describe{
  \item{individual}{The parent already chosen (the female, in sexual simulations)}
  \item{genome1}{One genome of the parent already chosen}
  \item{genome2}{The other genome of the parent already chosen}
  \item{subpop}{The subpopulation into which the offspring will be placed}
  \item{sourceSubpop}{The subpopulation from which the parents are being chosen}
  \item{weights}{The standard fitness-based weights for all individuals}
}
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
modified from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016-2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\examples{
# From page 600 of the SLiM Manual
slim_block(mateChoice(p2), {

  ## in subpopulation 2, high fitness individuals are preferred
  return(weights^2)

})
}
\seealso{
Other callbacks: 
\code{\link{early}()},
\code{\link{first}()},
\code{\link{fitness}()},
\code{\link{initialize}()},
\code{\link{interaction}()},
\code{\link{late}()},
\code{\link{modifyChild}()},
\code{\link{recombination}()},
\code{\link{reproduction}()},
\code{\link{slim_callbacks}()}
}
\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
\concept{callbacks}
