% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{minimal_slim_sim}
\alias{minimal_slim_sim}
\title{Generate a minimal SLiM simulation script}
\usage{
minimal_slim_sim(command = print("Hello world!"))
}
\arguments{
\item{command}{An R expression to be inserted in the `slimr` script}
}
\value{
An R `expression`.
}
\description{
This function is mainly used to quickly create a script which does
not does not do much, but can be used to test other code.
}
\examples{
minimal_slim_sim()
}
