% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{minimal_slimr_script}
\alias{minimal_slimr_script}
\title{Generate a minimal slimr_script object}
\usage{
minimal_slimr_script(command = print("Hello world!"))
}
\arguments{
\item{command}{An R expression to be inserted in the `slimr` script}
}
\value{
A `slimr_script` object
}
\description{
This function is mainly used to quickly create a script which does
not does not do much, but can be used to test other code that work
on `slimr_script` objects. Mainly used for documentation examples.
}
\examples{
minimal_slimr_script()
}
