% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{motifRegex}
\alias{motifRegex}
\title{SLiM regular expressions from ELM database}
\format{An object of class \code{list} of length 263.}
\source{
ELM classes \url{http://elm.eu.org/elms}
}
\usage{
motifRegex
}
\value{
A list object
}
\description{
List of available SLiM classes and their regular expression defined in the
ELM database as of 5th of August, 2016
}
\details{
This list can be updated using the getElmClasses() function motifRegex <-
as.list(as.vector(elms$RegEx)) names(motifRegex) <-
as.vector(elms$ELM_Identifier)
}
\keyword{datasets}
