% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_output.R
\name{mut_genotypes}
\alias{mut_genotypes}
\title{Load mutation genotypes and other information.}
\usage{
mut_genotypes(vcf, mut_type = NULL, pop_origin = NULL, t_min = -Inf,
  t_max = Inf)
}
\arguments{
\item{mut_type}{Mutation type using SLiM conventions ("m0", "m1", etc.).}

\item{pop_origin}{Population in which a mutation originated ("p0", "p1", etc.).}

\item{t_min}{Lower bound on time of origin.}

\item{t_max}{Upper bound on time of origin.}
}
\value{
GRanges object.
}
\description{
Load mutation genotypes and other information.
}
