% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifAnalysis.R
\name{mutateSequence}
\alias{mutateSequence}
\title{mutateSequence}
\usage{
mutateSequence(sequence, pos, wtAA, mutAA)
}
\arguments{
\item{sequence}{Character string of amino acids}

\item{pos}{The amino acid number (positive integer) at which a mutation
should be created}

\item{wtAA}{One letter code of the Wild-type amino-acid (to double check if
the desired mutation matches the actual coordinates of the wild-type
sequence)}

\item{mutAA}{One letter code of the mutant amino acid}
}
\value{
A character string where wtAA is replaced with mutAA if the
  amino-acid at the given position of the given sequence actually has the
  same amino acid.
}
\description{
Given a protein sequence, create a mutated copy of the sequence at the
desired position and amino acids.
}
\examples{

s <- 'AAAAA'
mutateSequence(s, 3, 'A', 'R')

}
