% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{mutationCountsInGenomes}
\alias{mutationCountsInGenomes}
\alias{Genome$mutationCountsInGenomes}
\alias{.G$mutationCountsInGenomes}
\title{SLiM method mutationCountsInGenomes}
\usage{
mutationCountsInGenomes(mutations)
}
\arguments{
\item{mutations}{An object of type null or Mutation object. The default value is
\code{NULL}. See details for description.}
}
\value{
An object of type integer.
}
\description{
Documentation for SLiM function \code{mutationCountsInGenomes}, which is a
method of the SLiM class \code{\link{Genome}}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Documentation for this function can be found in the official \href{http://
benhaller.com/slim/SLiM_Manual.pdf#page=536}{SLiM manual: page 536}.

Return an integer vector with the frequency counts of all of the
Mutation objects passed in mutations, within the target Genome vector. If the
optional mutations argument is NULL (the default), frequency counts will be
returned for all of the active Mutation objects in the simulation – the same
Mutation objects, and in the same order, as would be returned by the mutations
property of sim, in other words. See the +mutationFrequenciesInGenomes() method
to obtain float frequencies instead of integer counts. See also the SLiMSim
methods mutationCounts() and mutationFrequencies(), which may be more efficient
for getting counts/frequencies for whole subpopulations or for the whole
simulation.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016–2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\seealso{
Other Genome: 
\code{\link{Genome}},
\code{\link{addMutations}()},
\code{\link{addNewDrawnMutation}()},
\code{\link{addNewMutation}()},
\code{\link{containsMarkerMutation}()},
\code{\link{containsMutations}()},
\code{\link{countOfMutationsOfType}()},
\code{\link{mutationFrequenciesInGenomes}()},
\code{\link{mutationsOfType}()},
\code{\link{nucleotides}()},
\code{\link{outputMS}()},
\code{\link{outputVCF}()},
\code{\link{output}()},
\code{\link{positionsOfMutationsOfType}()},
\code{\link{readFromMS}()},
\code{\link{readFromVCF}()},
\code{\link{removeMutations}()},
\code{\link{sumOfMutationsOfType}()}
}
\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
\concept{Genome}
