% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_callbacks.R
\name{mutationEffect}
\alias{mutationEffect}
\title{SLiM mutationEffect() callback}
\usage{
mutationEffect(mut_type_id, subpop_id)
}
\arguments{
\item{mut_type_id}{The id of the mutationType to which this callback should apply. Can be an
integer 1, 2, etc., or character "m1", "m2", etc.}

\item{subpop_id}{The id(s) of the subpopulation(s) to which this callback should apply. Can
be an integer 1, 2, etc., or character "p1", "p2", etc.}
}
\value{
None
}
\description{
A mutationEffect() callback is called by SLiM when it is determining the fitness effect of a
mutation carried by an individual. Normally, the fitness effect of a mutation is determined by the
selection coefficient s of the mutation and the dominance coefficient h of the mutation (the latter
used only if the individual is heterozygous for the mutation). For details on this callback
see \href{http://benhaller.com/slim/SLiM_Manual.pdf#page=712}{SLiM Manual: page 712}
}
\details{
Global variables available in mutationEffect callbacks:
\describe{
  \item{mut}{A Mutation object, the mutation whose relative fitness is being evaluated}
  \item{homozygous}{A value of T (the mutation is homozygous), F (heterozygous), or NULL (it is paired with a null chromosome, and is thus hemizygous or haploid)}
  \item{effect}{The default relative fitness value calculated by SLiM}
  \item{individual}{The individual carrying this mutation (an object of class Individual)}
  \item{subpop}{The subpopulation in which that individual lives}
}
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
modified from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016-2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\examples{
# From page 714 of the SLiM Manual
slim_block(mutationEffect(), {
  if (homozygous) {
    return(1.0 + mut.selectionCoeff);
  } else {
    return(1.0 + mut.mutationType.dominanceCoeff * mut.selectionCoeff);
  }
})
}
\seealso{
Other callbacks: 
\code{\link{early}()},
\code{\link{first}()},
\code{\link{fitnessEffect}()},
\code{\link{fitness}()},
\code{\link{initialize}()},
\code{\link{interaction}()},
\code{\link{late}()},
\code{\link{mateChoice}()},
\code{\link{modifyChild}()},
\code{\link{mutation}()},
\code{\link{recombination}()},
\code{\link{reproduction}()},
\code{\link{slim_callbacks}()},
\code{\link{survival}()}
}
\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
\concept{callbacks}
