% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_output.R
\name{nea_per_ind}
\alias{nea_per_ind}
\title{Calculate the number of Nea. mutations (given in a GRanges object) in all
simulated individuals. Return result as an integer vector.}
\usage{
nea_per_ind(gr)
}
\description{
Nea. ancestry proportion can be calculated as: result / (2 * mut_count)
}
