% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{nearestNeighborsOfPoint}
\alias{nearestNeighborsOfPoint}
\alias{InteractionType$nearestNeighborsOfPoint}
\alias{.IT$nearestNeighborsOfPoint}
\title{SLiM method nearestNeighborsOfPoint}
\usage{
nearestNeighborsOfPoint(point, exerterSubpop, count)
}
\arguments{
\item{point}{An object of type float. See details for description.}

\item{exerterSubpop}{An object of type integer or Subpopulation object. Must be
of length 1 (a singleton). See details for description.}

\item{count}{An object of type integer. Must be of length 1 (a singleton). The
default value is \code{1}. See details for description.}
}
\value{
An object of type Individual object.
}
\description{
Documentation for SLiM function \code{nearestNeighborsOfPoint}, which is a
method of the SLiM class \code{\link{InteractionType}}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Documentation for this function can be found in the official
\href{http://benhaller.com/slim/SLiM_Manual.pdf#page=696}{SLiM manual: page
696}.

Returns up to count individuals in exerterSubpop that are spatially
closest to point, according to the distance metric of the InteractionType. The
subpopulation may be supplied either as an integer ID, or as a Subpopulation
object. To obtain all of the individuals within the maximum interaction distance
of point, simply pass a value for count that is greater than or equal to the
size of exerterSubpop. Note that if fewer than count individuals are within the
maximum interaction distance, the vector returned may be shorter than count,
or even zero-length; it is important to check for this possibility even when
requesting a single neighbor. The evaluate() method must have been previously
called for exerterSubpop, and positions saved at evaluation time will be used.
If the InteractionType is non-spatial, this method may not be called.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016-2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\seealso{
Other InteractionType: 
\code{\link{IT}},
\code{\link{clippedIntegral}()},
\code{\link{distance}()},
\code{\link{distanceFromPoint}()},
\code{\link{drawByStrength}()},
\code{\link{evaluate}()},
\code{\link{interactingNeighborCount}()},
\code{\link{interactionDistance}()},
\code{\link{localPopulationDensity}()},
\code{\link{nearestInteractingNeighbors}()},
\code{\link{nearestNeighbors}()},
\code{\link{neighborCount}()},
\code{\link{neighborCountOfPoint}()},
\code{\link{setConstraints}()},
\code{\link{setInteractionFunction}()},
\code{\link{strength}()},
\code{\link{testConstraints}()},
\code{\link{totalOfNeighborStrengths}()},
\code{\link{unevaluate}()}
}
\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
\concept{InteractionType}
