% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_script.R
\name{new_slimr_script_coll}
\alias{new_slimr_script_coll}
\title{Make a new \code{slimr_script_coll} object}
\usage{
new_slimr_script_coll(x = list())
}
\arguments{
\item{x}{A list of \code{slir_script} objects}
}
\value{
A \code{slimr_script_coll} object
}
\description{
Make a new \code{slimr_script_coll} object
}
