% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{nucleotideFrequencies}
\alias{nucleotideFrequencies}
\alias{SLiMBuiltin$nucleotideFrequencies}
\alias{.SB$nucleotideFrequencies}
\title{SLiM method nucleotideFrequencies}
\usage{
nucleotideFrequencies(sequence)
}
\arguments{
\item{sequence}{An object of type integer or string. See details for
description.}
}
\value{
An object of type string or float or float or float or integer or float
or integer.
}
\description{
Documentation for SLiM function \code{nucleotideFrequencies}, which is a method
of the SLiM class \code{SLiMBuiltin}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
A convenience function that returns a float vector of length four,
providing the frequencies of occurrences of A / C / G / T nucleotides,
respectively, in the supplied nucleotide sequence. The parameter sequence may be
a singleton string (e.g., "TATA"), a string vector of single characters (e.g.,
"T", "A", "T", "A"), or an integer vector (e.g., 3, 0, 3, 0), using SLiM’s
standard code of A=0, C=1, G=2, T=3.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016–2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
