% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{openDocument}
\alias{openDocument}
\alias{SLiMgui$openDocument}
\alias{.SG$openDocument}
\title{SLiM method openDocument}
\usage{
openDocument(filePath)
}
\arguments{
\item{filePath}{An object of type string. Must be of length 1 (a singleton). See
details for description.}
}
\value{
An object of type void or void.
}
\description{
Documentation for SLiM function \code{openDocument}, which is a method of the
SLiM class \code{\link{SLiMgui}}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Documentation for this function can be found in the official
\href{http://benhaller.com/slim/SLiM_Manual.pdf#page=711}{SLiM manual: page
711}.

Open the document at filePath in SLiMgui, if possible. Supported
document types include SLiM model files (typically with a .slim path
extension), text files (typically with a .txt path extension, and opened as
untitled model files), and PNG, JPG/JPEG, BMP, and GIF image file formats
(typically .png / .jpg / .jpeg / .bmp / .gif, respectively). (Note that in
SLiMguiLegacy, PDF files (.pdf) are supported but these other image file formats
are not.) This method can be particularly useful for opening images created by
the simulation itself, often by sublaunching a plotting process in R or another
environment; see section 14.8 for an example.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016-2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\seealso{
Other SLiMgui: 
\code{\link{SG}},
\code{\link{pauseExecution}()}
}
\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
\concept{SLiMgui}
