% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{parseMutation}
\alias{parseMutation}
\title{parseMutation}
\usage{
parseMutation(mutations)
}
\arguments{
\item{mutations}{Vector of mutations}
}
\value{
A data.frame object where columns are wtAA, mutAA, and pos
}
\description{
Given a vector of mutation substitutions
(e.g. "p.His160Arg") -> split "p.His160Arg" into "H 160 R"
}
