% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mean_expression_plot.R
\name{plot_mean_expression}
\alias{plot_mean_expression}
\title{Title}
\usage{
plot_mean_expression(object, features, assay = NULL, cluster_col = NULL)
}
\arguments{
\item{cluster_col}{}
}
\description{
Title
}
