% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{pointUniform}
\alias{pointUniform}
\alias{Subpopulation$pointUniform}
\alias{.P$pointUniform}
\title{SLiM method pointUniform}
\usage{
pointUniform(n)
}
\arguments{
\item{n}{An object of type integer. Must be of length 1 (a singleton). The
default value is \code{1}. See details for description.}
}
\value{
An object of type float.
}
\description{
Documentation for SLiM function \code{pointUniform}, which is a method of the
SLiM class \code{Subpopulation}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Returns a new point (or points, for n > 1) generated from uniform draws
for each coordinate, within the spatial boundaries of the subpopulation. The
returned vector will contain n points, each comprised of a number of coordinates
equal to the dimensionality of the simulation, so it will be of total length
n*dimensionality. This may only be called in simulations for which continuous
space has been enabled with initializeSLiMOptions().
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016–2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
