% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{positionsOfMutationsOfType}
\alias{positionsOfMutationsOfType}
\alias{Genome$positionsOfMutationsOfType}
\alias{.G$positionsOfMutationsOfType}
\title{SLiM method positionsOfMutationsOfType}
\usage{
positionsOfMutationsOfType(mutType)
}
\arguments{
\item{mutType}{An object of type integer or MutationType object. Must be of
length 1 (a singleton). See details for description.}
}
\value{
An object of type integer.
}
\description{
Documentation for SLiM function \code{positionsOfMutationsOfType}, which is a
method of the SLiM class \code{\link{Genome}}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Documentation for this function can be found in the official \href{http://
benhaller.com/slim/SLiM_Manual.pdf#page=538}{SLiM manual: page 538}.

Returns the positions of mutations that are of the type specified by
mutType, out of all of the mutations in the genome. If you need a vector of the
matching Mutation objects, rather than just positions, use -mutationsOfType().
This method is provided for speed; it is much faster than the corresponding
Eidos code.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016–2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\seealso{
Other Genome: 
\code{\link{Genome}},
\code{\link{addMutations}()},
\code{\link{addNewDrawnMutation}()},
\code{\link{addNewMutation}()},
\code{\link{containsMarkerMutation}()},
\code{\link{containsMutations}()},
\code{\link{countOfMutationsOfType}()},
\code{\link{mutationCountsInGenomes}()},
\code{\link{mutationFrequenciesInGenomes}()},
\code{\link{mutationsOfType}()},
\code{\link{nucleotides}()},
\code{\link{outputMS}()},
\code{\link{outputVCF}()},
\code{\link{output}()},
\code{\link{readFromMS}()},
\code{\link{readFromVCF}()},
\code{\link{removeMutations}()},
\code{\link{sumOfMutationsOfType}()}
}
\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
\concept{Genome}
