% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariationData.R
\name{processVEP}
\alias{processVEP}
\title{processVEP}
\usage{
processVEP(vcfFilePath, nodeN = 8)
}
\arguments{
\item{vcfFilePath}{path to VCF file containing variation data}
}
\value{
A data.table object
}
\description{
This function processes the output of Variant Effect Predictor to
select missense variants and create some columns that are useful
to assess the pathogenicity of variants
}

