% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_output.R
\name{r_output_sex}
\alias{r_output_sex}
\alias{slimr_output_sex}
\title{Utility function to tell SLim to output sexes of individuals}
\usage{
r_output_sex(name = "sex", ...)

slimr_output_sex(name = "sex", ...)
}
\arguments{
\item{name}{Name of output to use to label it in \code{slimr_results object}. Default is \code{"sex"}.}

\item{...}{Other arguments to be passed to \code{\link{r_output}}}
}
\value{
None
}
\description{
Utility function to tell SLim to output sexes of individuals
}
\examples{
if(slim_is_avail()) {
  test_sim <- slim_script(
    slim_block_init_minimal(mutation_rate = 1e-6),
    slim_block_add_subpops(1, 100),
    slim_block(1, 20, late(), {
      r_output_sex("sex", do_every = 10)
    })
  ) \%>\%
    slim_run()
  slim_results_to_data(test_sim$output_data)
}
}
