% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{read_coordinates}
\alias{read_coordinates}
\title{Read positions of simulated coordinates (regions or sites) in their
original coordinate system.}
\usage{
read_coordinates(file)
}
\arguments{
\item{file}{Path to a BED-like file. See the required format bellow.}
}
\value{
GRanges object.
}
\description{
SLiM simulates everything as part of a single continuous "chromosome".
Therefore, matching simulated coordinates to their real coordinates
requires their transposition into a 0-based continuous coordinate system.
This function reads a table in a BED-like format that makes it possible
to "reverse transpose" the simulated 0-based coordinates into their
original form using the "transpose_sites" function.
}
\details{
The coordinate file must be in a TSV format and must contain columns
chrom, start, end, slim_start and slim_end.
}
\seealso{
admixr::transpose_coordinates
}
