% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_output.R
\name{read_genomes}
\alias{read_genomes}
\title{Parse the genome section of the SLiM file into a data frame.}
\usage{
read_genomes(slim_file, pop_origin = "")
}
\arguments{
\item{slim_file}{SLiM text output file.}

\item{pop_origin}{Population of origin.}
}
\description{
Parse the genome section of the SLiM file into a data frame.
}
