% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_output.R
\name{read_mutations}
\alias{read_mutations}
\title{Parse the mutation section of the SLiM file into a data frame.}
\usage{
read_mutations(slim_file, mut_type = "", pop_origin = "", gen_origin = 0)
}
\arguments{
\item{slim_file}{SLiM text output file.}

\item{mut_type}{Mutation type id (m0, m1, etc.).}

\item{pop_origin}{Population of origin.}

\item{gen_origin}{Generation of origin.}
}
\description{
Parse the mutation section of the SLiM file into a data frame.
}
