% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_output.R
\name{read_populations}
\alias{read_populations}
\title{Parse the population section of the SLiM file into a data frame.}
\usage{
read_populations(slim_file)
}
\arguments{
\item{slim_file}{Lines of the whole SLiM output file.}
}
\description{
Parse the population section of the SLiM file into a data frame.
}
