% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_output.R
\name{read_regions}
\alias{read_regions}
\title{Read BED file with coordinates of a given type of genomic elements.}
\usage{
read_regions(regions_bed, slim = FALSE)
}
\description{
Read BED file with coordinates of a given type of genomic elements.
}
