% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_output.R
\name{read_section_data}
\alias{read_section_data}
\title{Get the subset of SLiM output file contents for a given section.}
\usage{
read_section_data(slim_file, section_name)
}
\arguments{
\item{slim_file}{SLiM text output file.}

\item{section_name}{Section of the output file.}
}
\description{
Get the subset of SLiM output file contents for a given section.
}
