% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_output.R
\name{read_section_delims}
\alias{read_section_delims}
\title{Read the positions of the SLiM output file section headers
used to break up the whole file into individual chunks
(population description, mutations, individuals and genomes).}
\usage{
read_section_delims(slim_file)
}
\arguments{
\item{slim_file}{SLiM text output file.}
}
\description{
Read the positions of the SLiM output file section headers
used to break up the whole file into individual chunks
(population description, mutations, individuals and genomes).
}
