% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{read_sites}
\alias{read_sites}
\title{Read positions of simulated sites in their original coordinate system.}
\usage{
read_sites(file)
}
\arguments{
\item{file}{Path to a BED-like file. See the required format bellow.}
}
\value{
GRanges object.
}
\description{
SLiM simulates all mutations as part of a single continuous "chromosome".
Therefore, simulating fixed mutations that are present in a real genome
requires their transposition into a 0-based continuous coordinate system.
This function reads a table in a BED-like format that makes it possible
to "reverse transpose" the simulated 0-based coordinates into their
original form using the "transpose_sites" function.
}
\details{
The file has a strict (tab-separated) format and must contain the following
columns in the same order: original_chromosome_id original_start_pos
original_end_pos SLiM_start_pos SLiM_end_pos.
}
\seealso{
admixr::transpose_sites
}
