% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_output.R
\name{read_slim}
\alias{read_slim}
\title{Read SLiM output file generated by the sim.outputFull() method.}
\usage{
read_slim(path)
}
\arguments{
\item{path}{Path to a SLiM output file.}
}
\description{
Read SLiM output file generated by the sim.outputFull() method.
}
