% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_output.R
\name{read_slim_file}
\alias{read_slim_file}
\title{Read SLiM output file generated by the sim.outputFull() method.}
\usage{
read_slim_file(slim_file)
}
\arguments{
\item{slim_file}{SLiM text output file.}
}
\description{
Read SLiM output file generated by the sim.outputFull() method.
}
