% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_output.R
\name{read_vcf}
\alias{read_vcf}
\title{Read a simulated VCF file.}
\usage{
read_vcf(path)
}
\arguments{
\item{path}{Path to a VCF file.}
}
\value{
VCF object.
}
\description{
Read a simulated VCF file.
}
\seealso{
VariantAnnotation::readVCF
}
