% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{readFromMS}
\alias{readFromMS}
\alias{Genome$readFromMS}
\alias{.G$readFromMS}
\title{SLiM method readFromMS}
\usage{
readFromMS(filePath, mutationType)
}
\arguments{
\item{filePath}{An object of type string. Must be of length 1 (a singleton). See
details for description.}

\item{mutationType}{An object of type integer or MutationType object. Must be of
length 1 (a singleton). See details for description.}
}
\value{
An object of type Mutation object.
}
\description{
Documentation for SLiM function \code{readFromMS}, which is a method of the SLiM
class \code{Genome}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Read new mutations from the MS format file at filePath and add them
to the target genomes. The number of target genomes must match the number
of genomes represented in the MS file. To read into all of the genomes in
a given subpopulation pN, simply call pN.genomes.readFromMS(), assuming the
subpopulation’s size matches that of the MS file. A vector containing all of the
mutations created by readFromMS() is returned. Each mutation is created at the
position specified in the file, using the mutation type given by mutationType.
Positions are expected to be in [0,1], and are scaled to the length of the
chromosome by multiplying by the last valid base position of the chromosome
(i.e., one less than the chromosome length). Selection coefficients are drawn
from the mutation type. The population of origin for each mutation is set
to -1, and the generation of origin is set to the current generation. In a
nucleotidebased model, if mutationType is nucleotide-based, a random nucleotide
different from the ancestral nucleotide at the position will be chosen with
equal probability. The target genomes correspond, in order, to the call lines
in the MS file. In sex-based models that simulate the X or Y chromosome, null
genomes in the target vector will be skipped, and will not be used to correspond
to any call line; however, care should be taken in this case that the lines in
the MS file correspond to the target genomes in the manner desired.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016–2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
