% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{readIUPred}
\alias{readIUPred}
\title{readIUpred}
\usage{
readIUPred(iupredResultFiles)
}
\arguments{
\item{A}{vector of file paths containing the results of IUPred}
}
\value{
A list of data.frame objects
}
\description{
Given a list of files each containing iupred results for a single protein,
return a list of data frames where each data frame holds the iupred results.
}
