% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{registerEarlyEvent}
\alias{registerEarlyEvent}
\alias{SLiMSim$registerEarlyEvent}
\alias{.SS$registerEarlyEvent}
\title{SLiM method registerEarlyEvent}
\usage{
registerEarlyEvent(id, source, start, end)
}
\arguments{
\item{id}{An object of type null or integer or string. Must be of length 1 (a
singleton). See details for description.}

\item{source}{An object of type string. Must be of length 1 (a singleton). See
details for description.}

\item{start}{An object of type null or integer. Must be of length 1 (a
singleton). The default value is \code{NULL}. See details for description.}

\item{end}{An object of type null or integer. Must be of length 1 (a singleton).
The default value is \code{NULL}. See details for description.}
}
\value{
An object of type SLiMEidosBlock object. Return will be of length 1 (a
singleton)
}
\description{
Documentation for SLiM function \code{registerEarlyEvent}, which is a method of
the SLiM class \code{SLiMSim}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Register a block of Eidos source code, represented as the string
singleton source, as an Eidos early() event in the current simulation, with
optional start and end generations limiting its applicability. The script block
will be given identifier id (specified as an integer, or as a string symbolic
name such as "s5"); this may be NULL if there is no need to be able to refer
to the block later. The registered event is added to the end of the list of
registered SLiMEidosBlock objects, and is active immediately; it may be eligible
to execute in the current generation (see section 24.9 for details). The new
SLiMEidosBlock will be defined as a global variable immediately by this method
(see section 23.10), and will also be returned by this method.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016–2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
