% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_callbacks.R
\name{reproduction}
\alias{reproduction}
\title{SLiM reproduction() callback}
\usage{
reproduction(subpop_id, sex)
}
\arguments{
\item{subpop_id}{The id(s) of the subpopulation(s) to which this callback should apply. Can
be an integer 1, 2, etc., or character "p1", "p2", etc.}

\item{sex}{The sex of the individuals to apply this callback to.}
}
\value{
None
}
\description{
This callback specifies that a code block is providing logic for the production
of offspring. It is called once per individual for each generation in which it is active,
and is expected to add new individuals to the simulation using the SLiM functions provided
for that purpose. This is only used in "nonWF" simulations.
For more information on how to use \code{reproduction()} callback
see \href{http://benhaller.com/slim/SLiM_Manual.pdf#page=606}{SLiM Manual: page 606}
}
\details{
Global variables available in reproduction callbacks:
\describe{
  \item{individual}{The focal individual that is expected to reproduce}
  \item{genome1}{One genome of the focal individual}
  \item{genome2}{The other genome of the focal individual}
  \item{subpop}{The subpopulation to which the focal individual belongs}
}
}
\examples{
slim_block(reproduction()) {

  subpop.addCrossed(individual, subpop.sampleIndividuals(1));

}
}
