% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{rescale}
\alias{rescale}
\alias{SpatialMap$rescale}
\alias{.SM$rescale}
\title{SLiM method rescale}
\usage{
rescale(min, max)
}
\arguments{
\item{min}{An object of type numeric or numeric. Must be of length 1 (a
singleton). The default value is \code{0.0}. See details for description.}

\item{max}{An object of type numeric or numeric. Must be of length 1 (a
singleton). The default value is \code{1.0}. See details for description.}
}
\value{
An object of type SpatialMap object. Return will be of length 1 (a
singleton)
}
\description{
Documentation for SLiM function \code{rescale}, which is a method of the SLiM
class \code{\link{SpatialMap}}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Documentation for this function can be found in the official
\href{http://benhaller.com/slim/SLiM_Manual.pdf#page=716}{SLiM manual: page
716}.

Rescales the values of the spatial map to the range [min, max]. By
default, the rescaling is to the range [0.0, 1.0]. It is required that min be
less than max, and that both be finite. Note that the final range may not be
exactly [min, max] due to numerical error. The target spatial map is returned,
to allow easy chaining of operations.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016-2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\seealso{
Other SpatialMap: 
\code{\link{SM}},
\code{\link{add}()},
\code{\link{blend}()},
\code{\link{changeColors}()},
\code{\link{changeValues}()},
\code{\link{divide}()},
\code{\link{exp}()},
\code{\link{gridValues}()},
\code{\link{interpolate}()},
\code{\link{mapColor}()},
\code{\link{mapImage}()},
\code{\link{mapValue}()},
\code{\link{multiply}()},
\code{\link{power}()},
\code{\link{range}()},
\code{\link{sampleImprovedNearbyPoint}()},
\code{\link{sampleNearbyPoint}()},
\code{\link{smooth}()},
\code{\link{subtract}()}
}
\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
\concept{SpatialMap}
