% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{runIUPred}
\alias{runIUPred}
\title{runIUPred}
\usage{
runIUPred(iupredPath, fastaFile, nodeN = 1)
}
\arguments{
\item{iupredPath}{The path to the folder containing the source code of the
IUPred tool}

\item{fastaFile}{A fasta file containing the amino acid sequences}

\item{nodeN}{default: 1. Positive integer for the number of parallel cores to
use for downloading and processing the files}
}
\value{
A list of numerical arrays, where each value in the array correspond
  to the disorder score propensity of the corresponding residue in the
  protein sequence.
}
\description{
run IUPred tool to get disorder propensity scores of a given protein sequence
in fasta format.
}
\details{
Given a fasta file containing one or more protein sequences; calculate
per-base iupred disorder scores. IUPred source code can be dowloaded from
here: http://iupred.enzim.hu/Downloads.php After unpacking the source code,
cd to the src directory. Compile the code with "cc iupred.c -o iupred"
}
