% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariationData.R
\name{runVEP}
\alias{runVEP}
\title{runVEP}
\usage{
runVEP(vepPATH = "/home/buyar/.local/bin/variant_effect_predictor.pl",
  vcfFilePath, overwrite = FALSE)
}
\arguments{
\item{vepPATH}{path to the variant_effect_predictor.pl script}

\item{vcfFilePath}{path to VCF file containing variation data}
}
\value{
a data.table data.frame containing variation data read from VEP
  output
}
\description{
A wrapper function to run Ensembl's variant_effect_predictor script for a
given vcf file.
}

