% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleFasta}
\alias{sampleFasta}
\title{Sample sequence extracted from a fasta file}
\format{data.frame object with 22 rows and two columns}
\source{
Downloaded from uniprot \url{www.uniprot.org/uniprot/P11166.fasta}
}
\usage{
sampleFasta
}
\value{
A data.frame object
}
\description{
The amino acid sequence for P11166 protein (GLUT1)
}
\keyword{datasets}

