% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{sampleImprovedNearbyPoint}
\alias{sampleImprovedNearbyPoint}
\alias{SpatialMap$sampleImprovedNearbyPoint}
\alias{.SM$sampleImprovedNearbyPoint}
\title{SLiM method sampleImprovedNearbyPoint}
\usage{
sampleImprovedNearbyPoint(point, maxDistance, functionType, ...)
}
\arguments{
\item{point}{An object of type float. See details for description.}

\item{maxDistance}{An object of type float. Must be of length 1 (a singleton).
See details for description.}

\item{functionType}{An object of type string. Must be of length 1 (a singleton).
See details for description.}

\item{...}{An object of type NA. NA See details for description.}
}
\value{
An object of type float.
}
\description{
Documentation for SLiM function \code{sampleImprovedNearbyPoint}, which is a
method of the SLiM class \code{\link{SpatialMap}}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Documentation for this function can be found in the official
\href{http://benhaller.com/slim/SLiM_Manual.pdf#page=716}{SLiM manual: page
716}.

This variant of sampleNearbyPoint() samples a Metropolis-Hastings
move on the spatial map. See sampleNearbyPoint() for discussion of the basic
idea. This method proposes a nearby point drawn from the given kernel. If the
drawn point has a larger map value than the original point, the new point is
returned. If the drawn point has a smaller map value than the original point, it
is returned with a probability equal to the ratio between its map value and the
original map value, otherwise the original point is returned. The distribution
of individuals that move (or not) to new locations governed by this method will
converge upon the map itself, in a similar manner to how MCMC converges upon
the posterior distribution (assuming no other forces, such as birth or death,
influence the distribution of individuals). Movement governed by this method
is "improved" in the sense that individuals will tend to remain where they are
unless the new sampled point is an improvement for them - a higher map value.
Note that unlike sampleNearbyPoint(), this method requires that all map values
are non-negative.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016-2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\seealso{
Other SpatialMap: 
\code{\link{SM}},
\code{\link{add}()},
\code{\link{blend}()},
\code{\link{changeColors}()},
\code{\link{changeValues}()},
\code{\link{divide}()},
\code{\link{exp}()},
\code{\link{gridValues}()},
\code{\link{interpolate}()},
\code{\link{mapColor}()},
\code{\link{mapImage}()},
\code{\link{mapValue}()},
\code{\link{multiply}()},
\code{\link{power}()},
\code{\link{range}()},
\code{\link{rescale}()},
\code{\link{sampleNearbyPoint}()},
\code{\link{smooth}()},
\code{\link{subtract}()}
}
\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
\concept{SpatialMap}
