% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database_ScType.R
\docType{data}
\name{ScType}
\alias{ScType}
\title{ScType dataset}
\format{
A tibble with 3 columns:
\describe{
\item{tissue_type}{Tissue type (e.g., "Immune system", "Brain", "Liver")}
\item{cell_name}{Cell type name, formatted as "cellName(shortName)" when a
short name is available, or "cellName" otherwise}
\item{marker}{Gene symbol of the marker}
}
}
\source{
\url{https://github.com/IanevskiAleksandr/sc-type}
}
\usage{
ScType
}
\description{
A processed long-format dataset containing marker genes for different cell
types from the ScType database. Each row represents one marker gene for a
given tissue type and cell type.
}
\details{
This dataset is used to filter and create a standardized marker list.
The dataset can be filtered based on tissue type and cell name to generate
a list of marker genes for specific cell types using
\code{\link{Markers_filter_ScType}}.
}
\seealso{
Other Section_0_Database:
\code{\link{Cellmarker2}},
\code{\link{Cellmarker2_raw}},
\code{\link{Cellmarker2_table}},
\code{\link{Markers_list_PCTAM}},
\code{\link{Markers_list_PCTIT}},
\code{\link{Markers_list_TCellSI}},
\code{\link{Markers_list_scIBD}},
\code{\link{PanglaoDB}},
\code{\link{PanglaoDB_raw}},
\code{\link{PanglaoDB_table}},
\code{\link{ScType_raw}},
\code{\link{ScType_table}}
}
\concept{Section_0_Database}
\keyword{datasets}
