% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database_ScType.R
\docType{data}
\name{ScType_table}
\alias{ScType_table}
\title{ScType metadata table}
\format{
A list with 2 elements:
\describe{
\item{tissue_type}{Frequency table of tissue types}
\item{cell_name}{Frequency table of cell type names}
}
}
\source{
\url{https://github.com/IanevskiAleksandr/sc-type}
}
\usage{
ScType_table
}
\description{
A list of frequency tables summarizing the ScType database, useful for
exploring available tissue types and cell types before filtering.
}
\seealso{
Other Section_0_Database:
\code{\link{Cellmarker2}},
\code{\link{Cellmarker2_raw}},
\code{\link{Cellmarker2_table}},
\code{\link{Markers_list_PCTAM}},
\code{\link{Markers_list_PCTIT}},
\code{\link{Markers_list_TCellSI}},
\code{\link{Markers_list_scIBD}},
\code{\link{PanglaoDB}},
\code{\link{PanglaoDB_raw}},
\code{\link{PanglaoDB_table}},
\code{\link{ScType}},
\code{\link{ScType_raw}}
}
\concept{Section_0_Database}
\keyword{datasets}
