% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{setAncestralNucleotides}
\alias{setAncestralNucleotides}
\alias{Chromosome$setAncestralNucleotides}
\alias{.Ch$setAncestralNucleotides}
\title{SLiM method setAncestralNucleotides}
\usage{
setAncestralNucleotides(sequence)
}
\arguments{
\item{sequence}{An object of type integer or string. See details for
description.}
}
\value{
An object of type integer. Return will be of length 1 (a singleton)
}
\description{
Documentation for SLiM function \code{setAncestralNucleotides}, which is a
method of the SLiM class \code{\link{Chromosome}}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Documentation for this function can be found in the official
\href{http://benhaller.com/slim/SLiM_Manual.pdf#page=633}{SLiM manual: page
633}.

This method, which may be called only in nucleotide-based models
(see section 1.8), replaces the ancestral nucleotide sequence for the
model. The sequence parameter is interpreted exactly as it is in the
initializeAncestralSequence() function; see that documentation for details
(section 24.1). The length of the ancestral sequence is returned. It is unusual
to replace the ancestral sequence in a running simulation, since the nucleotide
states of segregating and fixed mutations will depend upon the original
ancestral sequence. It can be useful when loading a new population state with
readFromMS() or readFromVCF(), such as when resetting the simulation state to an
earlier state in a conditional simulation; however, that is more commonly done
using readFromPopulationFile() with a SLiM or .trees file.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016-2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\seealso{
Other Chromosome: 
\code{\link{Ch}},
\code{\link{ancestralNucleotides}()},
\code{\link{drawBreakpoints}()},
\code{\link{setGeneConversion}()},
\code{\link{setHotspotMap}()},
\code{\link{setMutationRate}()},
\code{\link{setRecombinationRate}()}
}
\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
\concept{Chromosome}
