% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{setCloningRate}
\alias{setCloningRate}
\alias{Subpopulation$setCloningRate}
\alias{.P$setCloningRate}
\title{SLiM method setCloningRate}
\usage{
setCloningRate(rate)
}
\arguments{
\item{rate}{An object of type numeric. See details for description.}
}
\value{
An object of type void.
}
\description{
Documentation for SLiM function \code{setCloningRate}, which is a method of the
SLiM class \code{Subpopulation}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Set the cloning rate of this subpopulation. The rate is changed to
rate, which should be between 0.0 and 1.0, inclusive. Clonal reproduction can
be enabled in both non-sexual (i.e. hermaphroditic) and sexual simulations. In
non-sexual simulations, rate must be a singleton value representing the overall
clonal reproduction rate for the subpopulation. In sexual simulations, rate
may be either a singleton (specifying the clonal reproduction rate for both
sexes) or a vector containing two numeric values (the female and male cloning
rates specified separately, at indices 0 and 1 respectively). During mating and
offspring generation, the probability that any given offspring individual will
be generated by cloning – by asexual reproduction without gametes or meiosis –
will be equal to the cloning rate (for its sex, in sexual simulations) set in
the parental (not the offspring!) subpopulation.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016–2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
