% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{setMigrationRates}
\alias{setMigrationRates}
\alias{Subpopulation$setMigrationRates}
\alias{.P$setMigrationRates}
\title{SLiM method setMigrationRates}
\usage{
setMigrationRates(sourceSubpops, rates)
}
\arguments{
\item{sourceSubpops}{An object of type integer or Subpopulation object. See
details for description.}

\item{rates}{An object of type numeric. See details for description.}
}
\value{
An object of type void.
}
\description{
Documentation for SLiM function \code{setMigrationRates}, which is a method of
the SLiM class \code{Subpopulation}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Set the migration rates to this subpopulation from the subpopulations
in sourceSubpops to the corresponding rates specified in rates; in other words,
rates gives the expected fractions of the children in this subpopulation that
will subsequently be generated from parents in the subpopulations sourceSubpops
(see section 21.2.1). This method will only set the migration fractions from
the subpopulations given; migration rates from other subpopulations will be
left unchanged (explicitly set a zero rate to turn off migration from a given
subpopulation). The type of sourceSubpops may be either integer, specifying
subpopulations by identifier, or object, specifying subpopulations directly.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016–2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
