% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{setMutationMatrix}
\alias{setMutationMatrix}
\alias{GenomicElementType$setMutationMatrix}
\alias{.GET$setMutationMatrix}
\title{SLiM method setMutationMatrix}
\usage{
setMutationMatrix(mutationMatrix)
}
\arguments{
\item{mutationMatrix}{An object of type float. See details for description.}
}
\value{
An object of type any but object or void or void or void.
}
\description{
Documentation for SLiM function \code{setMutationMatrix}, which is a method of
the SLiM class \code{GenomicElementType}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Sets a new nucleotide mutation matrix for the genomic
element type. This replaces the mutation matrix originally set by
initializeGenomicElementType(). This method may only be called in nucleotide-
based models.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016–2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
