% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sets_equal}
\alias{sets_equal}
\title{Compares two sets and returns TRUE if they are the same.}
\usage{
sets_equal(actual, expected)
}
\arguments{
\item{actual}{The actual set.}

\item{expected}{The expected set.}
}
\value{
A logical scalar, TRUE if the sets are equal, FALSE if they are not equal
}
\description{
This code modified from code (by fangly) found here: https://github.com/r-lib/testthat/issues/473. Should be equivalent to the test performed in testthat::expect_setequal
}
