% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{setSpatialBounds}
\alias{setSpatialBounds}
\alias{Subpopulation$setSpatialBounds}
\alias{.P$setSpatialBounds}
\title{SLiM method setSpatialBounds}
\usage{
setSpatialBounds(bounds)
}
\arguments{
\item{bounds}{An object of type numeric. See details for description.}
}
\value{
An object of type void.
}
\description{
Documentation for SLiM function \code{setSpatialBounds}, which is a method of
the SLiM class \code{\link{Subpopulation}}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Documentation for this function can be found in the official \href{http://
benhaller.com/slim/SLiM_Manual.pdf#page=587}{SLiM manual: page 587}.

Set the spatial boundaries of the subpopulation to bounds. This
method may be called only for simulations in which continuous space has been
enabled with initializeSLiMOptions(). The length of bounds must be double the
spatial dimensionality, so that it supplies both minimum and maximum values
for each coordinate. More specifically, for a dimensionality of "x", bounds
should supply (x0, x1) values; for dimensionality "xy" it should supply (x0,
y0, x1, y1) values; and for dimensionality "xyz" it should supply (x0, y0,
z0, x1, y1, z1) (in that order). These boundaries will be used by SLiMgui to
calibrate the display of the subpopulation, and will be used by methods such
as pointInBounds(), pointReflected(), pointStopped(), and pointUniform(). The
default spatial boundaries for all subpopulations span the interval [0,1] in
each dimension. Spatial dimensions that are periodic (as established with the
periodicity parameter to initializeSLiMOptions()) must have a minimum coordinate
value of 0.0 (a restriction that allows the handling of periodicity to be
somewhat more efficient). The current spatial bounds for the subpopulation may
be obtained through the spatialBounds property.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016-2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\seealso{
Other Subpopulation: 
\code{\link{Subpopulation}},
\code{\link{addCloned}()},
\code{\link{addCrossed}()},
\code{\link{addEmpty}()},
\code{\link{addRecombinant}()},
\code{\link{addSelfed}()},
\code{\link{cachedFitness}()},
\code{\link{configureDisplay}()},
\code{\link{defineSpatialMap}()},
\code{\link{outputMSSample}()},
\code{\link{outputSample}()},
\code{\link{outputVCFSample}()},
\code{\link{pointInBounds}()},
\code{\link{pointPeriodic}()},
\code{\link{pointReflected}()},
\code{\link{pointStopped}()},
\code{\link{pointUniform}()},
\code{\link{removeSubpopulation}()},
\code{\link{sampleIndividuals}()},
\code{\link{setCloningRate}()},
\code{\link{setMigrationRates}()},
\code{\link{setSelfingRate}()},
\code{\link{setSexRatio}()},
\code{\link{setSubpopulationSize}()},
\code{\link{spatialMapColor}()},
\code{\link{spatialMapValue}()},
\code{\link{subsetIndividuals}()},
\code{\link{takeMigrants}()}
}
\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
\concept{Subpopulation}
