% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_slim.R
\name{setup_slim}
\alias{setup_slim}
\title{Attempt to install and / or setup SLiM for use with slimr}
\usage{
setup_slim(platform = get_os(), install_dir = "default")
}
\arguments{
\item{platform}{Character with platform. One of "windows", "unix", or "osx". By default \code{setup_slim()}
will attempt to determine the platform automatically. Note that on Windows, you must have Windows
subsystem for linux installed, which requires at least Windows 10.}

\item{install_dir}{Directory to install SLiM to. If "default" \code{setup_slim()} will install in the default
directory. For simplicity we reccommend sticking with "default" unless you have a good reason not to.}
}
\description{
Attempt to install and / or setup SLiM for use with slimr
}
\examples{
\donttest{
setup_slim()
}
}
