% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{setValue}
\alias{setValue}
\alias{LogFile$setValue}
\alias{.LF$setValue}
\title{SLiM method setValue}
\usage{
setValue(key, value)
}
\arguments{
\item{key}{An object of type string or any. Must be of length 1 (a singleton).
See details for description.}

\item{value}{An object of type string or any. See details for description.}
}
\value{
An object of type void or void or void or void or void or void or void
or void or void or void or void or void or void or void or void or void or void
or void or logical.
}
\description{
Documentation for SLiM function \code{setValue}, which is a method of the SLiM
class \code{\link{LogFile}}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Documentation for this function can be found in the official
\href{http://benhaller.com/slim/SLiM_Manual.pdf#page=703}{SLiM manual: page
703}.

This Dictionary method has an override in LogFile to make it illegal to
call, since LogFile manages its Dictionary entries.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016-2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\seealso{
Other LogFile: 
\code{\link{LF}},
\code{\link{addCustomColumn}()},
\code{\link{addCycleStage}()},
\code{\link{addCycle}()},
\code{\link{addKeysAndValuesFrom}()},
\code{\link{addMeanSDColumns}()},
\code{\link{addPopulationSexRatio}()},
\code{\link{addPopulationSize}()},
\code{\link{addSubpopulationSexRatio}()},
\code{\link{addSubpopulationSize}()},
\code{\link{addSuppliedColumn}()},
\code{\link{addTick}()},
\code{\link{clearKeysAndValues}()},
\code{\link{flush}()},
\code{\link{logRow}()},
\code{\link{setFilePath}()},
\code{\link{setLogInterval}()},
\code{\link{setSuppliedValue}()},
\code{\link{willAutolog}()}
}
\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
\concept{LogFile}
