% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{sharedParentCount}
\alias{sharedParentCount}
\alias{Individual$sharedParentCount}
\alias{.I$sharedParentCount}
\title{SLiM method sharedParentCount}
\usage{
sharedParentCount(individuals)
}
\arguments{
\item{individuals}{An object of type Individual object. See details for
description.}
}
\value{
An object of type integer.
}
\description{
Documentation for SLiM function \code{sharedParentCount}, which is a method of
the SLiM class \code{\link{Individual}}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Documentation for this function can be found in the official
\href{http://benhaller.com/slim/SLiM_Manual.pdf#page=684}{SLiM manual: page
684}.

Returns a vector containing the number of parents shared between
the receiver and each of the individuals in individuals. The number of shared
parents between A and B is always 2 if A and B are actually the same individual;
this facility works even if SLiM's optional pedigree tracking is not enabled
(in which case all other relatedness values will be 0). Otherwise, if pedigree
tracking is turned on with initializeSLiMOptions(keepPedigrees=T), this method
will use the pedigree information described in section 25.7.1 to construct a
relatedness estimate. More specifically, this method uses the parental pedigree
IDs from the pedigree records of a pair of individuals to count the number
of shared parents between them, such that full siblings (with all of the same
parents) have a count of 2, and half siblings (with half of the same parents)
have a count of 1. If possible parents of the two individuals are A, B, C, and
D, then the shared parent count is as follows, for some illustrative examples.
The first column showing the two parents of the first individual, the second
column showing the two parents of the second individual; note that the two
parents of an individual can be the same due to cloning or selfing: AB CD → 0
(no shared parents) AB CC → 0 (no shared parents) AB AC → 1 (half siblings) AB
AA → 1 (half siblings) AA AB → 1 (half siblings) AB AB → 2 (full siblings) AB
BA → 2 (full siblings) AA AA → 2 (full siblings) This method does not estimate
consanguinity. For example, if one individual is itself a parent of the other
individual, that is irrelevant for this method. Similarly, in simulations of
sex chromosomes, the sexes of the parents are irrelevant, even if no genetic
material would have been inherited from a given parent. See relatedness() for an
assessment of pedigree-based relatedness that does estimate the consanguinity of
individuals. The sharedParentCount() method is preferable if your exact question
is simply whether individuals are full siblings, half siblings, or non-siblings;
in other cases, relatedness() is probably more useful.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016-2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\seealso{
Other Individual: 
\code{\link{In}},
\code{\link{containsMutations}()},
\code{\link{countOfMutationsOfType}()},
\code{\link{relatedness}()},
\code{\link{setSpatialPosition}()},
\code{\link{sumOfMutationsOfType}()},
\code{\link{uniqueMutationsOfType}()}
}
\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
\concept{Individual}
