% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canned_models.R
\name{simple_spatial_model}
\alias{simple_spatial_model}
\title{Title}
\usage{
simple_spatial_model(
  ngen = 10000,
  N = 500,
  nloc = 10000,
  mating_sd = 0.1,
  move_sd = 0.01,
  xlims = c(0, 1),
  ylims = c(0, 1),
  mutation_rate = 1e-06,
  recomb_rate = 1e-08,
  .seed = .Random.seed,
  output_prop = 1/100,
  output_file = "slim_output.txt"
)
}
\arguments{
\item{output_file}{}
}
\value{

}
\description{
Title
}
