% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{skipTick}
\alias{skipTick}
\alias{Species$skipTick}
\alias{.Sp$skipTick}
\title{SLiM method skipTick}
\usage{
skipTick(void)
}
\arguments{
\item{void}{An object of type . See details for description.}
}
\value{
An object of type void.
}
\description{
Documentation for SLiM function \code{skipTick}, which is a method of the SLiM
class \code{\link{Species}}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Documentation for this function can be found in the official
\href{http://benhaller.com/slim/SLiM_Manual.pdf#page=727}{SLiM manual: page
727}.

Deactivate the target species for the current tick. This sets
the active property of the species to F; it also set the active property
of all callbacks that belong to the species (with the species as their
species specifier) to F, and sets the active property of all events that are
synchronized with the species (with the species as their ticks specifier) to
F. The cycle counter for the species will not be incremented at the end of the
tick. This method may only be called in first() events, to ensure that species
are either active or inactive throughout a given tick.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016-2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\seealso{
Other Species: 
\code{\link{Sp}},
\code{\link{addSubpop}()},
\code{\link{addSubpopSplit}()},
\code{\link{countOfMutationsOfType}()},
\code{\link{individualsWithPedigreeIDs}()},
\code{\link{killIndividuals}()},
\code{\link{mutationCounts}()},
\code{\link{mutationFrequencies}()},
\code{\link{mutationsOfType}()},
\code{\link{outputFixedMutations}()},
\code{\link{outputFull}()},
\code{\link{outputMutations}()},
\code{\link{readFromPopulationFile}()},
\code{\link{recalculateFitness}()},
\code{\link{registerFitnessEffectCallback}()},
\code{\link{registerMateChoiceCallback}()},
\code{\link{registerModifyChildCallback}()},
\code{\link{registerMutationCallback}()},
\code{\link{registerMutationEffectCallback}()},
\code{\link{registerRecombinationCallback}()},
\code{\link{registerReproductionCallback}()},
\code{\link{registerSurvivalCallback}()},
\code{\link{simulationFinished}()},
\code{\link{subsetMutations}()},
\code{\link{treeSeqCoalesced}()},
\code{\link{treeSeqOutput}()},
\code{\link{treeSeqRememberIndividuals}()},
\code{\link{treeSeqSimplify}()}
}
\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
\concept{Species}
