% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_blocks.R
\name{slim_block_finish}
\alias{slim_block_finish}
\title{Generate simulationFinished Block}
\usage{
slim_block_finish(generation)
}
\arguments{
\item{generation}{Generation to end the simulation at.}
}
\value{
A \code{slimr_block} object
}
\description{
This generates a simple block to end a SLiM simulation.
}
\examples{
slim_script(
  slim_block_init_minimal(),
  slim_block_finish(100)
)
}
