% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_blocks.R
\name{slim_block_init_minimal}
\alias{slim_block_init_minimal}
\title{Generate minimal initialize() block}
\usage{
slim_block_init_minimal(
  mutation_rate = 1e-07,
  dominance = 0.5,
  selection = 0,
  dist_type = "f",
  genome_size = 1e+05,
  recombination_rate = 1e-08
)
}
\arguments{
\item{mutation_rate}{The overall mutation rate.}

\item{dominance}{The overall dominance value.}

\item{selection}{Mean selection strength for mutations}

\item{dist_type}{Distribution from which to draw mutation selection values (see
\code{\link{InitializeMutationType}} for possible values).}

\item{genome_size}{Genome size of the population, in number of loci.}

\item{recombination_rate}{Overall recombination rate.}
}
\value{
A \code{slimr_block} object
}
\description{
This function creates a minimal initialize block, allowing you to
set some basic parameters, but generally only allowing for one type
of mutation, which is distributed across the whole genome.
}
