% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_callbacks.R
\name{slim_callbacks}
\alias{slim_callbacks}
\title{SLiM Callbacks}
\usage{
slim_callbacks()
}
\description{
These functions are stubs that don't do anything in R, but
represent callback functions in SLiM. Call these functions
as part of a \code{\link{slim_block}} call to specify that the
code is designed to work with a particular callback.
Most callbacks are designed to be called within SLiM under special
circumstances, and will have 'pseudo-variables' that are accessible
only inside that particular callback. See help files for individuals
callbacks listed below for more details.
Available callbacks are:
\itemize{
\item{\code{\link{initialize}}}
\item{\code{\link{early}}}
\item{\code{\link{late}}}
\item{\code{\link{fitness}}}
\item{\code{\link{fitnessEffect}}}
\item{\code{\link{mateChoice}}}
\item{\code{\link{modifyChild}}}
\item{\code{\link{mutation}}}
\item{\code{\link{mutationEffect}}}
\item{\code{\link{recombination}}}
\item{\code{\link{interaction}}}
\item{\code{\link{reproduction}}}
\item{\code{\link{survival}}}
}
}
\seealso{
Other callbacks: 
\code{\link{early}()},
\code{\link{first}()},
\code{\link{fitnessEffect}()},
\code{\link{fitness}()},
\code{\link{initialize}()},
\code{\link{interaction}()},
\code{\link{late}()},
\code{\link{mateChoice}()},
\code{\link{modifyChild}()},
\code{\link{mutationEffect}()},
\code{\link{mutation}()},
\code{\link{recombination}()},
\code{\link{reproduction}()},
\code{\link{survival}()}
}
\concept{callbacks}
