% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_code.R
\name{slim_code_Rify}
\alias{slim_code_Rify}
\title{Rify some SLiM code}
\usage{
slim_code_Rify(code_snippet)
}
\arguments{
\item{code_snippet}{SLiM code to Rify as a character vector}
}
\value{
Rified code snippet
}
\description{
Utility code to convert SLiM code into a form that can be parsed by R
(e.g. in \code{styler} or \code{prettycode}). Don't forget to re-SLiMify
afterwards (via \code{\link{slim_code_SLiMify}} )!
}
