% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_code.R
\name{slim_code_SLiMify}
\alias{slim_code_SLiMify}
\title{Rify some SLiM code}
\usage{
slim_code_SLiMify(code_snippet, prettycode = FALSE)
}
\arguments{
\item{code_snippet}{SLiM code to Rify as a character vector}
}
\value{
SLiMified code snippet
}
\description{
Utility code to convert Rified (via \code{\link(slim_code_Rify)}) SLiM code back to valid SLiM code!
}
