% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_extract.R
\name{slim_extract_full}
\alias{slim_extract_full}
\title{Extract Elements from SLiM's outputFull()}
\usage{
slim_extract_full(
  output_full,
  type = c("mutations", "individuals", "genomes", "coordinates", "sexes", "ages",
    "full_individual"),
  join = TRUE,
  expand_mutations = FALSE
)
}
\arguments{
\item{output_full}{A character vector where each element is the result of a
call to \code{outputFull()} in SLiM}

\item{type}{Which type of data to return: "mutations", "individuals", "genomes", "coordinates",
"sexes", or "ages"}

\item{join}{If asking for multiple output type, should they be joined into one tibble (\code{join = TRUE})
or left as separate tibbles returned in a list (\code{join = FALSE})?}

\item{expand_mutations}{If asking for "genomes" output, should mutations be expanded into their own column (\code{expand_mutations = TRUE})
or left as a vector of mutation ids in a list column (\code{expand_mutations = FALSE})?}
}
\value{
A tibble
}
\description{
Extract Elements from SLiM's outputFull()
}
\examples{
if(slim_is_avail()) {
  test_sim <- slim_script(
    slim_block_init_minimal(mutation_rate = 1e-6),
    slim_block_add_subpops(1, 100),
    slim_block(1, 20, late(), {
      slimr_output(sim.outputFull(), "out", do_every = 10)
    })
  ) \%>\%
    slim_run()
  slim_extract_full(test_sim$output_data, type = "mutations")
}
}
