% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_extract.R
\name{slim_extract_genlight}
\alias{slim_extract_genlight}
\title{Extract data into a genlight object}
\usage{
slim_extract_genlight(x, ...)
}
\arguments{
\item{x}{\code{slimr_results} object containing data generated by \code{slimr_output} using \code{outputVCF()} or}

\item{...}{Arguments passed to or from other methods.
\code{outputFull()} in SLiM}
}
\value{
A \code{genlight} object
}
\description{
Extract data into a genlight object
}
