% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_extract.R
\name{slim_extract_genome}
\alias{slim_extract_genome}
\title{Extract Elements from SLiM's output() for genomes}
\usage{
slim_extract_genome(
  output,
  type = c("mutations", "genomes", "full"),
  join = TRUE,
  expand_mutations = FALSE
)
}
\arguments{
\item{output}{A character vector where each element is the result of a
call to \code{genomes.output()} in SLiM}

\item{type}{Which type of data to return: "mutations", or "genomes" or
both.}

\item{join}{If asking for multiple output type, should they be joined into one tibble (\code{join = TRUE})
or left as separate tibbles returned in a list (\code{join = FALSE})?}

\item{expand_mutations}{If asking for "genomes" output, should mutations be expanded into their own column (\code{expand_mutations = TRUE})
or left as a vector of mutation ids in a list column (\code{expand_mutations = FALSE})?}
}
\value{
A tibble
}
\description{
Extract Elements from SLiM's output() for genomes
}
\examples{
if(slim_is_avail()) {
  test_sim <- slim_script(
    slim_block_init_minimal(mutation_rate = 1e-6),
    slim_block_add_subpops(1, 100),
    slim_block(1, 20, late(), {
      slimr_output(p1.genomes.output(), "out", do_every = 10)
    })
  ) \%>\%
    slim_run()
  slim_extract_genome(test_sim$output_data, type = "mutations")
}
}
