% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_output.R
\name{slim_extract_output_data}
\alias{slim_extract_output_data}
\title{Extract data produced by SLiM}
\usage{
slim_extract_output_data(output)
}
\arguments{
\item{output}{Character vector produced from SLiM run.}
}
\value{
A \code{tibble} with four columns:
\describe{
\item{generation}{A vector of generations where output was produced.}
\item{name}{Names of the output data.}
\item{expression}{The SLiM expression used to generate the output.}
\item{data}{The raw data output from SLiM as a character vector.}
}
}
\description{
This function takes output produced from a \code{slimr_script} which uses
\code{\link{slimr_output}}, and converts it into a \code{tibble}
}
