% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_script_manip.R
\name{slim_find_block_starting_at}
\alias{slim_find_block_starting_at}
\title{Title}
\usage{
slim_find_block_starting_at(slim_script, start_gen)
}
\arguments{
\item{slim_script}{}
}
\value{

}
\description{
Title
}
