% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_recipes.R
\name{slim_get_recipe}
\alias{slim_get_recipe}
\title{Get a SLiM recipe}
\usage{
slim_get_recipe(recipe = 1)
}
\arguments{
\item{recipe}{Integer or Character specifying the recipe number, or the recipe name.}
}
\value{
A SLiM recipe as a length 1 character vector
}
\description{
Get a SLiM recipe
}
\examples{
slim_get_recipe(recipe = "4.1")
}
