% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_recipes.R
\name{slim_get_recipes}
\alias{slim_get_recipes}
\title{Get a list of SLiM recipes (as text)}
\usage{
slim_get_recipes(recipes = "all")
}
\arguments{
\item{recipes}{Integer or Character specifying the recipe numbers / names, or "all", to retrieve all recipes.}
}
\value{
List of SLiM recipes, each element a length 1 character vector.
}
\description{
Get a list of SLiM recipes (as text)
}
\examples{
recipes <- slim_get_recipes("all")
recipes[[12]]
}
