% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_setup.R
\name{slim_install_path}
\alias{slim_install_path}
\title{Find the SLiM executable path `slimr` is using}
\usage{
slim_install_path()
}
\description{
Returns the SLiM installation path used by `slimr` to run simulations.
}
\examples{
slim_install_path()

}
