% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_setup.R
\name{slim_is_avail}
\alias{slim_is_avail}
\title{Check if SLiM is installed and \code{slimr} can find it}
\usage{
slim_is_avail()
}
\value{
\code{TRUE} if SLiM is found and \code{FALSE} otherwise
}
\description{
Check if SLiM is installed and \code{slimr} can find it
}
\examples{
slim_is_avail()
}
