% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_code.R
\name{slim_load_globals}
\alias{slim_load_globals}
\title{Load SLiM Globals into R Global environment}
\usage{
slim_load_globals(
  max = 10,
  sim = TRUE,
  community = TRUE,
  self = TRUE,
  pseudo = TRUE
)
}
\arguments{
\item{max}{A single integer values indicating how many of
each numbered Globals to load, or a named integer vector where the names
refer to different Global types (see details for more information), and the value
refers to how many of that Global type to load.}

\item{sim}{Should the \code{sim} global be loaded?}

\item{community}{Should the \code{community} global be loaded?}

\item{self}{Should the \code{self} global be loaded?}

\item{pseudo}{Should 'pseudo-variables' be loaded? These are similar to SLiM
globals but are only available inside certain callbacks.}
}
\value{
None
}
\description{

}
\details{
By default this will load ten of each of the standard SLiM Globals (e.g. objects prefixed with "p",
"m", "i", "g", and "s"), as well as the "sim" and "self" singular Globals. Note: be careful with
this function. If you have any objects in your R session's global environment with these names
they will be overwritten.
}
\examples{
slim_load_globals(c(p = 4, g = 2))
}
