% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_input.R
\name{slim_make_pop_input}
\alias{slim_make_pop_input}
\title{Create a file to initialise a population in SliM}
\usage{
slim_make_pop_input(
  snps,
  file_name = tempfile(),
  sim_gen = 10000,
  ind_pops = NULL,
  ind_sex = NULL,
  mut_pos = NULL,
  mut_type = NULL,
  mut_sel = NULL,
  mut_dom = NULL,
  mut_pop = NULL,
  mut_gen = NULL,
  mut_prev = NULL,
  mut_nuc = NULL,
  ind_coord = NULL,
  ind_age = NULL,
  gen_type = NULL
)
}
\arguments{
\item{gen_type}{}
}
\value{

}
\description{
Make a SLiM file in the style of \code{outputFull()} which can be read by SLiM. This is useful to
start a simulation with a particular population state that you define in R
}
