% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_script_manip.R
\name{slim_modify_block_code}
\alias{slim_modify_block_code}
\title{Title}
\usage{
slim_modify_block_code(slim_script, block, what = NULL, where = NULL)
}
\arguments{
\item{where}{}
}
\value{

}
\description{
Title
}
