% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_run.R
\name{slim_open}
\alias{slim_open}
\title{Title}
\usage{
slim_open(slimr_script, slim_gui_path = Sys.getenv("slim_gui_path"))
}
\arguments{
\item{slimr_script}{A slimr_script object to open in SLiMGUI
(or QtSLiM if on Linux or Windows)}

\item{slim_gui_path}{Full path to SLiMGUI or QtSLiM executable}
}
\value{
Invisibly return the process object used to launch the
GUI (for debugging purposes)
}
\description{
Title
}
\examples{
if(interactive()) {
  slim_open(minimal_slimr_script())
}
}
