% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_output.R
\name{slim_output_extract_one}
\alias{slim_output_extract_one}
\title{Extract data from a single generation's output}
\usage{
slim_output_extract_one(output_one)
}
\arguments{
\item{output_one}{Lines output in a single generation from SLiM}
}
\value{
A four column tibble.
}
\description{
Extract data from a single generation's output
}
