% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_output.R
\name{slim_output_genlight}
\alias{slim_output_genlight}
\title{Extract data into a genlight object}
\usage{
slim_output_genlight(x, ...)
}
\arguments{
\item{x}{\code{slimr_results} object containing data generated by \code{slimr_output} using \code{outputVCF()} or
\code{outputFull()} in SLiM}

\item{name}{Name of the output to extract from. this should match the \code{name} argument of \code{slimr_output} used to generate the data}
}
\value{
A \code{genlight} object
}
\description{
Extract data into a genlight object
}
