% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_output.R
\name{slim_outputFull_extract}
\alias{slim_outputFull_extract}
\title{Extract Elements from SLiM's outputFull()}
\usage{
slim_outputFull_extract(
  output_full,
  type = c("mutations", "individuals", "genomes", "coordinates", "sexes", "ages")
)
}
\arguments{
\item{output_full}{A character vector where each element is the result of a
call to \code{outputFull()} in SLiM}

\item{type}{Which type of data to return: "mutations", "individuals", "genomes", "coordinates",
"sexes", or "ages"}
}
\value{
A tibble
}
\description{
Extract Elements from SLiM's outputFull()
}
