% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_output.R
\name{slim_outputGS_extract}
\alias{slim_outputGS_extract}
\title{Extract Elements from SLiM's output() for genomes}
\usage{
slim_outputGS_extract(
  output,
  type = c("mutations", "genomes", "full"),
  join = TRUE,
  expand_mutations = FALSE
)
}
\arguments{
\item{output}{A character vector where each element is the result of a
call to \code{genomes.output()} in SLiM}

\item{type}{Which type of data to return: "mutations", or "genomes" or
both.}

\item{join}{If asking for multiple output type, should they be joined into one tibble (\code{join = TRUE})
or left as separate tibbles returned in a list (\code{join = FALSE})?}

\item{expand_mutations}{If asking for "genomes" output, should mutations be expanded into their own column (\code{expand_mutations = TRUE})
or left as a vector of mutation ids in a list column (\code{expand_mutations = FALSE})?}
}
\value{
A tibble
}
\description{
Extract Elements from SLiM's output() for genomes
}
