% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{slim_recipes}
\alias{slim_recipes}
\title{Dataset of all recipes available in SLiM}
\format{
A List with 190 elements, each containing a character vector
}
\source{
\url{https://github.com/MesserLab/SLiM/releases/download/v4.1/SLiM_Manual.pdf}
}
\usage{
slim_recipes
}
\description{
A Dataset containing all SLiM recipes found in the SLiM Manual
}
\keyword{datasets}
