% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_code.R
\name{slim_register_event_code}
\alias{slim_register_event_code}
\title{Title}
\usage{
slim_register_event_code(
  code_snippet,
  start_gen = 1,
  last_gen = NULL,
  auto_bracket = TRUE
)
}
\arguments{
\item{last_gen}{}
}
\value{

}
\description{
Title
}
