% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_extract.R
\name{slim_results_to_data}
\alias{slim_results_to_data}
\title{Convert output data from a slimr_results object to a tibble}
\usage{
slim_results_to_data(dat, generations = NULL)
}
\arguments{
\item{dat}{A \code{slimr_results} object to extract data from.
You can alternatively specify the \code{output_data} from the \code{slimr_results}
directly in this parameter.}

\item{generations}{For what generations do you want to extract data?
Default is all generations that have data. The special value \code{0}
can be used to specify just the most recent generation in the data.}
}
\value{
A tibble with three columns, name: the name of the outputs,
generation: the generations of the outputs, and data: the outputs as converted data.
This final column will be a list column, where each element will usually
be a vector or a tibble, depending on what kind of data was returned
by the simulation.
}
\description{
This function tries to automate the process of converting
output in the \code{slimr_results} object returned by
\code{\link{slim_run}} function into usable data in the
form of a \code{tibble}. If this process fails you will
end up with the data as a character string, and you will
have to manually convert this into something you can use.
}
