% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_run_script.R
\name{slim_run_script_template}
\alias{slim_run_script_template}
\title{Title}
\usage{
slim_run_script_template(slim_script_template, param_list, slim_path = NULL)
}
\arguments{
\item{slim_script_template}{A slim script with sections enclosed in '<<' and '>>'. Anything between double
triangular brackets will be replaced by its corresponding named element in \code{param_list}. For example,
'<<x>>' will be replace by the value \code{param_list$x}}

\item{param_list}{}
}
\value{

}
\description{
Title
}
