% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_scripting.R
\name{slim_script}
\alias{slim_script}
\title{Create a SLiMR script}
\usage{
slim_script(...)
}
\arguments{
\item{...}{A list of \code{\link{slim_block}} objects comprising a SLiM script
(written in slimr code)}
}
\value{
A \code{slim_script} object that can be used with
\code{\link{slim_run}} or converted into a text file for use with
SLiM directly using \code{\link{as_slim_text}}.
}
\description{
Setup a SLiMR script. Each argument should be a call to \code{\link{slim_block}}.
See details for more information.
}
\examples{
slim_script(
   slim_block(initialize(),
          {
            initializeMutationRate(1e-7);
            initializeMutationType("m1", 0.5, "f", 0.0);
            initializeGenomicElementType("g1", m1, 1.0);
            initializeGenomicElement(g1, 0, 99999);
            initializeRecombinationRate(1e-8);
          }),
   slim_block(1,
          {
            sim.addSubpop("p1", 500);
          }),
   slim_block(10000,
          {
            sim.simulationFinished();
          })
)
}
