% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_script_manip.R
\name{slim_script_add_code}
\alias{slim_script_add_code}
\title{Add code into a block in a slim_script.}
\usage{
slim_script_add_code(slim_script, block, what = NULL, where = NULL)
}
\arguments{
\item{slim_script}{\code{slim_script} object to modify}

\item{block}{Block id to add code into}

\item{what}{Code to add into script, of class character}

\item{where}{Where in the block should it be added? Can be either line number, or a character specifying "start" or "end". Default is "end".}
}
\value{
Modified \code{slim_script} object
}
\description{
Add code into a block in a slim_script.
}
