% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_script_manip.R
\name{slim_script_change_duration}
\alias{slim_script_change_duration}
\title{Function to change a slimr_script object's duration}
\usage{
slim_script_change_duration(slim_script, duration)
}
\arguments{
\item{slim_script}{slimr_script object to change the duration of}

\item{duration}{New duration for the script, in terms of the last generation that
should be run.}
}
\description{
Function to change a slimr_script object's duration
}
