% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_script.R
\name{slim_script_duration}
\alias{slim_script_duration}
\title{Set the duration of a slimr_script}
\usage{
slim_script_duration(x, duration)
}
\arguments{
\item{x}{slimr_script object to set the duration of}

\item{duration}{Final generation number to run for this simulation}
}
\value{
A \code{slimr_script} object with new duration
}
\description{
Set the duration of a slimr_script
}
