% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_script_manip.R
\name{slim_script_from_text}
\alias{slim_script_from_text}
\title{Title}
\usage{
slim_script_from_text(slim_script_text)
}
\arguments{
\item{slim_script_text}{A character vector of length 1 giving the full SLiM script to convert to a \code{slim_script} object}
}
\value{
A \code{slim_script} object.
}
\description{
Title
}
