% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_scripting.R
\name{slim_script_render_dots}
\alias{slim_script_render_dots}
\title{Version of `slim_script_render()` that takes dots (named arguments) instead of a
template list}
\usage{
slim_script_render_dots(
  slimr_script,
  ...,
  replace_NAs = TRUE,
  reps = 1,
  parallel = FALSE,
  portable = FALSE
)
}
\arguments{
\item{slimr_script}{The \code{slimr_script} object to be rendered}

\item{...}{Named arguments that will be used to replace templated variables in
the script. The names of the arguments must match the names of the templated
variables.}

\item{replace_NAs}{Should \code{NA} values in the template be replaced by their default values?}

\item{reps}{Should the rendered script be replicated? If greater than 1, a \code{slimr_script_coll}
will be returned. This can also be used with a \code{slimr_script} object that has already been
rendered, in which case it will just repeat the rendered script in the result.}

\item{parallel}{Should the rendering be done in parallel when rendering multiple scripts? Requires
the \code{furrr} package and will use the plan set by \code{future::\link[future]{plan}}}

\item{portable}{If `TRUE`, the the script will be rendered in a 'portable' format, which allows
the script to be modified in another program such as SLiMGUI, and then reimported into R, while maintaining
`slimr` features. This has not been implemented yet and so this is currently not used.}
}
\value{
A rendered `slimr_script_coll` object
}
\description{
Version of `slim_script_render()` that takes dots (named arguments) instead of a
template list
}
\examples{
test_sim <- slim_script(
  slim_block_init_minimal(mutation_rate = r_template("mu", 1e-7)),
  slim_block_add_subpops(1, 100),
  slim_block(1, 20, late(), {
    r_output(sim.outputFull(), "out", do_every = 10)
  })
) \%>\%
  slim_script_render_dots(mu = c(1e-7, 1e-6, 1e-5))
test_sim
}
