% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_setup.R
\name{slim_setup}
\alias{slim_setup}
\title{Attempt to install and / or setup SLiM for use with slimr}
\usage{
slim_setup(install_dir = "~/slim", test_slim = TRUE, verbose = TRUE)
}
\arguments{
\item{install_dir}{Directory to install SLiM to. If "default" `slim_setup()` will install in the default
directory. Be carefult to make sure you have write and execution permissions for the installation folder
you specify. We recommend using the default directory "~/slim", in which case you will not have to set an
environmental variable to tell slimr where to find your slim installation.}

\item{test_slim}{Should SLiM be tested once installation is complete?}

\item{verbose}{Whether to print out progress of the installation.}
}
\description{
`slim_setup()` will attempt to determine the user's OS and install SLiM automatically.
Note that on Windows, you must have Windows subsystem for linux installed, which requires at least Windows 10.
}
\examples{
\dontrun{
slim_setup()
}
}
