% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_template.R
\name{slim_template_info}
\alias{slim_template_info}
\title{Get information on templating in a slimr_script}
\usage{
slim_template_info(script_temp)
}
\arguments{
\item{script_temp}{A templated \code{slimr_script} to retrieve information from}
}
\value{
A list of lists. The top-level is named for the blocks in which templated variables exist.
For each block with templated variables the element is a list named with all templated variables
in that block, and its values are equal to the default values for those variables.
}
\description{
Returns information on templated variables and their default values in a \code{slimr_script}
}
\examples{
test_sim <- slim_script(
  slim_block_init_minimal(mutation_rate = r_template("mu", 1e-7)),
  slim_block_add_subpops(1, 100),
  slim_block(1, 20, late(), {
    slimr_output(sim.outputFull(), "out", do_every = 10)
  })
)
}
