% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_code.R
\name{slim_unload_globals}
\alias{slim_unload_globals}
\title{Unload SLiM Globals from R Global environment}
\usage{
slim_unload_globals()
}
\description{
This will remove any objects added to the global environment by
\code{\link{slim_load_globals}}
}
\examples{
slim_load_globals(c(p = 1))
slim_unload_globals()
}
