% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_setup.R
\name{slim_version}
\alias{slim_version}
\title{Find out what version of SLiM is being used by \code{slimr}}
\usage{
slim_version()
}
\value{
A character string with the version number
}
\description{
Find out what version of SLiM is being used by \code{slimr}
}
\examples{
if(slim_is_avail()) {
  slim_version()
}
}
