% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\docType{data}
\name{SLiMBuiltin}
\alias{SLiMBuiltin}
\alias{.SB}
\title{SLiMBuiltin}
\format{
An object of class \code{list} of length 7.
}
\usage{
SLiMBuiltin
}
\description{
Documentation for SLiMBuiltin class from SLiM
}
\details{
This class has the following methods (functions):
\itemize{
\item{\code{\link{codonsToAminoAcids}}}
\item{\code{\link{mm16To256}}}
\item{\code{\link{mmJukesCantor}}}
\item{\code{\link{mmKimura}}}
\item{\code{\link{nucleotideCounts}}}
\item{\code{\link{nucleotideFrequencies}}}
\item{\code{\link{nucleotidesToCodons}}}
}
This class has the following properties:
\describe{
\item{None.}{This class has no properties.}
}
}
\keyword{datasets}
