% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_classes.R, R/slim_lang.R
\name{SB}
\alias{SB}
\alias{SLiMBuiltin}
\title{SLiMBuiltin}
\description{
Documentation for SLiMBuiltin class from SLiM
}
\details{
SLiM provides a small number of built-in functions, documented here.
Note that these are
distinct from the functions built into the Eidos language itself, which are
documented in the Eidos
manual.
This class has the following methods (functions):
\itemize{
\item{\code{\link{codonsToAminoAcids}}}
\item{\code{\link{mm16To256}}}
\item{\code{\link{mmJukesCantor}}}
\item{\code{\link{mmKimura}}}
\item{\code{\link{nucleotideCounts}}}
\item{\code{\link{nucleotideFrequencies}}}
\item{\code{\link{nucleotidesToCodons}}}
\item{\code{\link{calcFST}}}
\item{\code{\link{calcHeterozygosity}}}
\item{\code{\link{calcInbreedingLoad}}}
\item{\code{\link{calcPairHeterozygosity}}}
\item{\code{\link{calcVA}}}
\item{\code{\link{calcWattersonsTheta}}}
\item{\code{\link{summarizeIndividuals}}}
\item{\code{\link{treeSeqMetadata}}}
}
This class has the following properties:
\describe{
\item{None.}{This class has no properties.}
}
}
\seealso{
Other SLiMBuiltin: 
\code{\link{calcFST}()},
\code{\link{calcHeterozygosity}()},
\code{\link{calcInbreedingLoad}()},
\code{\link{calcPairHeterozygosity}()},
\code{\link{calcVA}()},
\code{\link{calcWattersonsTheta}()},
\code{\link{codonsToAminoAcids}()},
\code{\link{mm16To256}()},
\code{\link{mmJukesCantor}()},
\code{\link{mmKimura}()},
\code{\link{nucleotideCounts}()},
\code{\link{nucleotideFrequencies}()},
\code{\link{nucleotidesToCodons}()},
\code{\link{summarizeIndividuals}()},
\code{\link{treeSeqMetadata}()}
}
\concept{SLiMBuiltin}
