% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_classes.R, R/slim_lang.R
\name{SG}
\alias{SG}
\alias{SLiMgui}
\title{SLiMgui}
\description{
Documentation for SLiMgui class from SLiM
}
\details{
This class represents the SLiMgui application. When running under
SLiMgui, a global object
singleton constant of class SLiMgui will be defined, named slimgui. This object
can be used to
query and control the SLiMgui application. When running at the command line, the
slimgui
object will not exist; to determine whether the simulation is running under
SLiMgui, one may
therefore test exists("slimgui"). If a model needs to run both at the command
line and under
SLiMgui, all uses of the slimgui object should be protected by if
(exists("slimgui")) to avoid
errors.
This class has the following methods (functions):
\itemize{
\item{\code{\link{openDocument}}}
\item{\code{\link{pauseExecution}}}
}
This class has the following properties:
\describe{
\item{pid}{A property of type integer. It is of length one (a singleton). This
property is a constant, so it is not modifiable. \strong{Property Description:}
The Un*x process identifier (commonly called the "pid") of the running SLiMgui
application. This can be useful for scripts that wish to use system calls to
influence the SLiMgui application.}

}
}
\seealso{
Other SLiMgui: 
\code{\link{openDocument}()},
\code{\link{pauseExecution}()}
}
\concept{SLiMgui}
