% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{slimr_clip_original}
\alias{slimr_clip_original}
\title{Copy original slimrlang code used to create a slimr_script object to the clipboard}
\usage{
slimr_clip_original(slimr_script_name = NULL)
}
\arguments{
\item{slimr_script_name}{Name of slimr_script to copy its original code
to the clipboard, as a string}
}
\value{
The code as a string (invisibly)
}
\description{
Copy original slimrlang code used to create a slimr_script object to the clipboard
}
