% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_script.R
\name{slimr_code}
\alias{slimr_code}
\title{Create a \code{slimr_code} object}
\usage{
slimr_code(...)
}
\arguments{
\item{...}{A list of character vectors.}
}
\description{
Create a \code{slimr_code} object
}
\examples{
slimr_code('print("Hello")', 'print("World!")')
}
