% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_inline.R
\name{slimr_inline}
\alias{slimr_inline}
\title{Insert R objects into SLiM scripts}
\usage{
slimr_inline(object)
}
\arguments{
\item{object}{R object to inline into the SLiM script.}
}
\value{
A character vector with the code generated for inlining.
}
\description{
By calling this function inside a \code{\link{slim_block}} function call you can insert R objects
into the script using direct inlining. This function should generally only be used within a
\code{\link{slim_block}} call
}
\details{
Currently supported R objects include all atomic vectors, matrices and arrays. Non-atomic vectors like factors are
currently not supported and neither are any other special object types, though we plan to support some in the future.
}
