% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_inline.R
\name{slimr_name}
\alias{slimr_name}
\title{Convert a string to a SLiM object name}
\usage{
slimr_name(name)
}
\arguments{
\item{name}{A string with the SLiM object name}
}
\value{
A symbol
}
\description{
Convert a string to a SLiM object name
}
\examples{
slimr_name("p1")
}
