% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{slimr_open_original}
\alias{slimr_open_original}
\title{Open original slimrlang code used to create a slimr_script object in a new document}
\usage{
slimr_open_original(slimr_script_name = NULL)
}
\arguments{
\item{slimr_script_name}{Name of slimr_script to open in a new document as a string.}
}
\value{
The code as a string (invisibly)
}
\description{
Open the original code used to create a slimr_script in a new document. Requires RStudio.
}
