% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_output.R
\name{slimr_output_coords}
\alias{slimr_output_coords}
\title{Utility function to tell SLiM to output coordinates from spatial simulations}
\usage{
slimr_output_coords(dimensionality = c("x", "xy", "xyz"), ...)
}
\arguments{
\item{dimensionality}{What dimensionality should be output? Can be
"x", "xy", or "xyz".}

\item{...}{Other arguments to be passed to \code{\link{slimr_output}}}
}
\value{
None
}
\description{
Utility function to tell SLiM to output coordinates from spatial simulations
}
\details{
Outputs x, y, and z coordinates as separate entries in \code{slimr_results},
with names "x", "y", and "z".
}
