% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_output.R
\name{slimr_output_full}
\alias{slimr_output_full}
\title{Utility function to tell SLiM to output its outputFull() output}
\usage{
slimr_output_full(name = "full_output", ...)
}
\arguments{
\item{name}{Name of output to use to label it in \code{slimr_results object}. Default is \code{"full_output"}}

\item{...}{Other arguments to be passed to \code{\link{slimr_output}}}
}
\description{
Utility function to tell SLiM to output its outputFull() output
}
