% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_output.R
\name{slimr_output_nucleotides}
\alias{slimr_output_nucleotides}
\title{Utility function to tell SLiM to output Nucleotides}
\usage{
slimr_output_nucleotides(
  name = "seqs",
  subpops = FALSE,
  both_genomes = FALSE,
  inds = NULL,
  ...
)
}
\arguments{
\item{name}{Name of output to use to label it in \code{slimr_results object}. Default is \code{"seqs"}.}

\item{subpops}{Should the subpopulation of each sequence be outputted as well?}

\item{inds}{SLiM expression that returns the individuals to get nucleotides from. By default all
individuals are returned.}

\item{...}{Other arguments to be passed to \code{\link{slimr_output}}}
}
\value{
None
}
\description{
Utility function to tell SLiM to output Nucleotides
}
