% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_output.R
\name{slimr_output_sex}
\alias{slimr_output_sex}
\title{Utility function to tell SLim to output sexes of individuals}
\usage{
slimr_output_sex(name = "sex", ...)
}
\arguments{
\item{name}{Name of output to use to label it in \code{slimr_results object}. Default is \code{"sex"}.}

\item{...}{Other arguments to be passed to \code{\link{slimr_output}}}
}
\description{
Utility function to tell SLim to output sexes of individuals
}
