% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_script.R
\name{slimr_script_coll}
\alias{slimr_script_coll}
\title{Make a new \code{slimr_script_coll} object}
\usage{
slimr_script_coll(...)
}
\arguments{
\item{...}{A list of \code{slir_script} objects}
}
\value{
A \code{slimr_script_coll} object
}
\description{
Make a new \code{slimr_script_coll} object
}
