% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{slimr_which}
\alias{slimr_which}
\title{Get the proper path for the SLiM executable. Returns "" if it cannot be found}
\usage{
slimr_which(slim_path, os = c("linux", "osx", "windows"))
}
\arguments{
\item{slim_path}{Path to SLiM executable}

\item{os}{What operating system are we working in? Must be "windows", "osx", or "linux#', if left blank slimr will try and detrmine it automatically}
}
\value{
The full path to the executable as a character, or "" if not found.
}
\description{
Get the proper path for the SLiM executable. Returns "" if it cannot be found
}
