% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_script.R
\name{slimr_write}
\alias{slimr_write}
\title{Write a slimr_script object to a text file}
\usage{
slimr_write(x, file, ...)
}
\arguments{
\item{x}{\code{slimr_script} object to write to file}

\item{file}{File path to write to}

\item{...}{Further arguments to be passed to or from other objects}
}
\value{
Returns \code{x}, invisibly
}
\description{
Write a slimr_script object to a text file, which can be run in
SLiM as a SLiM script
}
\examples{
slim_script(
   slim_block(initialize(),
          {
            initializeMutationRate(1e-7);
            initializeMutationType("m1", 0.5, "f", 0.0);
            initializeGenomicElementType("g1", m1, 1.0);
            initializeGenomicElement(g1, 0, 99999);
            initializeRecombinationRate(1e-8);
          }),
   slim_block(1,
          {
            sim.addSubpop("p1", 500);
          }),
   slim_block(10000,
          {
            sim.simulationFinished();
          })
) -> script
temp_file <- tempfile(fileext = ".txt")
slimr_write(script, temp_file)
readLines(temp_file)
}
