% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_classes.R
\name{SS}
\alias{SS}
\title{SLiMSim was a class used prior to v4.0 in SLiM. It remains in `slimr`
for backwards compatibility but its use is now deprecated.}
\description{
SLiMSim was a class used prior to v4.0 in SLiM. It remains in `slimr`
for backwards compatibility but its use is now deprecated.
}
