% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{spatialMapColor}
\alias{spatialMapColor}
\alias{Subpopulation$spatialMapColor}
\alias{.P$spatialMapColor}
\title{SLiM method spatialMapColor}
\usage{
spatialMapColor(name, value)
}
\arguments{
\item{name}{An object of type string or numeric. Must be of length 1 (a
singleton). See details for description.}

\item{value}{An object of type string or numeric. See details for description.}
}
\value{
An object of type string.
}
\description{
Documentation for SLiM function \code{spatialMapColor}, which is a method of the
SLiM class \code{Subpopulation}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Looks up the spatial map indicated by name, and uses its color-
translation machinery (as defined by the valueRange and colors parameters to
defineSpatialMap()) to translate each element of value into a corresponding
color string. If the spatial map does not have color-translation capabilities,
an error will result. See the documentation for defineSpatialMap() for
information regarding the details of color translation. See the Eidos manual for
further information on color strings.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016–2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
