% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{spatialMapImage}
\alias{spatialMapImage}
\alias{Subpopulation$spatialMapImage}
\alias{.P$spatialMapImage}
\title{SLiM method spatialMapImage}
\usage{
spatialMapImage(name, width, height, centers, color)
}
\arguments{
\item{name}{An object of type string. Must be of length 1 (a singleton). See
details for description.}

\item{width}{An object of type null or integer. Must be of length 1 (a
singleton). The default value is \code{NULL}. See details for description.}

\item{height}{An object of type null or integer. Must be of length 1 (a
singleton). The default value is \code{NULL}. See details for description.}

\item{centers}{An object of type logical. Must be of length 1 (a singleton). The
default value is \code{F}. See details for description.}

\item{color}{An object of type logical. Must be of length 1 (a singleton). The
default value is \code{T}. See details for description.}
}
\value{
An object of type Image object. Return will be of length 1 (a singleton)
}
\description{
Documentation for SLiM function \code{spatialMapImage}, which is a method of the
SLiM class \code{\link{Subpopulation}}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Documentation for this function can be found in the official
\href{http://benhaller.com/slim/SLiM_Manual.pdf#page=745}{SLiM manual: page
745}.

This method has been deprecated, and may be removed in a future release
of SLiM. In SLiM 4.1 and later, use the SpatialMap method mapImage() instead,
and see that method's documentation. (This method differs only in taking a name
parameter, which is used to look up the spatial map from those that have been
added to the subpopulation.)
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016-2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\seealso{
Other Subpopulation: 
\code{\link{P}},
\code{\link{addCloned}()},
\code{\link{addCrossed}()},
\code{\link{addEmpty}()},
\code{\link{addRecombinant}()},
\code{\link{addSelfed}()},
\code{\link{addSpatialMap}()},
\code{\link{cachedFitness}()},
\code{\link{configureDisplay}()},
\code{\link{defineSpatialMap}()},
\code{\link{outputMSSample}()},
\code{\link{outputSample}()},
\code{\link{outputVCFSample}()},
\code{\link{pointDeviated}()},
\code{\link{pointInBounds}()},
\code{\link{pointPeriodic}()},
\code{\link{pointReflected}()},
\code{\link{pointStopped}()},
\code{\link{pointUniform}()},
\code{\link{removeSpatialMap}()},
\code{\link{removeSubpopulation}()},
\code{\link{sampleIndividuals}()},
\code{\link{setCloningRate}()},
\code{\link{setMigrationRates}()},
\code{\link{setSelfingRate}()},
\code{\link{setSexRatio}()},
\code{\link{setSpatialBounds}()},
\code{\link{setSubpopulationSize}()},
\code{\link{spatialMapColor}()},
\code{\link{spatialMapValue}()},
\code{\link{subsetIndividuals}()},
\code{\link{takeMigrants}()}
}
\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
\concept{Subpopulation}
