% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{spatialMapValue}
\alias{spatialMapValue}
\alias{Subpopulation$spatialMapValue}
\alias{.P$spatialMapValue}
\title{SLiM method spatialMapValue}
\usage{
spatialMapValue(name, point)
}
\arguments{
\item{name}{An object of type string or float. Must be of length 1 (a
singleton). See details for description.}

\item{point}{An object of type string or float. See details for description.}
}
\value{
An object of type float.
}
\description{
Documentation for SLiM function \code{spatialMapValue}, which is a method of the
SLiM class \code{Subpopulation}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Looks up the spatial map indicated by name, and uses its mapping
machinery (as defined by the gridSize, values, and interpolate parameters to
defineSpatialMap()) to translate the coordinates of point into a corresponding
map value. The length of point must be equal to the spatiality of the spatial
map; in other words, for a spatial map with spatiality "xz", point must be
of length 2, specifying the x and z coordinates of the point to be evaluated.
Interpolation will automatically be used if it was enabled for the spatial map.
Point coordinates are clamped into the range defined by the spatial boundaries,
even if the spatial boundaries are periodic; use pointPeriodic() to wrap the
point coordinates first if desired. See the documentation for defineSpatialMap()
for information regarding the details of value mapping. Beginning in SLiM 3.3,
point may contain more than one point to be looked up. In this case, the length
of point must be an exact multiple of the spatiality of the spatial map; for a
spatial map with spatiality "xz", for example, the length of point must be an
exact multiple of 2, and successive pairs of elements from point (elements 0 and
1, then elements 2 and 3, etc.) will be taken as the x and z coordinates of the
points to be evaluated. This allows spatialMapValue() to be used in a vectorized
fashion.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016–2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
