% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{speciesWithIDs}
\alias{speciesWithIDs}
\alias{Community$speciesWithIDs}
\alias{.Co$speciesWithIDs}
\title{SLiM method speciesWithIDs}
\usage{
speciesWithIDs(ids)
}
\arguments{
\item{ids}{An object of type integer. See details for description.}
}
\value{
An object of type Species object.
}
\description{
Documentation for SLiM function \code{speciesWithIDs}, which is a method of the
SLiM class \code{\link{Community}}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Documentation for this function can be found in the official
\href{http://benhaller.com/slim/SLiM_Manual.pdf#page=668}{SLiM manual: page
668}.

Find and return the Species objects with id values matching the values
in ids. If no matching Species object can be found with a given id, an error
results.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016-2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\seealso{
Other Community: 
\code{\link{Co}},
\code{\link{createLogFile}()},
\code{\link{deregisterScriptBlock}()},
\code{\link{genomicElementTypesWithIDs}()},
\code{\link{interactionTypesWithIDs}()},
\code{\link{mutationTypesWithIDs}()},
\code{\link{outputUsage}()},
\code{\link{registerEarlyEvent}()},
\code{\link{registerFirstEvent}()},
\code{\link{registerInteractionCallback}()},
\code{\link{registerLateEvent}()},
\code{\link{rescheduleScriptBlock}()},
\code{\link{scriptBlocksWithIDs}()},
\code{\link{simulationFinished}()},
\code{\link{subpopulationsWithIDs}()},
\code{\link{usage}()}
}
\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
\concept{Community}
