% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{transpose_coordinates}
\alias{transpose_coordinates}
\title{Transpose coordinates in the SLiM 0-based coordinate system into
realistic coordinate system.}
\usage{
transpose_coordinates(sim_coords, real_coords)
}
\arguments{
\item{sim_coords}{GRanges object with SLiM, 0-based coordinates.}

\item{real_coords}{GRanges object with real coordinate.}
}
\value{
GRanges object with transposed coordinates.
}
\description{
SLiM simulates all mutations as part of a single continuous "chromosome".
Therefore, simulating fixed mutations that are present in a real genome
requires their transposition into a 0-based continuous coordinate system.
This function performs this transposition based on the given positions of
the sites in their original coordinate system.
}
\details{
The real_coords GRanges object must contain a DataFrame object with three
columns specifying a realistic chromosome, start and end coordinate of
a region.
}
\seealso{
read_coordinates
}
