% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_output.R
\name{transpose_sites}
\alias{transpose_sites}
\title{Transpose the simulated sites in the SLiM 0-based coordinate system into
the realistic coordinates.}
\usage{
transpose_sites(sim_sites, real_sites)
}
\description{
Transpose the simulated sites in the SLiM 0-based coordinate system into
the realistic coordinates.
}
