% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariationData.R
\name{validateVariants}
\alias{validateVariants}
\title{validateVariants}
\usage{
validateVariants(df, fasta, nodeN = 8)
}
\arguments{
\item{df}{A data.frame or data.table containing minimally the following
columns: 1. uniprotAccession 2. wtAA (wild-type amino acid) 3. pos (the position
of the wild-type amino acid in the protein sequence)}

\item{fasta}{AAStringSet object of protein sequences, in which names of the list items
should correspond to the uniprotAccession column of the 'df' input}

\item{nodeN}{Number of cores to use to parallelise the run}
}
\value{
A subset of the initial data.frame or data.table object such that the
  given positions of amino acids match the residues in the fasta sequences
}
\description{
This function is used to validate whether the missense variants are consistent
with the fasta sequences of the proteins
}
