% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{addEmpty}
\alias{addEmpty}
\alias{Subpopulation$addEmpty}
\alias{.P$addEmpty}
\title{SLiM method addEmpty}
\usage{
addEmpty(sex)
}
\arguments{
\item{sex}{An object of type null or float or string. Must be of length 1 (a
singleton). The default value is \code{NULL}. See details for description.}
}
\value{
An object of type null or Individual object. Return will be of length 1
(a singleton)
}
\description{
Documentation for SLiM function \code{addEmpty}, which is a method of the SLiM
class \code{Subpopulation}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Generates a new offspring individual with empty genomes (i.e.,
containing no mutations), queues it for addition to the target subpopulation,
and returns it. The new offspring will not be visible as a member of the target
subpopulation until the end of the offspring generation life cycle stage. No
recombination() or mutation() callbacks will be called. The target subpopulation
will be used to locate applicable modifyChild() callbacks governing the
generation of the offspring individual (unlike the other addX() methods, because
there is no parental individual to reference). The offspring is considered
to have no parents for the purposes of pedigree tracking. The sex parameter
is treated as in addCrossed(). Note that this method is only for use in nonWF
models. See addCrossed() for further general notes on the addition of new
offspring individuals.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016–2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
