% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability_parameter.R
\name{calculate_cluster_variability}
\alias{calculate_cluster_variability}
\title{Calculate Cluster Variability}
\usage{
calculate_cluster_variability(data.features, features)
}
\arguments{
\item{data.features}{Data frame containing expression data and cluster labels}

\item{features}{Feature names to include in analysis}
}
\value{
Numeric value representing cluster separation strength
}
\description{
Measures the degree of separation between different cell clusters
based on expression patterns.
}
\seealso{
Other Section_1_Functions_Use_in_Package: 
\code{\link{calculate_expression}()},
\code{\link{calculate_expression_skewness}()},
\code{\link{calculate_probability}()},
\code{\link{estimate_batch_effect}()},
\code{\link{extract_dataset_features}()},
\code{\link{generate_training_data}()},
\code{\link{postprocess_parameters}()},
\code{\link{predict_optimal_parameters}()},
\code{\link{train_parameter_model}()}
}
\concept{Section_1_Functions_Use_in_Package}
