% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability_parameter.R
\name{calculate_expression_skewness}
\alias{calculate_expression_skewness}
\title{Calculate Expression Distribution Skewness (Use in package)}
\usage{
calculate_expression_skewness(expression_matrix)
}
\arguments{
\item{expression_matrix}{Matrix of expression values}
}
\value{
Mean absolute skewness across all genes
}
\description{
Computes the average skewness of gene expression distributions
across all features.
}
\seealso{
Other Section_1_Functions_Use_in_Package: 
\code{\link{calculate_cluster_variability}()},
\code{\link{calculate_expression}()},
\code{\link{calculate_probability}()},
\code{\link{estimate_batch_effect}()},
\code{\link{extract_dataset_features}()},
\code{\link{generate_training_data}()},
\code{\link{postprocess_parameters}()},
\code{\link{predict_optimal_parameters}()},
\code{\link{train_parameter_model}()}
}
\concept{Section_1_Functions_Use_in_Package}
