% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cellmarker2.R
\docType{data}
\name{Cellmarker2}
\alias{Cellmarker2}
\title{Cellmarker2 dataset}
\format{
A data frame with 8 columns:
}
\source{
\url{http://117.50.127.228/CellMarker/}
}
\usage{
Cellmarker2
}
\description{
A dataset containing marker genes for different cell types from Cellmarker2
}
\details{
This dataset is used to filter and create a standardized marker list.
The dataset can be filtered based on species, tissue class, tissue type,
cancer type, and cell type to generate a list of marker genes for specific cell types.
}
\keyword{datasets}
