% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database_Cellmarker2_raw.R
\docType{data}
\name{Cellmarker2_raw}
\alias{Cellmarker2_raw}
\title{Cellmarker2 raw dataset}
\format{
A data frame with 20 columns contined in the Cellmarker2 database:
}
\source{
\url{http://117.50.127.228/CellMarker/}
}
\usage{
Cellmarker2_raw
}
\description{
A dataset containing marker genes for different cell types from Cellmarker2
}
\details{
This dataset is used to filter and create a standardized marker list.
The dataset can be filtered based on species, tissue class, tissue type,
cancer type, and cell type to generate a list of marker genes for specific cell types.
}
\seealso{
Other Section_0_Database:
\code{\link{Cellmarker2}},
\code{\link{Cellmarker2_table}},
\code{\link{Markers_list_PCTAM}},
\code{\link{Markers_list_PCTIT}},
\code{\link{Markers_list_TCellSI}},
\code{\link{Markers_list_scIBD}},
\code{\link{PanglaoDB}},
\code{\link{PanglaoDB_raw}},
\code{\link{PanglaoDB_table}},
\code{\link{ScType}},
\code{\link{ScType_raw}},
\code{\link{ScType_table}}
}
\concept{Section_0_Database}
\keyword{datasets}
