% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cellmarker2_table.R
\docType{data}
\name{Cellmarker2_table}
\alias{Cellmarker2_table}
\title{Cellmarker2 table}
\format{
A list contain different types like species, tissue_class, tissue_type,
cancer_type, cell_type
}
\source{
\url{http://117.50.127.228/CellMarker/}
}
\usage{
Cellmarker2_table
}
\description{
A dataset containing marker genes for different cell types from Cellmarker2
}
\details{
This list is used to choose filters for creation of standardized marker list.
}
\keyword{datasets}
