% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database_Cellmarker2_table.R
\docType{data}
\name{Cellmarker2_table}
\alias{Cellmarker2_table}
\title{Cellmarker2 table}
\format{
A list contain different types like species, tissue_class, tissue_type,
cancer_type, cell_type
}
\source{
\url{http://117.50.127.228/CellMarker/}
}
\usage{
Cellmarker2_table
}
\description{
A dataset containing marker genes for different cell types from Cellmarker2
}
\details{
This list is used to choose filters for creation of standardized marker list.
}
\seealso{
Other Section_0_Database:
\code{\link{Cellmarker2}},
\code{\link{Cellmarker2_raw}},
\code{\link{Markers_list_PCTAM}},
\code{\link{Markers_list_PCTIT}},
\code{\link{Markers_list_TCellSI}},
\code{\link{Markers_list_scIBD}},
\code{\link{PanglaoDB}},
\code{\link{PanglaoDB_raw}},
\code{\link{PanglaoDB_table}},
\code{\link{ScType}},
\code{\link{ScType_raw}},
\code{\link{ScType_table}}
}
\concept{Section_0_Database}
\keyword{datasets}
