% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celltype_Annotation.R
\name{Celltype_Annotation}
\alias{Celltype_Annotation}
\title{Annotate Seurat Object with SlimR Cell Type Predictions}
\usage{
Celltype_Annotation(seurat_obj, cluster_col, SlimR_anno_result)
}
\arguments{
\item{seurat_obj}{A Seurat object containing cluster information in meta.data}

\item{cluster_col}{Character string indicating the column name in meta.data that
contains cluster IDs}

\item{SlimR_anno_result}{List containing a data.frame in $Prediction_results with:
cluster_col (Cluster identifiers (should match cluster_col in meta.data))
Predicted_cell_type (Predicted cell types for each cluster)}
}
\value{
A Seurat object with updated meta.data containing the predicted cell types
}
\description{
This function assigns SlimR predicted cell types to a Seurat object
based on cluster annotations, and stores the results in the meta.data slot.
}
\examples{
\dontrun{
sce <- Celltype_Annotation(seurat_obj = sce,
    cluster_col = "seurat_clusters",
    SlimR_anno_result = SlimR_anno_result
    )
    }

}
\seealso{
Other Celltype_annotation: 
\code{\link{Celltype_Calculate}()}
}
\concept{Celltype_annotation}
