% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celltype_annotation_Bar.R
\name{Celltype_annotation_Bar}
\alias{Celltype_annotation_Bar}
\title{Title}
\usage{
Celltype_annotation_Bar(
  seurat_obj,
  gene_list,
  species,
  cluster_col = "seurat_clusters",
  assay = "RNA",
  save_path = "./Celltype_annotation_Excel/",
  metric_names = NULL
)
}
\arguments{
\item{metric_names}{}
}
\description{
Title
}
