% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celltype_annotation_Cellmarker2.R
\name{Celltype_annotation_Cellmarker2}
\alias{Celltype_annotation_Cellmarker2}
\title{Title}
\usage{
Celltype_annotation_Cellmarker2(
  seurat_obj,
  gene_list,
  species,
  cluster_col = "seurat_clusters",
  assay = "RNA",
  save_path = "./Celltype_annotation_Cellmarker2/",
  min_counts = 1
)
}
\arguments{
\item{min_counts}{}
}
\description{
Title
}
