% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celltype_annotation_Heatmap.R
\name{Celltype_annotation_Heatmap}
\alias{Celltype_annotation_Heatmap}
\title{Title}
\usage{
Celltype_annotation_Heatmap(
  seurat_obj,
  gene_list,
  species,
  cluster_col = "seurat_clusters",
  assay = "RNA"
)
}
\arguments{
\item{assay}{}
}
\description{
Title
}
