% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celltype_annotation_PanglaoDB.R
\name{Celltype_annotation_PanglaoDB}
\alias{Celltype_annotation_PanglaoDB}
\title{Title}
\usage{
Celltype_annotation_PanglaoDB(
  seurat_obj,
  gene_list,
  species,
  cluster_col = "seurat_clusters",
  assay = "RNA",
  save_path = "./Celltype_annotation_PanglaoDB/",
  metric_names = NULL
)
}
\arguments{
\item{metric_names}{}
}
\description{
Title
}
