% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariationData.R
\name{combineClinVarWithHumsavar}
\alias{combineClinVarWithHumsavar}
\title{combineClinVarWithHumsavar}
\usage{
combineClinVarWithHumsavar(vcfFilePath, vepFilePath, nodeN = 8)
}
\arguments{
\item{vcfFilePath}{path to VCF file containing variation data from ClinVar
database}
}
\value{
A data.table object
}
\description{
This function processes humsavar variants (output of getHumSavar()) and
clinvar variants (output of runVEP) and merges into a simplified data.table
object
}

