% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability_parameter.R
\name{compute_adaptive_parameters}
\alias{compute_adaptive_parameters}
\title{Compute Adaptive Parameters Based on Dataset Features (Use in package)}
\usage{
compute_adaptive_parameters(dataset_features, n_celltypes = 50)
}
\arguments{
\item{dataset_features}{List of dataset characteristics from extract_dataset_features()}

\item{n_celltypes}{Expected number of cell types in marker database}
}
\value{
List containing min_expression, specificity_weight, threshold, and rationale
}
\description{
Calculates optimal min_expression, specificity_weight, and threshold parameters
using continuous adaptive algorithms based on dataset characteristics.
}
\seealso{
Other Section_1_Functions_Use_in_Package: 
\code{\link{calculate_cluster_variability}()},
\code{\link{calculate_expression}()},
\code{\link{calculate_expression_skewness}()},
\code{\link{calculate_probability}()},
\code{\link{estimate_batch_effect}()},
\code{\link{extract_dataset_features}()}
}
\concept{Section_1_Functions_Use_in_Package}
