% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{createDB}
\alias{createDB}
\title{createDB}
\usage{
createDB(iupredPath = "/Users/buyar/Desktop/work/tools/iupred", uni,
  workingDirectory = getwd(), overwrite = FALSE)
}
\arguments{
\item{workingDirectory}{The name of the folder to write downloaded/processed
files and RDS objects}

\item{overwrite}{TRUE/FALSE (default: FALSE). Boolean to decide if downloaded
files should overwrite existing files}

\item{uniprotAccessions}{A vector of uniprot accessions e.g. c('P04637',
'P06300')}
}
\value{
A list object with variable data types
  directory
}
\description{
Given a uniprot accession, extract and process all available sequence
features including:
1. Fasta sequences
2. Disorder predictions using IUPred
3. Uniprot feature files downloaded in gff format
4. Uniprot variants including disease-causing and polymorphic substitutions
5. Short linear motifs - all substrings matching all available patterns
6. SLiM Changes - The collection of changes in the proteins' SLiM content
when variants at 4) are applied to the sequence
}

