% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{createDB}
\alias{createDB}
\title{createDB}
\usage{
createDB(uniprotAccessions, iupredPath = "/home/buyar/tools/iupred",
  motifRegex = slimR::motifRegex, workingDirectory = getwd(), nodeN = 1,
  overwrite = FALSE, updateFasta = TRUE, updateGff = TRUE,
  updateIupred = TRUE, updateSlims = TRUE, updateSlimChanges = TRUE)
}
\arguments{
\item{uniprotAccessions}{Vector of uniprot accessions (e.g. c('P04637',
'P11166'))}

\item{workingDirectory}{Path to folder where the database should be created}

\item{nodeN}{Number of cores to run the database generation.}

\item{overwrite}{TRUE/FALSE (default: FALSE) Boolean value to decide if the
previously downloaded or processed files be overwritten? (Applies to all
files except objects saved as RDS)}

\item{updateFasta}{TRUE/FALSE (default: TRUE) Boolean to decide if fasta
download module should be run to update fasta.RDS}

\item{updateGff}{TRUE/FALSE (default: TRUE) Boolean to decide if
gff download module should be run to update gff.RDS}

\item{updateIupred}{TRUE/FALSE (default: TRUE) Boolean to decide if
disorder prediction should be run to update iupred.RDS}

\item{updateSlims}{TRUE/FALSE (default: TRUE) Boolean to decide if
motif prediction should be run to update slims.RDS}

\item{updateSlimChanges}{TRUE/FALSE (default: TRUE) Boolean to
decide if motif changes should be calculated to update
slimChanges.RDS files for variants}
}
\description{
Given a vectore of uniProt accessions, extract and process all available
sequence features and save the resulting objects in folders and RDS objects
for: 1. Fasta sequences 2. Disorder predictions using IUPred 3. Uniprot
feature files downloaded in gff format 4. Uniprot variants including
disease-causing and polymorphic substitutions 5. Short linear motifs - all
substrings matching all available patterns 6. SLiM Changes due to disease
causing mutations - The collection of changes in the proteins' SLiM content
when diesase-causing variants at 4) are applied to the sequence 7. SLiM
Changes due to polymorphisms - The collection of changes in the proteins'
SLiM content when polymorphisms at 4) are applied to the sequence
}

