% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{deregisterScriptBlock}
\alias{deregisterScriptBlock}
\alias{SLiMSim$deregisterScriptBlock}
\alias{.SS$deregisterScriptBlock}
\title{SLiM method deregisterScriptBlock}
\usage{
deregisterScriptBlock(scriptBlocks)
}
\arguments{
\item{scriptBlocks}{An object of type integer or SLiMEidosBlock object. See
details for description.}
}
\value{
An object of type void.
}
\description{
Documentation for SLiM function \code{deregisterScriptBlock}, which is a method
of the SLiM class \code{SLiMSim}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
All SLiMEidosBlock objects specified by scriptBlocks (either with
SLiMEidosBlock objects or with integer identifiers) will be scheduled for
deregistration. The deregistered blocks remain valid, and may even still be
executed in the current stage of the current generation (see section 24.9); the
blocks are not actually deregistered and deallocated until sometime after the
currently executing script block has completed. To immediately prevent a script
block from executing, even when it is scheduled to execute in the current stage
of the current generation, use the active property of the script block (see
sections 23.10.1 and 24.9).
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016–2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
