% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Celltype_Calculate_PerCell.R
\name{.compute_weighted_scores}
\alias{.compute_weighted_scores}
\title{Compute weighted scores for per-cell annotation}
\usage{
.compute_weighted_scores(expr_matrix, marker_sets, min_expression)
}
\description{
Compute weighted scores for per-cell annotation
}
\keyword{internal}
