% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{drawByStrength}
\alias{drawByStrength}
\alias{InteractionType$drawByStrength}
\alias{.IT$drawByStrength}
\title{SLiM method drawByStrength}
\usage{
drawByStrength(individual, count)
}
\arguments{
\item{individual}{An object of type Individual object. Must be of length 1 (a
singleton). See details for description.}

\item{count}{An object of type integer. Must be of length 1 (a singleton). The
default value is \code{1}. See details for description.}
}
\value{
An object of type Individual object.
}
\description{
Documentation for SLiM function \code{drawByStrength}, which is a method of the
SLiM class \code{InteractionType}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Returns up to count individuals drawn from the subpopulation of
individual. The probability of drawing particular individuals is proportional
to the strength of interaction they exert upon individual. This method may be
used with either spatial or non-spatial interactions, but will be more efficient
with spatial interactions that set a short maximum interaction distance. Draws
are done with replacement, so the same individual may be drawn more than once;
sometimes using unique() on the result of this call is therefore desirable. If
more than one draw will be needed, it is much more efficient to use a single
call to drawByStrength(), rather than drawing individuals one at a time. Note
that if no individuals exert a non-zero interaction upon individual, the vector
returned will be zero-length; it is important to consider this possibility.
If the needed interaction strengths have already been calculated, those cached
values are simply used. Otherwise, calling this method triggers evaluation
of the needed interactions, including calls to any applicable interaction()
callbacks.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016–2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
