% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability_parameter.R
\name{estimate_batch_effect}
\alias{estimate_batch_effect}
\title{Estimate Batch Effect Strength (Use in package)}
\usage{
estimate_batch_effect(seurat_obj, assay)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{assay}{Assay name}
}
\value{
Batch effect score (0 indicates no detectable batch effect)
}
\description{
Roughly estimates the potential impact of batch effects
using available metadata.
}
\seealso{
Other Section_1_Functions_Use_in_Package: 
\code{\link{calculate_cluster_variability}()},
\code{\link{calculate_expression}()},
\code{\link{calculate_expression_skewness}()},
\code{\link{calculate_probability}()},
\code{\link{compute_adaptive_parameters}()},
\code{\link{extract_dataset_features}()}
}
\concept{Section_1_Functions_Use_in_Package}
