% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability_parameter.R
\name{estimate_batch_effect}
\alias{estimate_batch_effect}
\title{Estimate Batch Effect Strength}
\usage{
estimate_batch_effect(object, assay)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{Assay name}
}
\value{
Batch effect score (0 indicates no detectable batch effect)
}
\description{
Roughly estimates the potential impact of batch effects
using available metadata.
}
\keyword{internal}
