% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability_parameter.R
\name{extract_dataset_features}
\alias{extract_dataset_features}
\title{Extract Dataset Characteristics for Machine Learning}
\usage{
extract_dataset_features(object, features, assay = NULL, cluster_col = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{features}{Features to analyze}

\item{assay}{Assay name}

\item{cluster_col}{Cluster column name}
}
\value{
List of dataset characteristics including expression statistics,
variability measures, and cluster properties
}
\description{
Computes various statistical features from single-cell data that are used
as input for the parameter prediction model.
}
\keyword{internal}
