% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifAnalysis.R
\name{findMotifChanges}
\alias{findMotifChanges}
\title{findMotifChanges}
\usage{
findMotifChanges(sequence, variants, motifRegex = slimR::motifRegex)
}
\arguments{
\item{sequence}{A character string of amino acid sequence}

\item{variants}{A data.frame consisting of minimum four columns:
1.uniprotAccession, 2.wtAA, 3.mutAA, 4.pos where pos is the mutation position in
the sequence, wtAA is the wild-type amino acid (one letter code) in the
sequence and mutAA is the mutant amino acid (one letter code).}
}
\value{
data.table data.frame object
}
\description{
Find out which SLiMs are gained or lost (no longer matching the regex
pattern) via point amino acid substitutions in protein sequences
}
\examples{
c <- findMotifChanges(sequence = paste(glutFasta),
                      variants = glutMutations,
                     motifRegex = motifRegex)
}
