% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability_parameter.R
\name{generate_training_data}
\alias{generate_training_data}
\title{Generate Training Data for Machine Learning Model (Use in package)}
\usage{
generate_training_data(dataset_features, n_samples = 1000)
}
\arguments{
\item{dataset_features}{List of actual dataset characteristics}

\item{n_samples}{Number of synthetic samples to generate}
}
\value{
Data frame with synthetic features and optimal parameter targets
}
\description{
Creates synthetic training data based on empirical rules about
optimal parameter relationships with dataset characteristics.
}
\seealso{
Other Section_1_Functions_Use_in_Package: 
\code{\link{calculate_cluster_variability}()},
\code{\link{calculate_expression}()},
\code{\link{calculate_expression_skewness}()},
\code{\link{calculate_probability}()},
\code{\link{estimate_batch_effect}()},
\code{\link{extract_dataset_features}()},
\code{\link{postprocess_parameters}()},
\code{\link{predict_optimal_parameters}()},
\code{\link{train_parameter_model}()}
}
\concept{Section_1_Functions_Use_in_Package}
