% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getClinVarData}
\alias{getClinVarData}
\title{getClinVarData}
\usage{
getClinVarData(url = paste0("ftp://ftp.ncbi.nlm.nih.gov/",
  "pub/clinvar/tab_delimited/", "variant_summary.txt.gz"),
  overwrite = FALSE)
}
\arguments{
\item{url}{The url to the ftp location of the clinvar dataset (tab delimited
variant summary file) (e.g.
ftp://ftp.ncbi.nlm.nih.gov/pub/clinvar/tab_delimited/variant_summary.txt.gz)}

\item{overwrite}{default: FALSE, boolean value to decide if a fresh download
should overwrite the existing file}
}
\value{
Path to the downloaded file
}
\description{
This function will fetch the clinvar data from the given url and parse the
contents of the downloaded file.
}
\examples{
\dontrun{
#download
filePath <- getClinVarData()
#read first 100 lines
cv_head <- data.table::fread(text = readLines(filePath, n = 100))
}
}
