% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariationData.R
\name{getHumSavar}
\alias{getHumSavar}
\title{getHumSavar}
\usage{
getHumSavar()
}
\value{
A Granges object containing the coordinates of mutated sites in proteins
}
\description{
Download and parse protein mutation data from UniProt
}

