% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getUniprotData}
\alias{getUniprotData}
\title{getUniprotData}
\usage{
getUniprotData(outDir, format, reviewed = TRUE, organism = 9606,
  update = FALSE)
}
\arguments{
\item{outDir}{Target download folder path}

\item{format}{File format. Accepted formats are gff, fasta, txt}

\item{reviewed}{Only download reviewed uniprot entries (default: TRUE)}

\item{organism}{Organism code e.g. 9606 for human}

\item{update}{Boolean (default: FALSE), whether to download updated
annotations to the target folder}
}
\value{
Path to the downloaded file
}
\description{
Batch download different types of proteome-wide annotation data from Uniprot
using REST services
}
\examples{
#download Tobacco Rattle Virus proteome (it has only 6 genes)
#https://www.uniprot.org/proteomes/UP000001669
up <- getUniprotData(outDir = getwd(), format = 'fasta',
                    reviewed = TRUE, organism = 652939)
up_fasta <- Biostrings::readAAStringSet(up)
\dontrun{
#download gff annotation for human reviewed proteome
gffFile <- getUniprotData(outDir = getwd(), format = 'fasta',
                    reviewed = TRUE, organism = 9606)
gff <- rtracklayer::import.gff(gffFile)
}
}
