% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glutIUPred}
\alias{glutIUPred}
\title{GLUT1 disorder score predictions by IUPred}
\format{An object of class \code{data.frame} with 492 rows and 3 columns.}
\usage{
glutIUPred
}
\value{
A data.frame object
}
\description{
GLUT1 disorder score predictions by IUPred
}
\keyword{datasets}
