% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_scripting.R
\name{\%.\%}
\alias{\%.\%}
\title{slimrlang stub for the SLiM '.' operator}
\usage{
lhs \%.\% rhs
}
\arguments{
\item{lhs}{Object of class \code{rhs}, to extract methods or properties from}

\item{rhs}{SLiM class R object (such as \code{Subpopulation}, \code{.M}, etc.). Type \code{\link{slim_classes}} for a table of possible values.}
}
\description{
Use this in place of '.' from SLiM to specify a method or a property coming from a
particular SLiM class. Note that the R function is a stub, it does not do anything in R
(except bring up this documentation). It will only do anything useful when used inside
a \\code{\\link{slim_block}} function further nested in a \\code{\\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a full SLiM script.
}
\section{Copyright}{

 This is documentation for a function in the SLiM software, and has been reproduced from the official manual,
 which can be found here: \\url{http://benhaller.com/slim/SLiM_Manual.pdf}. This documentation is
 Copyright © 2016–2020 Philipp Messer. All rights reserved. More information about SLiM can be found
 on the official website: \\url{https://messerlab.org/slim/}
}

\author{
Benjamin C Haller (\\email{bhaller@benhaller.com}) and Philipp W Messer (\\email{messer@cornell.edu})
}
