% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\docType{data}
\name{Initialize}
\alias{Initialize}
\alias{.Init}
\title{Initialize}
\format{
An object of class \code{list} of length 14.
}
\usage{
Initialize
}
\description{
Documentation for Initialize class from SLiM
}
\details{
This class has the following methods (functions):
\itemize{
\item{\code{\link{initializeAncestralNucleotides}}}
\item{\code{\link{initializeGeneConversion}}}
\item{\code{\link{initializeGenomicElement}}}
\item{\code{\link{initializeGenomicElementType}}}
\item{\code{\link{initializeHotspotMap}}}
\item{\code{\link{initializeInteractionType}}}
\item{\code{\link{initializeMutationRate}}}
\item{\code{\link{initializeMutationType}}}
\item{\code{\link{initializeMutationTypeNuc}}}
\item{\code{\link{initializeRecombinationRate}}}
\item{\code{\link{initializeSex}}}
\item{\code{\link{initializeSLiMModelType}}}
\item{\code{\link{initializeSLiMOptions}}}
\item{\code{\link{initializeTreeSeq}}}
}
This class has the following properties:
\describe{
\item{None.}{This class has no properties.}
}
}
\keyword{datasets}
