% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\docType{data}
\name{InteractionType}
\alias{InteractionType}
\alias{.IT}
\title{InteractionType}
\format{
An object of class \code{list} of length 21.
}
\usage{
InteractionType
}
\description{
Documentation for InteractionType class from SLiM
}
\details{
This class has the following methods (functions):
\itemize{
\item{\code{\link{distance}}}
\item{\code{\link{distanceToPoint}}}
\item{\code{\link{drawByStrength}}}
\item{\code{\link{evaluate}}}
\item{\code{\link{getValue}}}
\item{\code{\link{interactingNeighborCount}}}
\item{\code{\link{interactionDistance}}}
\item{\code{\link{nearestInteractingNeighbors}}}
\item{\code{\link{nearestNeighbors}}}
\item{\code{\link{nearestNeighborsOfPoint}}}
\item{\code{\link{setInteractionFunction}}}
\item{\code{\link{setValue}}}
\item{\code{\link{strength}}}
\item{\code{\link{totalOfNeighborStrengths}}}
\item{\code{\link{unevaluate}}}
}
This class has the following properties:
\describe{
\item{id}{A property of type integer or float or logical or string or string
or integer. It is of length one (a singleton). This property is a constant, so
it is not modifiable. \strong{Property Description:} The identifier for this
interaction type; for interaction type i3, for example, this is 3. }
\item{maxDistance}{A property of type integer or float or logical or string
or string or integer. It is of length one (a singleton). This property is
a variable, so it is modifiable. \strong{Property Description:} The maximum
distance over which this interaction will be evaluated. For inter-individual
distances greater than maxDistance, the interaction strength will be zero. }
\item{reciprocal}{A property of type integer or float or logical or string
or string or integer. It is of length one (a singleton). This property is
a constant, so it is not modifiable. \strong{Property Description:} The
reciprocality of the interaction, as specified in initializeInteractionType().
This will be T for reciprocal interactions (those for which the interaction
strength of B upon A is equal to the interaction strength of A upon B), and F
otherwise. }
\item{sexSegregation}{A property of type integer or float or logical or string
or string or integer. It is of length one (a singleton). This property is a
constant, so it is not modifiable. \strong{Property Description:} The sex-
segregation of the interaction, as specified in initializeInteractionType(). For
nonsexual simulations, this will be "**". For sexual simulations, this string
value indicates the sex of individuals feeling the interaction, and the sex
of individuals exerting the interaction; see initializeInteractionType() for
details. }
\item{spatiality}{A property of type integer or float or logical or string
or string or integer. It is of length one (a singleton). This property is a
constant, so it is not modifiable. \strong{Property Description:} The spatial
dimensions used by the interaction, as specified in initializeInteractionType().
This will be "" (the empty string) for non-spatial interactions, or "x", "y",
"z", "xy", "xz", "yz", or "xyz", for interactions using those spatial dimensions
respectively. The specified dimensions are used to calculate the distances
between individuals for this interaction. The value of this property is always
the same as the value given to initializeInteractionType(). }
\item{tag}{A property of type integer or float or logical or string or string
or integer. It is of length one (a singleton). This property is a variable, so
it is modifiable. \strong{Property Description:} A user-defined integer value.
The value of tag is initially undefined, and it is an error to try to read
it; if you wish it to have a defined value, you must arrange that yourself by
explicitly setting its value prior to using it elsewhere in your code. The value
of tag is not used by SLiM; it is free for you to use. See also the getValue()
and setValue() methods, for another way of attaching state to interaction
types.}

}
}
\keyword{datasets}
