% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Markers_filter_Cellmarker2.R
\name{Markers_filter_Cellmarker2}
\alias{Markers_filter_Cellmarker2}
\title{Title}
\usage{
Markers_filter_Cellmarker2(
  df,
  species = NULL,
  tissue_class = NULL,
  tissue_type = NULL,
  cancer_type = NULL,
  cell_type = NULL,
  cell_name = NULL,
  marker = NULL,
  counts = NULL
)
}
\arguments{
\item{counts}{}
}
\description{
Title
}
