% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Markers_filter_ScType.R
\name{Markers_filter_ScType}
\alias{Markers_filter_ScType}
\title{Create Marker_list from the ScType database}
\usage{
Markers_filter_ScType(df, tissue_type = NULL, cell_name = NULL)
}
\arguments{
\item{df}{Standardized ScType database. It is read as \code{data(ScType)}
in the SlimR library.}

\item{tissue_type}{Tissue type information in the ScType database. The input
can be retrieved by \code{"ScType_table"}. For more information, please
refer to \url{https://github.com/IanevskiAleksandr/sc-type}.}

\item{cell_name}{Cell type name information in the ScType database. The input
can be retrieved by \code{"ScType_table"}. For more information, please
refer to \url{https://github.com/IanevskiAleksandr/sc-type}.}
}
\value{
The standardized "Marker_list" in the SlimR package
}
\description{
Create Marker_list from the ScType database
}
\examples{
ScType <- SlimR::ScType
Markers_list_ScType <- Markers_filter_ScType(
    ScType,
    tissue_type = "Immune system",
    cell_name = NULL
    )

}
\seealso{
Other Section_2_Standardized_Markers_List: 
\code{\link{Markers_filter_Cellmarker2}()},
\code{\link{Markers_filter_PanglaoDB}()},
\code{\link{Read_excel_markers}()},
\code{\link{Read_seurat_markers}()}
}
\concept{Section_2_Standardized_Markers_List}
