% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{nearestInteractingNeighbors}
\alias{nearestInteractingNeighbors}
\alias{InteractionType$nearestInteractingNeighbors}
\alias{.IT$nearestInteractingNeighbors}
\title{SLiM method nearestInteractingNeighbors}
\usage{
nearestInteractingNeighbors(individual, count)
}
\arguments{
\item{individual}{An object of type Individual object. Must be of length 1 (a
singleton). See details for description.}

\item{count}{An object of type integer. Must be of length 1 (a singleton). The
default value is \code{1}. See details for description.}
}
\value{
An object of type Individual object.
}
\description{
Documentation for SLiM function \code{nearestInteractingNeighbors}, which is a
method of the SLiM class \code{InteractionType}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Returns up to count interacting individuals that are spatially
closest to individual, according to the distance metric of the InteractionType.
More specifically, this method returns only individuals which can exert an
interaction upon the focal individual; it does not include individuals that only
feel an interaction from the focal individual. To obtain all of the interacting
individuals within the maximum interaction distance of individual, simply pass
a value for count that is greater than or equal to the size of individual’s
subpopulation. Note that if fewer than count interacting individuals are within
the maximum interaction distance, the vector returned may be shorter than count,
or even zero-length; it is important to check for this possibility even when
requesting a single neighbor. If only the number of interacting individuals
is needed, use interactingNeighborCount() instead. Note that this method uses
interaction eligibility as a criterion; it will not return neighbors that cannot
exert an interaction upon the focal individual (due to sex-segregation, e.g.).
(It will also never return the focal individual as a neighbor of itself.) To
find all neighbors of the focal individual, whether they can interact with it or
not, use nearestNeighbors().
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016–2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
