% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database_PanglaoDB.R
\docType{data}
\name{PanglaoDB}
\alias{PanglaoDB}
\title{PanglaoDB dataset}
\format{
A data frame with 9 columns:
}
\source{
\url{https://panglaodb.se/}
}
\usage{
PanglaoDB
}
\description{
A dataset containing marker genes for different cell types from PanglaoDB
}
\details{
This dataset is used to filter and create a standardized marker list.'
}
\seealso{
Other SlimR_Database: 
\code{\link{Cellmarker2}},
\code{\link{Cellmarker2_raw}},
\code{\link{Cellmarker2_table}},
\code{\link{Markers_list_TCellSI}},
\code{\link{Markers_list_scIBD}},
\code{\link{PanglaoDB_raw}},
\code{\link{PanglaoDB_table}}
}
\concept{SlimR_Database}
\keyword{datasets}
