% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanglaoDB.R
\docType{data}
\name{PanglaoDB}
\alias{PanglaoDB}
\title{PanglaoDB dataset}
\format{
A data frame with 9 columns:
}
\source{
\url{https://panglaodb.se/}
}
\usage{
PanglaoDB
}
\description{
A dataset containing marker genes for different cell types from PanglaoDB
}
\details{
This dataset is used to filter and create a standardized marker list.'
}
\keyword{datasets}
