% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanglaoDB_raw.R
\docType{data}
\name{PanglaoDB_raw}
\alias{PanglaoDB_raw}
\title{PanglaoDB raw dataset}
\format{
A data frame with 14 columns contined in the PanglaoDB database:
}
\source{
\url{https://panglaodb.se/}
}
\usage{
PanglaoDB_raw
}
\description{
A dataset containing marker genes for different cell types from PanglaoDB
}
\details{
This dataset is used to filter and create a standardized marker list.'
}
\keyword{datasets}
