% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database_PanglaoDB_raw.R
\docType{data}
\name{PanglaoDB_raw}
\alias{PanglaoDB_raw}
\title{PanglaoDB raw dataset}
\format{
A data frame with 14 columns contined in the PanglaoDB database:
}
\source{
\url{https://panglaodb.se/}
}
\usage{
PanglaoDB_raw
}
\description{
A dataset containing marker genes for different cell types from PanglaoDB
}
\details{
This dataset is used to filter and create a standardized marker list.'
}
\seealso{
Other Section_0_Database:
\code{\link{Cellmarker2}},
\code{\link{Cellmarker2_raw}},
\code{\link{Cellmarker2_table}},
\code{\link{Markers_list_PCTAM}},
\code{\link{Markers_list_PCTIT}},
\code{\link{Markers_list_TCellSI}},
\code{\link{Markers_list_scIBD}},
\code{\link{PanglaoDB}},
\code{\link{PanglaoDB_table}},
\code{\link{ScType}},
\code{\link{ScType_raw}},
\code{\link{ScType_table}}
}
\concept{Section_0_Database}
\keyword{datasets}
