% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanglaoDB_table.R
\docType{data}
\name{PanglaoDB_table}
\alias{PanglaoDB_table}
\title{PanglaoDB table}
\format{
A list contain different types like species, organ, cell type.
}
\source{
\url{https://panglaodb.se/}
}
\usage{
PanglaoDB_table
}
\description{
A dataset containing marker genes for different cell types from PanglaoDB
}
\details{
This list is used to choose filters for creation of standardized marker list.
}
\keyword{datasets}
