% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database_PanglaoDB_table.R
\docType{data}
\name{PanglaoDB_table}
\alias{PanglaoDB_table}
\title{PanglaoDB table}
\format{
A list contain different types like species, organ, cell type.
}
\source{
\url{https://panglaodb.se/}
}
\usage{
PanglaoDB_table
}
\description{
A dataset containing marker genes for different cell types from PanglaoDB
}
\details{
This list is used to choose filters for creation of standardized marker list.
}
\seealso{
Other Section_0_Database: 
\code{\link{Cellmarker2}},
\code{\link{Cellmarker2_raw}},
\code{\link{Cellmarker2_table}},
\code{\link{Markers_list_PCTAM}},
\code{\link{Markers_list_PCTIT}},
\code{\link{Markers_list_TCellSI}},
\code{\link{Markers_list_scIBD}},
\code{\link{PanglaoDB}},
\code{\link{PanglaoDB_raw}},
\code{\link{ScType}},
\code{\link{ScType_raw}},
\code{\link{ScType_table}}
}
\concept{Section_0_Database}
\keyword{datasets}
