% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expression_plot.R
\name{plot_mean_expression}
\alias{plot_mean_expression}
\title{Counts average expression of gene set and plots Boxplot (Use in package)}
\usage{
plot_mean_expression(object, features, assay = NULL, cluster_col = NULL)
}
\arguments{
\item{object}{Enter a Seurat object.}

\item{features}{Enter one or a set of markers.}

\item{assay}{Enter the assay used by the Seurat object, such as "RNA". Default
parameters use "assay = NULL".}

\item{cluster_col}{Enter the meta.data column in the Seurat object to be
annotated, such as "seurat_cluster". Default parameters use "cluster_col = NULL".}
}
\value{
Average expression box plot of genes in the input "Seurat" object
given "cluster_col" and given "features".
}
\description{
Counts average expression of gene set and plots Boxplot (Use in package)
}
