% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{pointPeriodic}
\alias{pointPeriodic}
\alias{Subpopulation$pointPeriodic}
\alias{.P$pointPeriodic}
\title{SLiM method pointPeriodic}
\usage{
pointPeriodic(point)
}
\arguments{
\item{point}{An object of type float. See details for description.}
}
\value{
An object of type float.
}
\description{
Documentation for SLiM function \code{pointPeriodic}, which is a method of the
SLiM class \code{Subpopulation}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Returns a revised version of point that has been brought inside
the periodic spatial boundaries of the subpopulation (as specified by the
periodicity parameter of initializeSLiMOptions()) by wrapping around periodic
spatial boundaries. In brief, if a coordinate of point lies beyond a periodic
spatial boundary, that coordinate is wrapped around the boundary, so that it
lies inside the spatial extent by the same magnitude that it previously lay
outside, but on the opposite side of the space; in effect, the two edges of
the periodic spatial boundary are seamlessly joined. This is done iteratively
until all coordinates lie inside the subpopulation’s periodic boundaries. Note
that non-periodic spatial boundaries are not enforced by this method; they
should be enforced using pointReflected(), pointStopped(), or some other means
of enforcing boundary constraints (which can be used after pointPeriodic() to
bring the remaining coordinates into bounds; coordinates already brought into
bounds by pointPeriodic() will be unaffected by those calls). This method is
useful for implementing periodic boundary conditions. This may only be called
in simulations for which continuous space and at least one periodic spatial
dimension have been enabled with initializeSLiMOptions(). The length of point
must be an exact multiple of the dimensionality of the simulation; in other
words, point may contain values comprising more than one point. In this case,
each point will be processed as described above and a new vector containing all
of the processed points will be returned.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016–2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
