% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability_parameter.R
\name{postprocess_parameters}
\alias{postprocess_parameters}
\title{Post-process Predicted Parameters (Use in package)}
\usage{
postprocess_parameters(predicted_params, dataset_features)
}
\arguments{
\item{predicted_params}{List of raw predicted parameters}

\item{dataset_features}{Characteristics of current dataset}
}
\value{
List of finalized parameters after post-processing
}
\description{
Applies constraints and dataset-specific adjustments to ensure
predicted parameters are within reasonable ranges.
}
\seealso{
Other Section_1_Functions_Use_in_Package: 
\code{\link{calculate_cluster_variability}()},
\code{\link{calculate_expression}()},
\code{\link{calculate_expression_skewness}()},
\code{\link{calculate_probability}()},
\code{\link{estimate_batch_effect}()},
\code{\link{extract_dataset_features}()},
\code{\link{generate_training_data}()},
\code{\link{predict_optimal_parameters}()},
\code{\link{train_parameter_model}()}
}
\concept{Section_1_Functions_Use_in_Package}
