% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability_parameter.R
\name{predict_optimal_parameters}
\alias{predict_optimal_parameters}
\title{Predict Optimal Parameters Using Trained Model}
\usage{
predict_optimal_parameters(model, dataset_features)
}
\arguments{
\item{model}{Trained machine learning model (now a list with two models)}

\item{dataset_features}{Extracted characteristics of current dataset}
}
\value{
List containing predicted min_expression and specificity_weight
}
\description{
Applies the trained machine learning model to predict optimal
parameters for the current dataset.
}
\seealso{
Other Section_1_Functions_Use_in_Package: 
\code{\link{calculate_cluster_variability}()},
\code{\link{calculate_expression}()},
\code{\link{calculate_expression_skewness}()},
\code{\link{calculate_probability}()},
\code{\link{estimate_batch_effect}()},
\code{\link{extract_dataset_features}()},
\code{\link{generate_training_data}()},
\code{\link{postprocess_parameters}()},
\code{\link{train_parameter_model}()}
}
\concept{Section_1_Functions_Use_in_Package}
