% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_output.R
\name{r_output_snp}
\alias{r_output_snp}
\alias{slimr_output_snp}
\title{Utility function to tell SLiM to output SNP format data}
\usage{
r_output_snp(name = "snp", subpops = FALSE, ...)

slimr_output_snp(name = "snp", subpops = FALSE, ...)
}
\arguments{
\item{name}{Name of output to use to label it in \code{slimr_results object}. Default is \code{"snp"}.}

\item{subpops}{Should the subpopulation of each sequence be outputted as well?}

\item{...}{Other arguments to be passed to \code{\link{r_output}}}
}
\value{
None
}
\description{
Use \code{\link{slim_results_to_data}} on the \code{\link{slim_run}} results to get this in a nice SNP matrix form.
}
\examples{
if(slim_is_avail()) {
  test_sim <- slim_script(
    slim_block_init_minimal(mutation_rate = 1e-6),
    slim_block_add_subpops(1, 100),
    slim_block(1, 20, late(), {
      r_output_snp("snp", do_every = 10)
    })
  ) \%>\%
    slim_run()
  slim_results_to_data(test_sim)
}
}
