% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_excel_markers.R
\name{read_excel_markers}
\alias{read_excel_markers}
\title{Create "Marker_list" from Excel files ".xlsx"}
\usage{
read_excel_markers(path)
}
\arguments{
\item{path}{The path information of Marker files stored in ".xlsx" format.
The Sheet name in the file is filled with cell type. The first line of
each Sheet is the table head, the first column is filled with markers
information, and the following column is filled with mertic information.}
}
\value{
The standardized "Marker_list" in the SlimR package.
}
\description{
Create "Marker_list" from Excel files ".xlsx"
}
\examples{
\dontrun{Markers_list_Excel <- read_excel_markers(
         "D:/Laboratory/Marker_load.xlsx"
         )
         }

}
