% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_output.R
\name{read_individuals}
\alias{read_individuals}
\title{Parse the individual section of the SLiM file into a data frame.}
\usage{
read_individuals(slim_file, p = "")
}
\description{
Parse the individual section of the SLiM file into a data frame.
}
