% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_output.R
\name{read_mutations}
\alias{read_mutations}
\title{Parse the mutation section of the SLiM file into a data frame.}
\usage{
read_mutations(slim_file, mut_type = "", pop = "", time = 0)
}
\arguments{
\item{slim_file}{Lines of the whole SLiM output file.}

\item{mut_type, pop, time}{Mutation type, population of origin
and generation of origin of mutations to filter from the file.}
}
\description{
Parse the mutation section of the SLiM file into a data frame.
}
