% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_seurat_markers.R
\name{read_seurat_markers}
\alias{read_seurat_markers}
\title{Title}
\usage{
read_seurat_markers(df, sort_by = "avg_log2FC", gene_fliter = 20)
}
\arguments{
\item{gene_fliter}{}
}
\description{
Title
}
