% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_seurat_markers.R
\name{read_seurat_markers}
\alias{read_seurat_markers}
\title{Create "Marker_list" from Seurat object}
\usage{
read_seurat_markers(
  df,
  sources = c("Seurat", "presto"),
  sort_by = NULL,
  gene_filter = 20
)
}
\arguments{
\item{df}{Dataframe generated by "FindAllMarkers" function, recommend
to use parameter "group.by = "Cell_type"" and "only.pos = TRUE".}

\item{sources}{Type of markers sources to use. Be one of: \code{"Seurat"} or \code{"presto"}.}

\item{sort_by}{Marker sorting parameter, select "avg_log2FC" or "p_val_adj".
Default parameters use "sort_by = 'avg_log2FC'".}

\item{gene_filter}{The number of markers left for each cell type based
on the "sort_by" parameter's level of difference. Default parameters use
"gene_fliter = 20"}
}
\value{
The standardized "Marker_list" in the SlimR package.
}
\description{
Create "Marker_list" from Seurat object
}
\examples{
\dontrun{
# Example for Seurat sources markers
library(Seurat)
seurat_markers <- FindAllMarkers(
    seurat_obj = sce,
    group.by = "Cell_type",
    only.pos = TRUE)

Markers_list_Seurat <- read_seurat_markers(seurat_markers,
    sources = "Seurat",
    sort_by = "avg_log2FC",
    gene_filter = 10
    )

# Example for presto sources markers
library(presto)
library(dplyr)
seurat_markers <- wilcoxauc(
    seurat_obj = sce,
    group_by = "Cell_type",
    seurat_assay = "RNA"
    ) \%>\%
    filter(padj < 0.05, logFC > 0.5)

Markers_list_Seurat <- read_seurat_markers(seurat_markers,
    sources = "presto",
    sort_by = "logFC",
    gene_filter = 10
    )
}


}
\seealso{
Other Markers_list_generation: 
\code{\link{Markers_filter_Cellmarker2}()},
\code{\link{Markers_filter_PanglaoDB}()},
\code{\link{read_excel_markers}()}
}
\concept{Markers_list_generation}
