% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_callbacks.R
\name{reproduction}
\alias{reproduction}
\title{SLiM reproduction() callback}
\usage{
reproduction(subpop_id, sex)
}
\arguments{
\item{subpop_id}{The id(s) of the subpopulation(s) to which this callback should apply. Can
be an integer 1, 2, etc., or character "p1", "p2", etc.}

\item{sex}{The sex of the individuals to apply this callback to.}
}
\value{
None
}
\description{
This callback specifies that a code block is providing logic for the production
of offspring. It is called once per individual for each generation in which it is active,
and is expected to add new individuals to the simulation using the SLiM functions provided
for that purpose. This is only used in "nonWF" simulations.
For more information on how to use \code{reproduction()} callback
see \href{http://benhaller.com/slim/SLiM_Manual.pdf#page=606}{SLiM Manual: page 606}
}
\details{
Global variables available in reproduction callbacks:
\describe{
  \item{individual}{The focal individual that is expected to reproduce}
  \item{genome1}{One genome of the focal individual}
  \item{genome2}{The other genome of the focal individual}
  \item{subpop}{The subpopulation to which the focal individual belongs}
}
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
modified from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016-2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\examples{
slim_block(reproduction(), {

  subpop.addCrossed(individual, subpop.sampleIndividuals(1))

})
}
\seealso{
Other callbacks: 
\code{\link{early}()},
\code{\link{first}()},
\code{\link{fitness}()},
\code{\link{initialize}()},
\code{\link{interaction}()},
\code{\link{late}()},
\code{\link{mateChoice}()},
\code{\link{modifyChild}()},
\code{\link{recombination}()},
\code{\link{slim_callbacks}()}
}
\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
\concept{callbacks}
