% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database_ScType.R
\docType{data}
\name{ScType_raw}
\alias{ScType_raw}
\title{ScType raw dataset}
\format{
A tibble with 5 columns:
\describe{
\item{tissueType}{Tissue type}
\item{cellName}{Full cell type name}
\item{geneSymbolmore1}{Comma-separated positive marker genes}
\item{geneSymbolmore2}{Comma-separated negative marker genes (not used in processing)}
\item{shortName}{Abbreviated cell type name}
}
}
\source{
\url{https://github.com/IanevskiAleksandr/sc-type}
}
\usage{
ScType_raw
}
\description{
The original ScType marker database before processing.
}
\seealso{
Other Section_0_Database: 
\code{\link{Cellmarker2}},
\code{\link{Cellmarker2_raw}},
\code{\link{Cellmarker2_table}},
\code{\link{Markers_list_PCTAM}},
\code{\link{Markers_list_PCTIT}},
\code{\link{Markers_list_TCellSI}},
\code{\link{Markers_list_scIBD}},
\code{\link{PanglaoDB}},
\code{\link{PanglaoDB_raw}},
\code{\link{PanglaoDB_table}},
\code{\link{ScType}},
\code{\link{ScType_table}}
}
\concept{Section_0_Database}
\keyword{datasets}
