% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifAnalysis.R
\name{searchSLiMs}
\alias{searchSLiMs}
\title{searchMotifs}
\usage{
searchSLiMs(sequence, motifRegex)
}
\arguments{
\item{sequence}{Protein sequence using amino acid alphabet}

\item{motifRegex}{A list object where names of the list are ELM identifiers
and each list item has one ELM regex}
}
\description{
Using regular expressions of SLiMs, find all matches of given SLiM patterns
in a given sequence
}
\examples{
download.file(url = 'www.uniprot.org/uniprot/P11166.fasta',
         destfile = file.path(getwd(), 'sample.fasta'))
sequence <- Biostrings::readAAStringSet(filepath = file.path(getwd(), 'sample.fasta'))
elms <- getElmClasses()
motifRegex <- as.list(as.vector(elms$RegEx))
names(motifRegex) <- as.vector(elms$ELM_Identifier)

searchSLiMs(sequence, motifRegex)

}

