% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifAnalysis.R
\name{searchSLiMs}
\alias{searchSLiMs}
\title{searchMotifs}
\usage{
searchSLiMs(sequence, motifRegex)
}
\arguments{
\item{sequence}{Protein sequence using amino acid alphabet}

\item{motifRegex}{A list object where names of the list are ELM identifiers
and each list item has one ELM regex}
}
\value{
A list of data.frame objects.
}
\description{
Using regular expressions of SLiMs, find all matches of given SLiM patterns
in a given sequence
}
\examples{
data("glutFasta")
motifRegex <- list('TRG_ENDOCYTIC_2' = 'Y..[LMVIF]')
searchSLiMs(paste(glutFasta), motifRegex)
}

