% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{setSelfingRate}
\alias{setSelfingRate}
\alias{Subpopulation$setSelfingRate}
\alias{.P$setSelfingRate}
\title{SLiM method setSelfingRate}
\usage{
setSelfingRate(rate)
}
\arguments{
\item{rate}{An object of type numeric. Must be of length 1 (a singleton). See
details for description.}
}
\value{
An object of type void.
}
\description{
Documentation for SLiM function \code{setSelfingRate}, which is a method of the
SLiM class \code{Subpopulation}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Set the selfing rate of this subpopulation. The rate is changed to
rate, which should be between 0.0 and 1.0, inclusive. Selfing can only be
enabled in non-sexual (i.e. hermaphroditic) simulations. During mating and
offspring generation, the probability that any given offspring individual will
be generated by selfing – by self-fertilization via gametes produced by meiosis
by a single parent – will be equal to the selfing rate set in the parental (not
the offspring!) subpopulation (see section 21.2.1).
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016–2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
