% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{simulationFinished}
\alias{simulationFinished}
\alias{Community$simulationFinished}
\alias{.Co$simulationFinished}
\alias{Species$simulationFinished}
\alias{.Sp$simulationFinished}
\title{SLiM method simulationFinished}
\usage{
simulationFinished(void)

simulationFinished(void)
}
\arguments{
\item{void}{An object of type . See details for description.}
}
\value{
An object of type void.

An object of type void.
}
\description{
Documentation for SLiM function \code{simulationFinished}, which is a method of
the SLiM class \code{\link{Community}}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.

Documentation for SLiM function \code{simulationFinished}, which is a method of
the SLiM class \code{\link{Species}}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Documentation for this function can be found in the official
\href{http://benhaller.com/slim/SLiM_Manual.pdf#page=668}{SLiM manual: page
668}.

Declare the current simulation finished. Normally SLiM ends a
simulation when, at the end of a tick, there are no script events or callbacks
registered for any future tick (excluding scripts with no declared end tick).
If you wish to end a simulation before this condition is met, a call to
simulationFinished() will cause the current simulation to end at the end of
the current tick. For example, a simulation might self-terminate if a test for
a dynamic equilibrium condition is satisfied. Note that the current tick will
finish executing; if you want the simulation to stop immediately, you can use
the Eidos method stop(), which raises an error condition.

Documentation for this function can be found in the official
\href{http://benhaller.com/slim/SLiM_Manual.pdf#page=727}{SLiM manual: page
727}.

Declare the current simulation finished. This method is equivalent to
the Community method simulationFinished(), except that this method is only legal
to call in single-species models (to provide backward compatibility). It is
recommended that new code should call the Community method; this method may be
deprecated in the future.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016-2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}


This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016-2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\seealso{
Other Community: 
\code{\link{Co}},
\code{\link{createLogFile}()},
\code{\link{deregisterScriptBlock}()},
\code{\link{genomicElementTypesWithIDs}()},
\code{\link{interactionTypesWithIDs}()},
\code{\link{mutationTypesWithIDs}()},
\code{\link{outputUsage}()},
\code{\link{registerEarlyEvent}()},
\code{\link{registerFirstEvent}()},
\code{\link{registerInteractionCallback}()},
\code{\link{registerLateEvent}()},
\code{\link{rescheduleScriptBlock}()},
\code{\link{scriptBlocksWithIDs}()},
\code{\link{speciesWithIDs}()},
\code{\link{subpopulationsWithIDs}()},
\code{\link{usage}()}

Other Species: 
\code{\link{Sp}},
\code{\link{addSubpop}()},
\code{\link{addSubpopSplit}()},
\code{\link{countOfMutationsOfType}()},
\code{\link{individualsWithPedigreeIDs}()},
\code{\link{killIndividuals}()},
\code{\link{mutationCounts}()},
\code{\link{mutationFrequencies}()},
\code{\link{mutationsOfType}()},
\code{\link{outputFixedMutations}()},
\code{\link{outputFull}()},
\code{\link{outputMutations}()},
\code{\link{readFromPopulationFile}()},
\code{\link{recalculateFitness}()},
\code{\link{registerFitnessEffectCallback}()},
\code{\link{registerMateChoiceCallback}()},
\code{\link{registerModifyChildCallback}()},
\code{\link{registerMutationCallback}()},
\code{\link{registerMutationEffectCallback}()},
\code{\link{registerRecombinationCallback}()},
\code{\link{registerReproductionCallback}()},
\code{\link{registerSurvivalCallback}()},
\code{\link{skipTick}()},
\code{\link{subsetMutations}()},
\code{\link{treeSeqCoalesced}()},
\code{\link{treeSeqOutput}()},
\code{\link{treeSeqRememberIndividuals}()},
\code{\link{treeSeqSimplify}()}
}
\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
\concept{Community}
\concept{Species}
