% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_scripting.R
\name{slim_block}
\alias{slim_block}
\title{Setup a SLiM code block}
\usage{
slim_block(...)
}
\arguments{
\item{...}{A list of arguments corresponding to elements in SLiM code blocks.
See details for more information on how to specify these arguments.}
}
\value{
A slimr_block object. This is of little use outside a \code{\link{slim_script}}
function call.
}
\description{
Setup a SLiM code block
}
\examples{
slim_script(slim_block({print("Hello World!")}))
}
