% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_blocks.R
\name{slim_block_progress}
\alias{slim_block_progress}
\title{Insert slim_block to track progress when using slim_run}
\usage{
slim_block_progress(update_every = 1)
}
\arguments{
\item{update_every}{How often to update progress, expressed as the number
of generations after which an update should be tracked.}
}
\value{
A \code{slimr_block} object
}
\description{
Insert slim_block to track progress when using slim_run
}
