% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_classes.R
\docType{data}
\name{slim_classes}
\alias{slim_classes}
\title{SLiM Classes}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{class_name}{The name of SLiM class}
  \item{class_abbr}{An abbreviation of the class name that can be used to save typing. They
  always begin with '.' to reduce the chance of name collisions with short variable names}
}
}
\usage{
slim_classes
}
\description{
A data.frame containing the name of the SLiM classes and an abbreviation that can be
used to refer to them in R.
}
\keyword{datasets}
