% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_code.R
\name{slim_code_Rify}
\alias{slim_code_Rify}
\title{Rify some SLiM code}
\usage{
slim_code_Rify(code_snippet)
}
\arguments{
\item{code_snippet}{SLiM code to Rify as a character vector}
}
\value{
Rified code snippet
}
\description{

}
