% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_extract.R
\name{slim_extract_genlight}
\alias{slim_extract_genlight}
\title{Extract data into a genlight object}
\usage{
slim_extract_genlight(x, ...)
}
\arguments{
\item{x}{\code{slimr_results} object containing data generated by \code{r_output}
using \code{outputFull()} in SLiM}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A \code{genlight} object
}
\description{
Extract data into a genlight object
}
\examples{
if(slim_is_avail()) {
  test_sim <- slim_script(
    slim_block_init_minimal(mutation_rate = 1e-6),
    slim_block_add_subpops(1, 100),
    slim_block(1, 20, late(), {
      r_output(sim.outputFull(), "out", do_every = 10)
    })
  ) \%>\%
    slim_run()
  if(require("adegenet", quietly = TRUE)) {
    plot(slim_extract_genlight(test_sim$output_data))
  }
}
}
