% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_extract.R
\name{slim_extract_output_data}
\alias{slim_extract_output_data}
\title{Extract data produced by SLiM}
\usage{
slim_extract_output_data(output)
}
\arguments{
\item{output}{Character vector produced from SLiM run.}
}
\value{
A \code{tibble} with four columns:
\describe{
\item{generation}{A vector of generations where output was produced.}
\item{name}{Names of the output data.}
\item{expression}{The SLiM expression used to generate the output.}
\item{data}{The raw data output from SLiM as a character vector.}
}
}
\description{
This function takes output produced from a \code{slimr_script} which uses
\code{\link{r_output}}, and converts it into a \code{tibble}
}
\examples{
if(slim_is_avail()) {
  test_sim <- slim_script(
    slim_block_init_minimal(),
    slim_block_add_subpops(1, 100),
    slim_block(1, 20, late(), {
      r_output(p1$size(), "MS", do_every = 10)
    })
  ) \%>\%
    slim_run(simple = TRUE)
  slim_extract_output_data(test_sim$output)
}
}
