% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_scripting.R
\name{slim_function}
\alias{slim_function}
\title{Specify an Eidos function to be included in the SLiM script}
\usage{
slim_function(..., name, return_type = "f$", body)
}
\arguments{
\item{...}{List of arguments specified using Eidos's argument syntax (which includes type specification; see details)}

\item{name}{Name of function being created.}

\item{return_type}{Type of the functions return, using Eidos' type syntax (see details)}

\item{body}{SLiM / Eidos code to be executed in the body of the function.}
}
\value{
A \code{slimr_block} object (only useful with the context of a \code{\link{slim_script}}) call.
}
\description{
Specify an Eidos function to be included in the SLiM script
}
