% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_code.R
\name{slim_gen_output_code}
\alias{slim_gen_output_code}
\title{Title}
\usage{
slim_gen_output_code(
  output_type = c("individuals", "genomes", "coordinates", "sexes"),
  output_format = c("VCF", "slim_tibble"),
  output_sample = 0,
  subpops = "all"
)
}
\arguments{
\item{output_format}{}
}
\value{

}
\description{
Title
}
