% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_setup.R
\name{slim_install_path}
\alias{slim_install_path}
\title{A simple exported version of install_path
Returns the SLiM installation path.}
\usage{
slim_install_path()
}
\description{
A simple exported version of install_path
Returns the SLiM installation path.
}
\examples{
slim_install_path()

}
