% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_run.R
\name{slim_run}
\alias{slim_run}
\alias{slim_run.character}
\alias{slim_run.slimr_script}
\title{Run a SLiM script from R}
\usage{
slim_run(
  x,
  slim_path = NULL,
  script_file = NULL,
  simple_run = FALSE,
  capture_output = TRUE,
  show_output = !capture_output,
  callbacks = NULL,
  new_grdev = FALSE,
  ...
)

\method{slim_run}{character}(x, ...)

\method{slim_run}{slimr_script}(x, ...)
}
\arguments{
\item{callbacks}{A list of functions to be called during the SLiM run. This can be used to
dynamically transform or visualise output from the simulation while it is running. It should
be of the form \code{function(data, ...) \{do something..\}}. If using \code{\link{slimr_output}}
to get formatted data output from SLiM, data will be a four column \code{tibble}
containing output from the current iteration of the simulation. Columns are:
\describe{
\item{generation}{A vector of generations processed in the current iteration}
\item{name}{Names of the output data.}
\item{expression}{The SLiM expression used to generate the output}
\item{data}{The raw data output from SLiM as a character vector}
}}

\item{new_grdev}{Should a new graphics device window  be opened on RStudio? This is mainly useful if you are using
custom callbacks that generate live figures, and want a faster plotting experience. This is because the
default plot viewer in RStudio can be quite slow.}

\item{...}{Additional arguments to be passed to any callback functions.}
}
\value{
A \code{slimr_results} object which has the following components:
\describe{
\item{output}{A character vector of raw output. Will be NULL if \code{capture_output is \code{FALSE}}}
\item{exit_status}{The exit status code returned by the SLiM process. 0 means success.}
}
}
\description{
This function runs a SLiM script, specified as a \code{slimr_script} object,
a character vector, or a text file.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Run a SLiM script from character vector

\item \code{slimr_script}: Run a SLiM script from slimr_script object
}}

