% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_run_script.R
\name{slim_run_script}
\alias{slim_run_script}
\title{Title}
\usage{
slim_run_script(slim_script)
}
\arguments{
\item{slim_script}{A character vector (of length 1) containing the SLiM script to run}
}
\value{

}
\description{
Title
}
