% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_run_script.R
\name{slim_run_script}
\alias{slim_run_script}
\title{Run a SLiM script and optionally return output from it to R.}
\usage{
slim_run_script(
  slim_script = NULL,
  script_file = NULL,
  slim_path = NULL,
  output = NULL,
  output_every = 1,
  output_sample = NULL,
  output_format = c("genlight", "slim_tibble", "vcf_tibble"),
  subpops = "all",
  vis_data_live = NULL,
  .progress = TRUE,
  write_data_live = FALSE,
  log_output_to = NULL,
  remove_console_output = TRUE,
  allow_output_files = TRUE,
  slim_wd = NULL
)
}
\arguments{
\item{slim_script}{This can either be a character vector (of length 1) containing the SLiM script to run or a \code{slim_script} object such as created by \code{\link{slim_script_from_text}}. If this is \code{NULL} then you must provide a text file with your script in \code{script_file}.}

\item{script_file}{A file path specifying a text file with a valid SLiM script to run. If you specify both this and \code{slim_script}, \code{slim_script} will be run and this parameter will be ignored.}

\item{slim_path}{Path to SLiM executable. If \code{NULL} \code{slimr} will try and find the SLiM executable through default locations or environmental variables. See documentation for \code{\link{slim_setup()}} for details.}

\item{output}{What would you like to output from SLiM to R? A character vector containing the names of output to generate. See details for valid output names. Can also be a list of character vectors, for a more complex sampling scenario, see details for a full explanation.}

\item{output_every}{How often should R collect output from SLiM? Output will be collected once every \code{output_every} generations.}

\item{output_sample}{When generating output from SLiM, how many samples of the population to take? 0 means sample all extant members of the population. This should be a numeric vector of length one, or of length equal to \code{output}, if \code{output} is a list (see details).}

\item{output_format}{In what format should the genomic data be outputted? Valid values are: "VCF", "genlight", "text", or "tibble". Regardless of your choice the genomes will be contained in a list column of the \code{slim_result} \link[tibble:tibble-package]{tibble} output (named "genomes").}

\item{subpops}{Character vector specifying what subpopulations in the SLiM simulation to collect output for. SLiM subpopulation names are of the form "p1", "p2", etc. The default is the special argument "all", which tells slimr to sample from all subpopulations in the simulation.}

\item{vis_data_live}{Should the output data from SLiM be visualised live (as the script is running)? This should be a vector or list specifying which elements in \code{output} should be visualised. The default \code{NULL}, mean no visualisation.}

\item{.progress}{Should a progress bar be displayed while the SLiM script is running?}

\item{write_data_live}{Should the output being read into R from SLiM be saved as it is running? Use \code{log_output_to} to specify a file to save to.}

\item{log_output_to}{File path to save output to. Ignored if \code{write_data_live} is \code{FALSE}}

\item{remove_console_output}{Should slimr remove any existing output generated by the SLiM script itself? Default is TRUE, and this is recommended if you asking slimr to collect output via the \code{output} argument, as otherwise existing SLiM output will interfere with slimr's output system.}

\item{allow_output_files}{Should slimr allow SLiM to produce output files? If FALSE, slimr will remove any file output, if TRUE, slimr will track produced files, and their filenames will be available as part of the \code{slim_results} object returned.}

\item{slim_wd}{What directory should SLiM use as its working directory? This argument is only important if \code{allow_output_files} is \code{TRUE}, otherwise it is ignored. If left NULL, a temporary directory will be corrected.}
}
\value{
A \code{slim_sim} object with output from the simulation if \code{output} is specified. If \code{output} is not specified or if \code{slim_script} is of class \code{character}, this will be the exit code of the SLiM run (exit code 0 means it ran successfully)
}
\description{
Run a SLiM script and optionally return output from it to R.
}
\details{
Valid output names that can be added to the \code{output} parameter include:
\itemize{
\item{"genomes"}{"Add this to get output on simulated genomes."}
\item{"individuals}{"Add this for output on individuals, specifically their (two) genomes."}
\item{"coordinates}{"If running a spatially explicit simulation, add this to output spatial coordinates of individuals."}
\item{"sexes}{"If running a simulation with sex, add this to output the sexes of individuals."}
}
}
