% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_run_script.R
\name{slim_run_script}
\alias{slim_run_script}
\title{Title}
\usage{
slim_run_script(
  slim_script = NULL,
  script_file = NULL,
  slim_path = NULL,
  output = NULL,
  output_every = 1,
  output_sample = NULL,
  .progress = TRUE,
  write_data_live = FALSE,
  log_output_to = NULL
)
}
\arguments{
\item{slim_script}{A character vector (of length 1) containing the SLiM script to run}

\item{log_output_to}{}
}
\value{

}
\description{
Title
}
