% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_scripting.R
\name{slim_script_render}
\alias{slim_script_render}
\title{Render a SLiM script with special slimrlang formatting}
\usage{
slim_script_render(
  slimr_script,
  template = NULL,
  replace_NAs = TRUE,
  reps = 1,
  parallel = FALSE,
  portable = FALSE
)
}
\arguments{
\item{slimr_script}{The \code{slimr_script} object to be rendered}

\item{template}{A list or data.frame containing values for any templated variables. If a list,
it must be named, where the names correspond to the variables. If a list of lists, the inner
lists must be named with the variable names, and \code{slim_script_render} will render a
separate \code{slimr_script} for each top-level list element and return it as a \code{slimr_script_coll}
object. If a \code{data.frame} (or \code{tibble}), then the column names should match the templated
variables, and \code{slim_script_render} will render a separate \code{slimr_script} for each row
and return it as a \code{slimr_script_coll} object.}

\item{replace_NAs}{Should \code{NA} values in the template be replaced by their default values?}

\item{reps}{Should the rendered script be replicated? If greater than 1, a \code{slimr_script_coll}
will be returned. This can also be used with a \code{slimr_script} object that has already been
rendered, in which case it will just repeat the rendered script in the result.}

\item{parallel}{Should the rendering be done in parallel when rendering multiple scripts? Requires
the \code{furrr} package and will use the plan set by \code{future::\link[future]{plan}}}

\item{portable}{If `TRUE`, the the script will be rendered in a 'portable' format, which allows
the script to be modified in another program such as SLiMGUI, and then reimported into R, while maintaining
`slimr` features. See details for more information on how this works.}
}
\value{
returns a rendered slim script
}
\description{
If your \code{slimr_script} object has made use of special \code{slimrlang}
syntax \code{\link{slimr_template}}, this function will 'render' the \code{slimr_script}
into valid SLiM syntax, ready to be run with SLiM or \code{\link{slim_run}}
}
