% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\docType{data}
\name{SLiMEidosBlock}
\alias{SLiMEidosBlock}
\alias{.SEB}
\title{SLiMEidosBlock}
\format{
An object of class \code{list} of length 7.
}
\usage{
SLiMEidosBlock
}
\description{
Documentation for SLiMEidosBlock class from SLiM
}
\details{
This class has the following methods (functions):
\itemize{
\item{None. This class has no methods.}
}
This class has the following properties:
\describe{
\item{active}{A property of type integer or integer or integer or string or
integer or integer or string. It is of length one (a singleton). This property
is a variable, so it is modifiable. \strong{Property Description:} If this
evaluates to logical F (i.e., is equal to 0), the script block is inactive and
will not be called. The value of active for all registered script blocks is
reset to -1 at the beginning of each generation, prior to script events being
called, thus activating all blocks. Any integer value other than -1 may be
used instead of -1 to represent that a block is active; for example, active
may be used as a counter to make a block execute a fixed number of times in
each generation. This value is not cached by SLiM; if it is changed, the new
value takes effect immediately. For example, a callback might be activated and
inactivated repeatedly during a single generation. }
\item{end}{A property of type integer or integer or integer or string or integer
or integer or string. It is of length one (a singleton). This property is a
constant, so it is not modifiable. \strong{Property Description:} The last
generation in which the script block is active. }
\item{id}{A property of type integer or integer or integer or string or integer
or integer or string. It is of length one (a singleton). This property is a
constant, so it is not modifiable. \strong{Property Description:} The identifier
for this script block; for script s3, for example, this is 3. A script block for
which no id was given will have an id of -1. }
\item{source}{A property of type integer or integer or integer or string or
integer or integer or string. It is of length one (a singleton). This property
is a constant, so it is not modifiable. \strong{Property Description:} The
source code string of the script block. }
\item{start}{A property of type integer or integer or integer or string or
integer or integer or string. It is of length one (a singleton). This property
is a constant, so it is not modifiable. \strong{Property Description:} The first
generation in which the script block is active. }
\item{tag}{A property of type integer or integer or integer or string or integer
or integer or string. It is of length one (a singleton). This property is a
variable, so it is modifiable. \strong{Property Description:} A user-defined
integer value. The value of tag is initially undefined, and it is an error to
try to read it; if you wish it to have a defined value, you must arrange that
yourself by explicitly setting its value prior to using it elsewhere in your
code. The value of tag is not used by SLiM; it is free for you to use. }
\item{type}{A property of type integer or integer or integer or string or
integer or integer or string. It is of length one (a singleton). This property
is a constant, so it is not modifiable. \strong{Property Description:} The
type of the script block; this will be "early" or "late" for the two types
of Eidos events, or "initialize", "fitness", "mateChoice", "modifyChild", or
"recombination" for the respective types of Eidos callbacks (see section 23.1
and chapter 24).}

}
}
\keyword{datasets}
