% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_blocks.R
\name{slimr_block_progress}
\alias{slimr_block_progress}
\title{Insert slim_block to track progress when using slim_run}
\usage{
slimr_block_progress(update_every = 1)
}
\arguments{
\item{update_every}{How often to update progress, expressed as the number
of generations after which an update should be tracked.}
}
\value{

}
\description{
Insert slim_block to track progress when using slim_run
}
