% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_output.R
\name{slimr_output}
\alias{slimr_output}
\title{Tell SLiM to produce easily parseable output}
\usage{
slimr_output(slimr_expr, name, do_every = 1)
}
\arguments{
\item{slimr_expr}{A SLiM expression to generate output. This can either be
a SLiM expression designed to create output, such as \code{outputFull()},
or an object created in the SLiM code, in which case \code{slimr_output}
will automatically concatenate it to a string and output it}

\item{name}{The name to use to identify this output.}

\item{do_every}{How often should the output be produced? Expressed
as an integer saying how many generations to run before producing output.
e.g. \code{do_every = 10} means to output every 10 generations of the
simulation.}
}
\value{
An expression with the code to be run in SLiM.
}
\description{
Use this function in a \code{\link{slim_block}} call and it will be
converted in the SLiM script into code to make formatted output.
This output can easily be read into R and even dynamically read
during simulation runs with \code{\link[slimr]{slim_run}} from the
\code{slimr} package. This function should generally only be
used within a \code{\link{slim_block}} call
}
