% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_output.R
\name{slimr_output_nucleotides}
\alias{slimr_output_nucleotides}
\title{Utility function to tell SLiM to output Nucleotides}
\usage{
slimr_output_nucleotides(name = "seqs", subpops = FALSE, ...)
}
\arguments{
\item{name}{Name of data column to hold sequences}

\item{subpops}{Should the subpopulation of each sequence be outputted as well?}
}
\description{
Utility function to tell SLiM to output Nucleotides
}
