% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\docType{data}
\name{SLiMSim}
\alias{SLiMSim}
\alias{.SS}
\title{SLiMSim}
\format{
An object of class \code{list} of length 48.
}
\usage{
SLiMSim
}
\description{
Documentation for SLiMSim class from SLiM
}
\details{
This class has the following methods (functions):
\itemize{
\item{\code{\link{addSubpop}}}
\item{\code{\link{addSubpopSplit}}}
\item{\code{\link{countOfMutationsOfType}}}
\item{\code{\link{deregisterScriptBlock}}}
\item{\code{\link{getValue}}}
\item{\code{\link{mutationCounts}}}
\item{\code{\link{mutationFrequencies}}}
\item{\code{\link{mutationsOfType}}}
\item{\code{\link{outputFixedMutations}}}
\item{\code{\link{outputFull}}}
\item{\code{\link{outputMutations}}}
\item{\code{\link{outputUsage}}}
\item{\code{\link{readFromPopulationFile}}}
\item{\code{\link{recalculateFitness}}}
\item{\code{\link{registerEarlyEvent}}}
\item{\code{\link{registerFitnessCallback}}}
\item{\code{\link{registerInteractionCallback}}}
\item{\code{\link{registerLateEvent}}}
\item{\code{\link{registerMateChoiceCallback}}}
\item{\code{\link{registerModifyChildCallback}}}
\item{\code{\link{registerRecombinationCallback}}}
\item{\code{\link{registerMutationCallback}}}
\item{\code{\link{registerReproductionCallback}}}
\item{\code{\link{rescheduleScriptBlock}}}
\item{\code{\link{setValue}}}
\item{\code{\link{simulationFinished}}}
\item{\code{\link{treeSeqCoalesced}}}
\item{\code{\link{treeSeqOutput}}}
\item{\code{\link{treeSeqRememberIndividuals}}}
\item{\code{\link{treeSeqSimplify}}}
}
This class has the following properties:
\describe{
\item{chromosome}{A property of type Chromosome object. It is of length
one (a singleton). This property is a constant, so it is not modifiable.
\strong{Property Description:} The Chromosome object used by the simulation. }
\item{chromosomeType}{A property of type string. It is of length one
(a singleton). This property is a constant, so it is not modifiable.
\strong{Property Description:} The type of chromosome being simulated; this will
be one of "A", "X", or "Y". }
\item{dimensionality}{A property of type string. It is of length one
(a singleton). This property is a constant, so it is not modifiable.
\strong{Property Description:} The spatial dimensionality of the simulation,
as specified in initializeSLiMOptions(). This will be "" (the empty string) for
non-spatial simulations (the default), or "x", "xy", or "xyz", for simulations
using those spatial dimensions respectively. }
\item{dominanceCoeffX}{A property of type float. It is of length one (a
singleton). This property is a variable, so it is modifiable. \strong{Property
Description:} The dominance coefficient value used to modify the selection
coefficients of mutations present on the single X chromosome of an XY male (see
the SLiM documentation for details). Used only when simulating an X chromosome;
setting a value for this property in other circumstances is an error. Changing
this will normally affect the fitness values calculated at the end of the
current generation; if you want current fitness values to be affected, you can
call SLiMSim’s method recalculateFitness() – but see the documentation of that
method for caveats. }
\item{generation}{A property of type integer. It is of length one (a singleton).
This property is a variable, so it is modifiable. \strong{Property Description:}
The current generation number. }
\item{genomicElementTypes}{A property of type GenomicElementType object. This
property is a constant, so it is not modifiable. \strong{Property Description:}
The GenomicElementType objects being used in the simulation. }
\item{inSLiMgui}{A property of type logical. It is of length one (a singleton).
This property is a constant, so it is not modifiable. \strong{Property
Description:} This property has been deprecated, and may be removed in a future
release of SLiM. In SLiM 3.2.1 and later, use exists("slimgui") instead. If T,
the simulation is presently running inside SLiMgui; if F, it is running at the
command line. In general simulations should not care where they are running,
but in special circumstances such as opening plot windows it may be necessary to
know the runtime environment. }
\item{interactionTypes}{A property of type InteractionType object. This property
is a constant, so it is not modifiable. \strong{Property Description:} The
InteractionType objects being used in the simulation. }
\item{modelType}{A property of type string. It is of length one (a singleton).
This property is a constant, so it is not modifiable. \strong{Property
Description:} The type of model being simulated, as specified in
initializeSLiMModelType(). This will be "WF" for WF models (Wright–Fisher
models, the default), or "nonWF" for nonWF models (non-Wright–Fisher models; see
section 1.6 for discussion). }
\item{mutationTypes}{A property of type MutationType object. This property
is a constant, so it is not modifiable. \strong{Property Description:} The
MutationType objects being used in the simulation. }
\item{mutations}{A property of type Mutation object. This property is a
constant, so it is not modifiable. \strong{Property Description:} The Mutation
objects that are currently active in the simulation. }
\item{nucleotideBased}{A property of type logical. It is of length one
(a singleton). This property is a constant, so it is not modifiable.
\strong{Property Description:} If T, the model is nucleotide-based; if
F, it is not. See the discussion of the nucleotideBased parameter to
initializeSLiMOptions() for discussion. }
\item{periodicity}{A property of type string. It is of length one (a singleton).
This property is a constant, so it is not modifiable. \strong{Property
Description:} The spatial periodicity of the simulation, as specified
in initializeSLiMOptions(). This will be "" (the empty string) for non-
spatial simulations and simulations with no periodic spatial dimensions
(the default). Otherwise, it will be a string representing the subset of
spatial dimensions that have been declared to be periodic, as specified to
initializeSLiMOptions(). }
\item{scriptBlocks}{A property of type SLiMEidosBlock object. This property is a
constant, so it is not modifiable. \strong{Property Description:} All registered
SLiMEidosBlock objects in the simulation. }
\item{sexEnabled}{A property of type logical. It is of length one (a singleton).
This property is a constant, so it is not modifiable. \strong{Property
Description:} If T, sex is enabled in the simulation; if F, individuals are
hermaphroditic. }
\item{subpopulations}{A property of type Subpopulation object. This property
is a constant, so it is not modifiable. \strong{Property Description:} The
Subpopulation instances currently defined in the simulation. }
\item{substitutions}{A property of type Substitution object. This property is
a constant, so it is not modifiable. \strong{Property Description:} A vector of
Substitution objects, representing all mutations that have been fixed. }
\item{tag}{A property of type integer. It is of length one (a singleton). This
property is a variable, so it is modifiable. \strong{Property Description:}
A user-defined integer value. The value of tag is initially undefined, and
it is an error to try to read it; if you wish it to have a defined value, you
must arrange that yourself by explicitly setting its value prior to using it
elsewhere in your code. The value of tag is not used by SLiM; it is free for
you to use. See also the getValue() and setValue() methods, for another way of
attaching state to the simulation.}

}
}
\keyword{datasets}
