% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{subsetIndividuals}
\alias{subsetIndividuals}
\alias{Subpopulation$subsetIndividuals}
\alias{.P$subsetIndividuals}
\title{SLiM method subsetIndividuals}
\usage{
subsetIndividuals(exclude, sex, minAge, maxAge, migrant)
}
\arguments{
\item{exclude}{An object of type null or Individual object. Must be of length 1
(a singleton). The default value is \code{NULL}. See details for description.}

\item{sex}{An object of type null or string. Must be of length 1 (a singleton).
The default value is \code{NULL],[Ni$ tag = NULL}. See details for description.}

\item{minAge}{An object of type null or integer. Must be of length 1 (a
singleton). The default value is \code{NULL}. See details for description.}

\item{maxAge}{An object of type null or integer. Must be of length 1 (a
singleton). The default value is \code{NULL}. See details for description.}

\item{migrant}{An object of type null or logical. Must be of length 1 (a
singleton). The default value is \code{NULL}. See details for description.}
}
\value{
An object of type Individual object.
}
\description{
Documentation for SLiM function \code{subsetIndividuals}, which is a method of
the SLiM class \code{Subpopulation}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Returns a vector of individuals subset from the individuals in the
target subpopulation. The parameters specify constraints upon the subset
of individuals that will be returned. Parameter exclude, if non- NULL, may
specify a specific individual that should not be included (typically the focal
individual in some operation). Parameter sex, if non-NULL, may specify a sex
("M" or "F") for the individuals to be returned, in sexual models. Parameter
tag, if non-NULL, may specify a tag value for the individuals to be returned;
only individuals whose tag property matches this value will be returned.
Parameters minAge and maxAge, if non-NULL, may specify a minimum or maximum age
for the individuals to be returned, in nonWF models. Parameter migrant, if non-
NULL, may specify a required value for the migrant property of the individuals
to be returned (so T will require that individuals be migrants, F will require
that they not be). This method is shorthand for getting the individuals property
of the subpopulation, and then using operator [] to select only individuals with
the desired properties; besides being much simpler than the equivalent Eidos
code, it is also much faster. See sampleIndividuals() for a similar method that
returns a sample taken from a chosen subset of individuals.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016–2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
