% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{transpose_sites}
\alias{transpose_sites}
\title{Transpose the simulated sites in the SLiM 0-based coordinate system into
realistic coordinate system.}
\usage{
transpose_sites(sim_sites, real_sites)
}
\arguments{
\item{sim_sites}{GRanges object with SLiM VCF output data.}

\item{real_sites}{GRanges object with real coordinates of sim_sites.}
}
\value{
GRanges object with transposed coordinates.
}
\description{
SLiM simulates all mutations as part of a single continuous "chromosome".
Therefore, simulating fixed mutations that are present in a real genome
requires their transposition into a 0-based continuous coordinate system.
This function performs this transposition based on the given positions of
the sites in their original coordinate system.
}
\seealso{
read_sites
}
