% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{treeSeqOutput}
\alias{treeSeqOutput}
\alias{SLiMSim$treeSeqOutput}
\alias{.SS$treeSeqOutput}
\title{SLiM method treeSeqOutput}
\usage{
treeSeqOutput(path, simplify)
}
\arguments{
\item{path}{An object of type string or logical. Must be of length 1 (a
singleton). See details for description.}

\item{simplify}{An object of type string or logical. Must be of length 1 (a
singleton). The default value is \code{T}. See details for description.}
}
\value{
An object of type void.
}
\description{
Documentation for SLiM function \code{treeSeqOutput}, which is a method of the
SLiM class \code{SLiMSim}.
Note that the R function is a stub, it does not do anything in R (except bring
up this documentation). It will only do
anything useful when used inside a \code{\link{slim_block}} function further
nested in a \code{\link{slim_script}}
function call, where it will be translated into valid SLiM code as part of a
full SLiM script.
}
\details{
Outputs the current tree sequence recording tables to the path
specified by path. This method may only be called if tree sequence recording
has been turned on with initializeTreeSeq(). If simplify is T (the default),
simplification will be done immediately prior to output; this is almost always
desirable, unless a model wishes to avoid simplification entirely. A binary
tree sequence file will be written to the specified path; a filename extension
of .trees is suggested for this type of file.
}
\section{Copyright}{

This is documentation for a function in the SLiM software, and has been
reproduced from the official manual,
which can be found here: \url{http://benhaller.com/slim/SLiM_Manual.pdf}. This
documentation is
Copyright © 2016–2020 Philipp Messer. All rights reserved. More information
about SLiM can be found
on the official website: \url{https://messerlab.org/slim/}
}

\author{
Benjamin C Haller (\email{bhaller@benhaller.com}) and Philipp W Messer
(\email{messer@cornell.edu})
}
