% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{validateVariants}
\alias{validateVariants}
\title{validateVariants}
\usage{
validateVariants(df, fasta, nodeN = 1)
}
\arguments{
\item{df}{A data.frame or data.table containing minimally the following
columns: 1. uniprotAccession 2. wtAA (wild-type amino acid) 3. pos (the position
of the wild-type amino acid in the protein sequence)}

\item{fasta}{AAStringSet object of protein sequences, in which names of the list items
should correspond to the uniprotAccession column of the 'df' input}

\item{nodeN}{Number of cores to use to parallelise the run}
}
\value{
A data.frame consisting of the input data.frame and two additional columns
for
1. mappedResidue (the actual residue at the variant position)
2. validity (boolean TRUE/FALSE) showing if the mapped residue matches the
wild-type residue in the corresponding variant.
}
\description{
This function is used to validate whether the missense variants are consistent
with the fasta sequences of the proteins
}
